+++ Project.toml
@@ -30,7 +30,7 @@ DataFrames = "1"
 Flux = "0.12, 0.13"
 LaplaceRedux = "0.1"
 MLJ = "0.18"
-MLUtils = "0.2"
+MLUtils = "0.2, 0.3"
 MultivariateStats = "0.9, 0.10"
 NearestNeighborModels = "0.2"
 Parameters = "0.12"
