+++ docs/src/cats_dogs.qmd
@@ -1,124 +1,0 @@
-``` @meta
-CurrentModule = CounterfactualExplanations
-```
-
-# From 🐱 to 🐶 - a motivating example
-
-Suppose we have a sample of cats and dogs with information about two features: height and tail length. Based on these two features we have trained two black box classifiers to distinguish between cats and dogs: firstly, an artificial neural network with weight regularization and secondly, that same neural network but its Bayesian counterpart (@fig-predictive below). One individual cat – let’s call her Kitty 🐱 – is friends with a lot of cool dogs and wants to be part of that group. Let's see how we can generate counterfactual paths for her.
-
-```{julia}
-#| echo: false
-# RUN IN REPL
-using Pkg.Artifacts
-www_path = "docs/src/www"
-artifact_dir = artifact"cats_dogs_images"
-img_path = joinpath(artifact_dir, "cats_dogs_images")
-if !isdir(www_path)
-    mkdir(www_path)
-end
-[cp(joinpath(img_path, file), joinpath(www_path, file), force=true) for file in readdir(img_path)]
-```
-
-```{julia}
-#| echo: false
-using Images
-mlp = load(joinpath(img_path, "predictive_mlp.png"))
-la = load(joinpath(img_path, "predictive_laplace.png"))
-img = mosaic(mlp, la, ncol=2)
-save(joinpath(www_path,"predictive.png"),img)
-```
-
-![Classification for toy dataset of cats and dogs. The contour indicates confidence in predicted labels. Left: MLP with weight regularization. Right: That same MLP, but with Laplace approximation for posterior predictive.](www/predictive.png){#fig-predictive}
-
-### From basic principles ...
-
-Counterfactual search happens in the feature space: we are interested in understanding how we need to change 🐱's attributes in order to change the output of the black-box classifier. We will start with the first model, that relies on simple plugin estimates to produce its predictions. The model was built and pre-trained using Flux and can be loaded as follows:
-
-```{julia}
-using CounterfactualExplanations
-using CounterfactualExplanations.Data: cats_dogs_model
-model = cats_dogs_model()
-```
-
-The training data can be loaded and pre-processed as follows:
-
-```{julia}
-using CounterfactualExplanations.Data: cats_dogs_data
-X, ys = cats_dogs_data()
-counterfactual_data = CounterfactualData(X,ys')
-```
-
-In order to make the Flux.jl model compatible with CounterfactualExplanations.jl we need to run the following (more on this in the [models tutorial](https://www.paltmeyer.com/CounterfactualExplanations.jl/dev/tutorials/models/)):
-
-```{julia}
-using CounterfactualExplanations.Models
-import CounterfactualExplanations.Models: logits, probs # import functions in order to extend
-
-# Step 1)
-struct NeuralNetwork <: Models.AbstractFittedModel
-    model::Any
-end
-
-# Step 2)
-logits(M::NeuralNetwork, X::AbstractArray) = M.model(X)
-probs(M::NeuralNetwork, X::AbstractArray)= σ.(logits(M, X))
-M = NeuralNetwork(model);
-```
-
-Let `x` be the 2D-feature vector describing Kitty 🐱. Based on those features she is currently labelled as `y = 0.0`. We have set the target label to `1.0` and the desired confidence in the prediction to `γ = 0.75`.  Now we can use the `GenericGenerator` for our counterfactual search as follows:
-
-```{julia}
-# Define generator:
-generator = GenericGenerator()
-# Generate recourse:
-counterfactual = generate_counterfactual(x, target, counterfactual_data, M, generator)
-```
-
[…]
