+++ docs/src/intro.qmd
@@ -2,21 +2,7 @@
 
 ## News 📣
 
-```{julia}
-#| output: true
-versioninfo()
-```
-
-
 **JuliaCon 2022**: This project was presented at JuliaCon 2022 in July 2022. See [here](https://pretalx.com/juliacon-2022/talk/HU8FVH/) for details.
-
-### Developmemts since v0.1.2
-
-- Added support for generating multiple counterfactuals instead of just one. 
-- Implemented DiCE [@mothilal2020explaining].
-- Added option to add a random perturbation to the initial factual values, which is an easy way to mitigate adversarial attacks on CE [@slack2021counterfactual].
-- Visualizations: added various plotting methods to visualize results. For $n$-dimensional data where $n>2$ we use dimensionality reduction to embed the data in 2D. The decision boundary is then visualized using its a Voronoi based representation following @migut2015visualizing. Thanks to my student [Aleksander Buszydlik](https://github.com/abuszydlik) for bringing this idea to my attention.
-- Added the option to specify strict convergence: instead of terminating search once the probability threshold $\gamma$ has been reached, search continues as long the objective function still improves.
 
 ## Installation 🚩
 
