+++ docs/src/index.qmd
@@ -1,8 +1,3 @@
----
-jupyter: julia-1.8
----
-
-
 ```@meta
 CurrentModule = CounterfactualExplanations
 ```

[…]
