### 2022-11-02 7baac64
improved plotting
