+++ src/data_preprocessing/plotting.jl
@@ -1,3 +1,4 @@
+using CategoricalArrays
 using MultivariateStats
 using Plots
 using Random
@@ -65,5 +66,6 @@ end
 
 function Plots.scatter!(data::CounterfactualData; dim_red::Symbol=:pca, kwargs...)
     X, y, _ = prepare_for_plotting(data; dim_red=dim_red)
-    Plots.scatter!(X[:,1],X[:,2],group=Int.(y),color=Int.(y); kwargs...)
+    _c = Int.(CategoricalArrays.categorical(y).refs)
+    Plots.scatter!(X[:,1],X[:,2],colour=_c; kwargs...)
 end
