### 2022-11-03 8bb764c
fixed a small issue related to initialization of latent space geenrators
