+++ dev/examples/timeseries/UCR_ECG200.qmd
@@ -1,40 +1,0 @@
----
-jupyter: julia-1.6
----
-
-```@meta
-CurrentModule = CounterfactualExplanations 
-```
-
-# UCR ECG 200
-
-In this example we will look at a time series classification task...
-
-```{julia}
-using CounterfactualExplanations, Flux
-data = CounterfactualExplanations.Data.ucr_data()
-model = CounterfactualExplanations.Data.ucr_model()
-```
-
-```{julia}
-y_train, X_train, X_test, y_test = values(data)
-n_train = size(X_train)[1]
-input_dim = size(X_train)[2]
-output_dim = 1;
-using Flux: DataLoader
-bs = Int(round(n_train/10))
-data = DataLoader((X_train[:,:,1]',Flux.unsqueeze(y_train,1)),batchsize=bs)
-test_data = zip(Flux.unstack(X_test[:,:,1],1),Flux.unsqueeze(y_test,1));
-```
-
-```{julia}
-using Statistics
-loss(x, y) = Flux.Losses.logitbinarycrossentropy(model(x), y)
-avg_loss(data) = mean(map(d -> loss(d[1],d[2]), data))
-accuracy(data) = mean(map(d -> mean(round.(Flux.σ.(model(d[1]))) .== d[2]), data));
-```
-
-```{julia}
-accuracy(test_data)
-```
-
