+++ dev/tutorials/generators.qmd
@@ -1,27 +1,11 @@
----
-format: 
-  commonmark:
-    variant: -raw_html
-    wrap: none
-    self-contained: true
-crossref:
-  fig-prefix: Figure
-  tbl-prefix: Table
-bibliography: https://raw.githubusercontent.com/pat-alt/bib/main/bib.bib
-output: asis
-execute: 
-  echo: true
-  eval: false
-jupyter: julia-1.6
----
-
 ```@meta
 CurrentModule = CounterfactualExplanations 
 ```
 
 ```{julia}
 #| echo: false
-using Flux, Random, Plots, PlotThemes, CounterfactualExplanations
+using Pkg; Pkg.activate("docs")
+using Flux, Random, Plots, CounterfactualExplanations
 theme(:wong)
 using Logging
 disable_logging(Logging.Info)
