+++ dev/tutorials/binary.qmd
@@ -1,20 +1,3 @@
-r---
-format: 
-  commonmark:
-    variant: -raw_html
-    wrap: none
-    self-contained: true
-crossref:
-  fig-prefix: Figure
-  tbl-prefix: Table
-bibliography: https://raw.githubusercontent.com/pat-alt/bib/main/bib.bib
-output: asis
-execute: 
-  echo: true
-  eval: false
-jupyter: julia-1.6
----
-
 ```@meta
 CurrentModule = CounterfactualExplanations 
 ```
