### 2022-11-05 c0a85ec
CompatHelper: add new compat entry for CategoricalArrays at version 0.10, (keep existing compat)
