### 2022-11-05 57cc11e
CompatHelper: add new compat entry for PkgTemplates at version 0.7, (keep existing compat)
