+++ Project.toml
@@ -37,6 +37,7 @@ MLUtils = "0.2"
 MultivariateStats = "0.9, 0.10"
 NearestNeighborModels = "0.2"
 Parameters = "0.12"
+PkgTemplates = "0.7"
 Plots = "1"
 SliceMap = "0.2"
 StatsBase = "0.33"
