### 2022-11-05 eb1b478
CompatHelper: add new compat entry for Revise at version 3, (keep existing compat)
