+++ Project.toml
@@ -38,6 +38,7 @@ MultivariateStats = "0.9, 0.10"
 NearestNeighborModels = "0.2"
 Parameters = "0.12"
 Plots = "1"
+Revise = "3"
 SliceMap = "0.2"
 StatsBase = "0.33"
 UMAP = "0.1"
