+++ Project.toml
@@ -17,10 +17,8 @@ MLUtils = "f1d291b0-491e-4a28-83b9-f70985020b54"
 MultivariateStats = "6f286f6a-111f-5878-ab1e-185364afe411"
 NearestNeighborModels = "636a865e-7cf4-491e-846c-de09b730eb36"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
-PkgTemplates = "14b8a8f1-9102-5b29-a752-f990bacb7fe1"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 SliceMap = "82cb661a-3f19-5665-9e27-df437c7e54c8"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
