+++ src/counterfactuals/functions.jl
@@ -59,14 +59,9 @@ function CounterfactualExplanation(
 
     # Counterfactual state variable:
     size_ = Int.(vcat(ones(maximum([ndims(x),2])),num_counterfactuals))
-    s′ = copy(x)  # start from factual
-    s′ = repeat(x, outer=size_)
-    # Initialization
-    if initialization == :add_perturbation
-        scale = std(data.X, dims=2) .* 0.1
-        s′ = mapslices(s -> s .+ scale .* randn(size(scale,1)), s′, dims=(1,2))
-    end
-    f(s) = s # default mapping to feature space 
+    s′ = copy(x)                    # start from factual
+    s′ = repeat(x, outer=size_)     # augment to account for specified number of counterfactuals
+    f(s) = s                        # default mapping f: S ↦ X
 
     # Parameters:
     params = Dict(
@@ -76,15 +71,14 @@ function CounterfactualExplanation(
         :initial_mutability => repeat(DataPreprocessing.mutability_constraints(data), outer=size_),
     )
 
-    # Latent space:
-    wants_latent_space = DataPreprocessing.has_pretrained_generative_model(data) || typeof(generator) <: Generators.AbstractLatentSpaceGenerator
-    latent_space = isnothing(latent_space) ? wants_latent_space : latent_space
-
     # Instantiate: 
     counterfactual_explanation = CounterfactualExplanation(
         x, target, nothing, s′, f, 
-        data, M, generator, latent_space, params, nothing, num_counterfactuals, initialization
+        data, M, generator, false, params, nothing, num_counterfactuals, initialization
     )
+
+    # Counterfactual initialization:
+    initialize!(counterfactual_explanation)
 
     # Encode target:
     counterfactual_explanation.target_encoded = encode_target(counterfactual_explanation)
@@ -110,6 +104,8 @@ function CounterfactualExplanation(
     )
 
     # Latent space:
+    wants_latent_space = DataPreprocessing.has_pretrained_generative_model(data) || typeof(generator) <: Generators.AbstractLatentSpaceGenerator
+    counterfactual_explanation.latent_space = isnothing(latent_space) ? wants_latent_space : latent_space
     if counterfactual_explanation.latent_space && !counterfactual_explanation.search[:terminated]
         @info "Searching in latent space using generative model."
         generative_model = DataPreprocessing.get_generative_model(counterfactual_explanation.data; generative_model_params...)
@@ -153,6 +149,22 @@ function encode_target(counterfactual_explanation::CounterfactualExplanation)
     target = out_dim > 1 ? Flux.onehot(target, 1:out_dim) : [target]
     target = repeat(target, outer=[1,1,counterfactual_explanation.num_counterfactuals])
     return target
+end
+
+function initialize!(counterfactual_explanation::CounterfactualExplanation)
+
+    s′ = counterfactual_explanation.s′
+    data = counterfactual_explanation.data
+
+    if counterfactual_explanation.initialization == :add_perturbation
+        scale = std(data.X, dims=2) .* 0.1
+        s′ = mapslices(s′, dims=(1,2)) do s
+            Δs′ = scale .* randn(size(scale,1))
+            Δs′ = apply_mutability(counterfactual_explanation, Δs′)    
+            s .+ Δs′
+        end
+        counterfactual_explanation.s′ = s′
+    end
 end
 
 # 1) Factual values
@@ -341,11 +353,17 @@ function embed_path(counterfactual_explanation::CounterfactualExplanation)
 end
 
 """
-    apply_mutability(Δx′::AbstractArray, counterfactual_explanation::CounterfactualExplanation)
+    apply_mutability(counterfactual_explanation::CounterfactualExplanation, Δs′::AbstractArray)
 
 A subroutine that applies mutability constraints to the proposed vector of feature perturbations.
 """
[…]
