+++ _quarto.yml
@@ -12,3 +12,5 @@ execute:
   echo: true
   output: false
 
+jupyter: julia-1.8
+
