+++ Project.toml
@@ -31,7 +31,7 @@ CategoricalArrays = "0.10"
 DataFrames = "1"
 Flux = "0.12, 0.13"
 LaplaceRedux = "0.1"
-MLJ = "0.18"
+MLJ = "0.18, 0.19"
 MLUtils = "0.2, 0.3"
 MultivariateStats = "0.9, 0.10"
 NearestNeighborModels = "0.2"
