+++ src/counterfactuals/functions.jl
@@ -11,11 +11,11 @@ mutable struct CounterfactualExplanation
     target::Number
     target_encoded::Union{Number, AbstractArray, Nothing}
     s′::AbstractArray
-    f::Function
     data::DataPreprocessing.CounterfactualData
     M::Models.AbstractFittedModel
     generator::Generators.AbstractGenerator
-    latent_space::Bool
+    latent_space::Union{Nothing, Bool}
+    generative_model_params::NamedTuple
     params::Dict
     search::Union{Dict,Nothing}
     num_counterfactuals::Int
@@ -70,8 +70,9 @@ function CounterfactualExplanation(
 
     # Instantiate: 
     counterfactual_explanation = CounterfactualExplanation(
-        x, target, nothing, s′, f, 
-        data, M, generator, false, params, nothing, num_counterfactuals, initialization
+        x, target, nothing, x,
+        data, M, generator, latent_space, generative_model_params, 
+        params, nothing, num_counterfactuals, initialization
     )
 
     # Initialization:
@@ -80,38 +81,18 @@ function CounterfactualExplanation(
     counterfactual_explanation.s′ = initialize_state(counterfactual_explanation)                # initialize the counterfactual state
     counterfactual_explanation.target_encoded = encode_target(counterfactual_explanation)       # encode the target variable
 
-    # Check for redundancy:
-    if threshold_reached(counterfactual_explanation)
-        @info "Factual already in target class and probability exceeds threshold γ."
-    end
-
     # Initialize search:
     counterfactual_explanation.search = Dict(
         :iteration_count => 0,
-        :times_changed_features => zeros(size(counterfactual_explanation.f(counterfactual_explanation.s′))),
+        :times_changed_features => zeros(size(decode_state(counterfactual_explanation))),
         :path => [counterfactual_explanation.s′],
-        :terminated => threshold_reached(counterfactual_explanation),
+        :terminated => threshold_reached(counterfactual_explanation, counterfactual_explanation.x),
         :converged => converged(counterfactual_explanation),
     )
 
-    # Latent space:
-    wants_latent_space = DataPreprocessing.has_pretrained_generative_model(data) || typeof(generator) <: Generators.AbstractLatentSpaceGenerator
-    counterfactual_explanation.latent_space = isnothing(latent_space) ? wants_latent_space : latent_space
-    if counterfactual_explanation.latent_space && !terminated(counterfactual_explanation)
-        @info "Searching in latent space using generative model."
-        generative_model = DataPreprocessing.get_generative_model(counterfactual_explanation.data; generative_model_params...)
-        # map counterfactual to latent space: s′=z′∼p(z|x)
-        counterfactual_explanation.s′, _, _ = GenerativeModels.rand(generative_model.encoder, counterfactual_explanation.s′)
-
-        # NOTE! This is not very clean, will be improved.
-        if generative_model.params.nll==Flux.Losses.logitbinarycrossentropy
-            counterfactual_explanation.f = function(s) Flux.σ.(generative_model.decoder(s)) end
-        else
-            counterfactual_explanation.f = function(s) generative_model.decoder(s) end
-        end
-        
-        # Path
-        counterfactual_explanation.search[:path] = [counterfactual_explanation.s′]
+    # Check for redundancy:
+    if terminated(counterfactual_explanation)
+        @info "Factual already in target class and probability exceeds threshold γ."
     end
 
     return counterfactual_explanation
@@ -149,22 +130,35 @@ Encodes counterfactual.
 function encode_state(counterfactual_explanation::CounterfactualExplanation)
 
     s′ = counterfactual_explanation.s′
+    data = counterfactual_explanation.data
 
     # Standardization:
[…]
