### 2022-11-16 8101f97
next i will try to get rid of the State constructor altogether
