### 2022-11-17 444c837
still struggling with decoding (reconstruction)
