+++ src/benchmark/functions.jl
@@ -7,7 +7,7 @@ using Statistics
 Computes the success rate.
 """
 function success_rate(counterfactual_explanation::CounterfactualExplanation; agg=mean)
-    agg(Counterfactuals.target_probs(counterfactual_explanation) .>= counterfactual_explanation.params[:γ])
+    agg(CounterfactualExplanations.target_probs(counterfactual_explanation) .>= counterfactual_explanation.params[:γ])
 end   
 
 function success_rate(counterfactual_explanations::Vector{CounterfactualExplanation}; agg=mean)
@@ -20,8 +20,8 @@ end
 Computes the distance.
 """
 function distance(counterfactual_explanation::CounterfactualExplanation; agg=mean)
-    x = Counterfactuals.factual(counterfactual_explanation)
-    x′ = Counterfactuals.counterfactual(counterfactual_explanation)
+    x = CounterfactualExplanations.factual(counterfactual_explanation)
+    x′ = CounterfactualExplanations.counterfactual(counterfactual_explanation)
     return agg(LinearAlgebra.norm.(x.-x′))
 end
 
@@ -35,8 +35,8 @@ end
 Computes the feature redundancy.
 """
 function redundancy(counterfactual_explanation::CounterfactualExplanation; agg=mean)
-    x′ = Counterfactuals.counterfactual(counterfactual_explanation)
-    redundant_x = mapslices(x -> sum(x .== 0)/size(x,1), x′, dims=[1,2])
+    x′ = CounterfactualExplanations.counterfactual(counterfactual_explanation)
+    redundant_x = mapslices(x -> sum(x .== 0)/size(x,1), x′, dims=(1,2))
     return agg(redundant_x)
 end
 
