+++ src/generators/gradient_based/GreedyGenerator.jl
@@ -1,5 +1,6 @@
 using Parameters
 using LinearAlgebra
+using SliceMap
 
 # -------- Schut et al (2020): 
 mutable struct GreedyGenerator <: AbstractGradientBasedGenerator
@@ -72,21 +73,21 @@ function GreedyGenerator(
 end
 
 """
-    ∇(generator::GreedyGenerator, counterfactual_state::CounterfactualState.State)    
+    ∇(generator::GreedyGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)    
 
 he default method to compute the gradient of the counterfactual search objective for a greedy generator. Since no complexity penalty is needed, this gradients just correponds to the partial derivative with respect to the loss function.
 
 """
-∇(generator::GreedyGenerator, M::Models.AbstractDifferentiableModel, counterfactual_state::CounterfactualState.State) = ∂ℓ(generator, M, counterfactual_state)
+∇(generator::GreedyGenerator, M::Models.AbstractDifferentiableModel, counterfactual_explanation::AbstractCounterfactualExplanation) = ∂ℓ(generator, M, counterfactual_explanation)
 
 """
-    generate_perturbations(generator::GreedyGenerator, counterfactual_state::CounterfactualState.State)
+    generate_perturbations(generator::GreedyGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)
 
 The default method to generate perturbations for a greedy generator. Only the most salient feature is perturbed.
 """
-function generate_perturbations(generator::GreedyGenerator, counterfactual_state::CounterfactualState.State) 
-    𝐠ₜ = ∇(generator, counterfactual_state.M, counterfactual_state) # gradient
-    𝐠ₜ[counterfactual_state.params[:mutability] .== :none] .= 0
+function generate_perturbations(generator::GreedyGenerator, counterfactual_explanation::AbstractCounterfactualExplanation) 
+    𝐠ₜ = ∇(generator, counterfactual_explanation.M, counterfactual_explanation) # gradient
+    𝐠ₜ[counterfactual_explanation.params[:mutability] .== :none] .= 0
     function choose_most_salient(x)
         s = -((abs.(x).==maximum(abs.(x),dims=1)) .* generator.ϵ .* sign.(x))
         non_zero_elements = findall(vec(s).!=0)
@@ -99,23 +100,23 @@ function generate_perturbations(generator::GreedyGenerator, counterfactual_state
         end
         return s
     end
-    Δs′ = mapslices(x -> choose_most_salient(x), 𝐠ₜ, dims=1) # choose most salient feature
+    Δs′ = SliceMap.slicemap(x -> choose_most_salient(x), 𝐠ₜ, dims=1) # choose most salient feature
     return Δs′
 end
 
 """
-    mutability_constraints(generator::GreedyGenerator, counterfactual_state::CounterfactualState.State)
+    mutability_constraints(generator::GreedyGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)
 
 The default method to return search state dependent mutability constraints for a greedy generator. Features that have been perturbed `n` times already can no longer be perturbed.
 """
-function mutability_constraints(generator::GreedyGenerator, counterfactual_state::CounterfactualState.State)
-    mutability = counterfactual_state.params[:mutability]
-    if all(counterfactual_state.search[:times_changed_features] .>= generator.n) 
+function mutability_constraints(generator::GreedyGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)
+    mutability = counterfactual_explanation.params[:mutability]
+    if all(counterfactual_explanation.search[:times_changed_features] .>= generator.n) 
         generator.passes += 1
         generator.n += generator.n/generator.passes
[…]
