+++ src/counterfactual_state/functions.jl
@@ -18,15 +18,15 @@ struct State
 end
 
 """
-    guess_loss(counterfactual_state::State)
+    guess_loss(counterfactual_explanation::CounterfactualExplanation)
 
 Guesses the loss function to be used for the counterfactual search in case `likelihood` field is specified for the [`AbstractFittedModel`](@ref) instance and no loss function was explicitly declared for [`AbstractGenerator`](@ref) instance.
 """
-function guess_loss(counterfactual_state::State)
-    if :likelihood in fieldnames(typeof(counterfactual_state.M))
-        if counterfactual_state.M.likelihood == :classification_binary
+function guess_loss(counterfactual_explanation::CounterfactualExplanation)
+    if :likelihood in fieldnames(typeof(counterfactual_explanation.M))
+        if counterfactual_explanation.M.likelihood == :classification_binary
             loss_fun = Flux.Losses.logitbinarycrossentropy
-        elseif counterfactual_state.M.likelihood == :classification_multi
+        elseif counterfactual_explanation.M.likelihood == :classification_multi
             loss_fun = Flux.Losses.logitcrossentropy
         else
             loss_fun = Flux.Losses.mse
