### 2022-11-17 7885c3c
reconstruction as part of function that is differentiated just isn't working. Might need a different approach altogether here :(
