+++ src/data_preprocessing/DataPreprocessing.jl
@@ -1,6 +1,7 @@
 module DataPreprocessing
 
 using ..GenerativeModels
+include("transformations.jl")
 include("functions.jl")
 export CounterfactualData, select_factual, apply_domain_constraints
 
