+++ Project.toml
@@ -12,7 +12,7 @@ Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 LaplaceRedux = "c52c1a26-f7c5-402b-80be-ba1e638ad478"
 LazyArtifacts = "4af54fe1-eca0-43a8-85a7-787d91b784e3"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-MLJ = "add582a8-e3ab-11e8-2d5e-e98b27df1bc7"
+MLJBase = "a7f614a8-145f-11e9-1d2a-a57a1082229d"
 MLUtils = "f1d291b0-491e-4a28-83b9-f70985020b54"
 MultivariateStats = "6f286f6a-111f-5878-ab1e-185364afe411"
 NearestNeighborModels = "636a865e-7cf4-491e-846c-de09b730eb36"
@@ -31,7 +31,6 @@ CategoricalArrays = "0.10"
 DataFrames = "1"
 Flux = "0.12, 0.13"
 LaplaceRedux = "0.1"
-MLJ = "0.18, 0.19"
 MLUtils = "0.2, 0.3"
 MultivariateStats = "0.9, 0.10"
 NearestNeighborModels = "0.2"
