### 2022-12-15 21282a4
got rid of mnist data functions, need to adjust docs
