+++ Project.toml
@@ -35,7 +35,6 @@ MLUtils = "0.2, 0.3"
 MultivariateStats = "0.9, 0.10"
 NearestNeighborModels = "0.2"
 Parameters = "0.12"
-Plots = "1"
 SliceMap = "0.2"
 StatsBase = "0.33"
 UMAP = "0.1"
