+++ Project.toml
@@ -31,10 +31,12 @@ CategoricalArrays = "0.10"
 DataFrames = "1"
 Flux = "0.12, 0.13"
 LaplaceRedux = "0.1"
+MLJBase = "0.21.3"
 MLUtils = "0.2, 0.3"
 MultivariateStats = "0.9, 0.10"
 NearestNeighborModels = "0.2"
 Parameters = "0.12"
+Plots = "1.37.2"
 SliceMap = "0.2"
 StatsBase = "0.33"
 UMAP = "0.1"
