+++ test/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+CompatHelperLocal = "5224ae11-6099-4aaa-941d-3aab004bd678"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
