+++ v0.1.4/index.qmd
@@ -1,0 +1,9 @@
+```@meta
+CurrentModule = CounterfactualExplanations
+```
+
+# CounterfactualExplanations
+
+Documentation for [CounterfactualExplanations.jl](https://github.com/pat-alt/CounterfactualExplanations.jl).
+
+{{< include intro.qmd >}}
