+++ v0.1.4/generators/gradient_based/generic.qmd
@@ -1,0 +1,88 @@
+---
+format: 
+  commonmark:
+    variant: -raw_html
+    wrap: none
+    self-contained: true
+crossref:
+  fig-prefix: Figure
+  tbl-prefix: Table
+bibliography: https://raw.githubusercontent.com/pat-alt/bib/main/bib.bib
+output: asis
+execute: 
+  eval: false
+  echo: true
+jupyter: julia-1.7
+---
+
+```@meta
+CurrentModule = CounterfactualExplanations 
+```
+
+```{julia}
+#| echo: false
+using Pkg; Pkg.activate("docs")
+using Flux, Random, Plots, PlotThemes, CounterfactualExplanations
+theme(:wong)
+www_path = "docs/src/generators/www/"
+include("docs/src/utils.jl")
+```
+
+# `GenericGenerator`
+
+```{julia}
+using MLJ
+N = 1000
+X, ys = make_blobs(N, 2; centers=2, as_table=false, center_box=(-5 => 5), cluster_std=0.5)
+ys .= ys.==2
+X = X'
+xs = Flux.unstack(X,2)
+data = zip(xs,ys)
+counterfactual_data = CounterfactualData(X,ys')
+```
+
+```{julia}
+using Flux
+nn = Chain(Dense(2,1))
+using Flux.Optimise: update!, Adam
+opt = Adam()
+epochs = 100
+loss(x, y) = Flux.Losses.logitbinarycrossentropy(nn(x), y)
+avg_loss(data) = mean(map(d -> loss(d[1],d[2]), data))
+show_every = epochs/10
+for epoch = 1:epochs
+  for d in data
+    gs = gradient(Flux.params(nn)) do
+      l = loss(d...)
+    end
+    update!(opt, Flux.params(nn), gs)
+  end
+  if epoch % show_every == 0
+    println("Epoch " * string(epoch))
+    @show avg_loss(data)
+  end
+end
+```
+
+```{julia}
+M = FluxModel(nn)
+```
+
+```{julia}
+x = select_factual(counterfactual_data, rand(1:size(X)[2])) 
+y = round(probs(M, x)[1])
+target = ifelse(y==1.0,0.0,1.0) # opposite label as target
+```
+
+```{julia}
+# Define generator:
+generator = GenericGenerator()
+# Generate recourse:
+counterfactual = generate_counterfactual(x, target, counterfactual_data, M, generator)
+```
+
+```{julia}
+using CounterfactualExplanations.Counterfactuals: animate_path
+anim = animate_path(counterfactual)
+gif(anim, joinpath(www_path, "generic.gif"), fps=10)
+```

[…]
