### 2022-12-17 3bfb79f
added generators. ClapROAR currently not working for multi-class
