+++ src/CounterfactualExplanations.jl
@@ -40,12 +40,14 @@ export State
 include("generators/Generators.jl")
 using .Generators
 export AbstractGenerator, AbstractGradientBasedGenerator
+export ClapROARGenerator, ClapROARGeneratorParams
 export GenericGenerator, GenericGeneratorParams
+export GravitationalGenerator, GravitationalGeneratorParams
 export GreedyGenerator, GreedyGeneratorParams
 export REVISEGenerator, REVISEGeneratorParams
 export DiCEGenerator, DiCEGeneratorParams
 export generator_catalog
-export generate_perturbations, conditions_satisified, mutability_constraints   
+export generate_perturbations, conditions_satisified, mutability_constraints  
 
 ### CounterfactualExplanation
 # argmin 
