### 2022-12-17 5f04080
adjusted new generators, ClapROAR still fails. Latent space issue still not addressed
