+++ src/data/functions.jl
@@ -12,7 +12,7 @@ A convenience function to load the cats and dogs toy dataset.
 # Cats and dogs:
 function cats_dogs_data()
     data_dir = artifact"cats_dogs_data"
-    data = BSON.load(joinpath(data_dir,"cats_dogs_data.bson"),@__MODULE__)[:data]
+    data = BSON.load(joinpath(data_dir, "cats_dogs_data.bson"), @__MODULE__)[:data]
     X, y = (data[:X], data[:y])
     return X, y
 end
@@ -24,7 +24,7 @@ A convenience function to load the pre-trained MLP to classify cats and dogs.
 """
 function cats_dogs_model()
     data_dir = artifact"cats_dogs_model"
-    model = BSON.load(joinpath(data_dir,"cats_dogs_model.bson"),@__MODULE__)[:model]
+    model = BSON.load(joinpath(data_dir, "cats_dogs_model.bson"), @__MODULE__)[:model]
     return model
 end
 
@@ -35,7 +35,7 @@ A convenience function to load the pre-trained MLP with Laplace approximation to
 """
 function cats_dogs_laplace()
     data_dir = artifact"cats_dogs_laplace"
-    la = BSON.load(joinpath(data_dir,"cats_dogs_laplace.bson"),@__MODULE__)[:la]
+    la = BSON.load(joinpath(data_dir, "cats_dogs_laplace.bson"), @__MODULE__)[:la]
     return la
 end
 
@@ -51,18 +51,19 @@ toy_data_linear()
 ```
 
 """
-function toy_data_linear(N=100, p=2)
+function toy_data_linear(N = 100, p = 2)
     # Number of points to generate.
     M = round(Int, N / 2)
 
     # Generate artificial data.
-    x1s = rand(M) * 4.5; x2s = rand(M) * 4.5; 
+    x1s = rand(M) * 4.5
+    x2s = rand(M) * 4.5
     xt1s = Array([[x1s[i] + 0.5; x2s[i] + 0.5] for i = 1:M])
     xt0s = Array([[x1s[i] - 5; x2s[i] - 5] for i = 1:M])
 
     # Store all the data for later.
     xs = [xt1s; xt0s]
-    ts = [ones(M); zeros(M)];
+    ts = [ones(M); zeros(M)]
     return xs, ts
 end
 
@@ -78,24 +79,28 @@ toy_data_non_linear()
 ```
 
 """
-function toy_data_non_linear(N=100)
+function toy_data_non_linear(N = 100)
     # Number of points to generate.
     M = round(Int, N / 4)
 
     # Generate artificial data.
-    x1s = rand(M) * 4.5; x2s = rand(M) * 4.5; 
+    x1s = rand(M) * 4.5
+    x2s = rand(M) * 4.5
     xt1s = Array([[x1s[i] + 0.5; x2s[i] + 0.5] for i = 1:M])
-    x1s = rand(M) * 4.5; x2s = rand(M) * 4.5; 
+    x1s = rand(M) * 4.5
+    x2s = rand(M) * 4.5
     append!(xt1s, Array([[x1s[i] - 5; x2s[i] - 5] for i = 1:M]))
 
-    x1s = rand(M) * 4.5; x2s = rand(M) * 4.5; 
+    x1s = rand(M) * 4.5
+    x2s = rand(M) * 4.5
     xt0s = Array([[x1s[i] + 0.5; x2s[i] - 5] for i = 1:M])
-    x1s = rand(M) * 4.5; x2s = rand(M) * 4.5; 
+    x1s = rand(M) * 4.5
+    x2s = rand(M) * 4.5
     append!(xt0s, Array([[x1s[i] - 5; x2s[i] + 0.5] for i = 1:M]))
 
     # Store all the data for later.
[…]
