+++ src/data_preprocessing/transformations.jl
@@ -5,8 +5,12 @@ using StatsBase
 
 Method extension to avoid array mutation issue with Zygote.jl.
 """
-function StatsBase.reconstruct(dt::ZScoreTransform, y::AbstractMatrix{<:Real}, idx::AbstractArray)
-    D = size(y,1)
+function StatsBase.reconstruct(
+    dt::ZScoreTransform,
+    y::AbstractMatrix{<:Real},
+    idx::AbstractArray,
+)
+    D = size(y, 1)
     _scale = ones(D)
     _scale[idx] .= dt.scale
     _mean = zeros(D)
