+++ src/counterfactuals/Counterfactuals.jl
@@ -5,4 +5,3 @@ using .Models
 
 include("functions.jl")
 include("plotting.jl")
-
