+++ docs/src/generators/gradient_based/latent_space_generator.qmd
@@ -21,11 +21,8 @@ CurrentModule = CounterfactualExplanations
 
 ```{julia}
 #| echo: false
-using Pkg; Pkg.activate("docs")
-using Flux, Random, Plots, PlotThemes, CounterfactualExplanations
-theme(:wong)
-www_path = "docs/src/generators/www/"
-include("docs/src/utils.jl")
+include("docs/setup_docs.jl")
+eval(setup_docs)
 ```
 
 # Latent Space Search
