### 2022-12-18 1a6a066
some progress wrt standardization
