+++ src/counterfactuals/functions.jl
@@ -88,7 +88,7 @@ function CounterfactualExplanation(;
     )
 
     # Initialization:
-    reshape!(counterfactual_explanation)                                                        # adjust shape to specified number of counterfactuals
+    adjust_shape!(counterfactual_explanation)                                                        # adjust shape to specified number of counterfactuals
     counterfactual_explanation.s′ = encode_state(counterfactual_explanation)                    # encode the counterfactual state
     counterfactual_explanation.s′ = initialize_state(counterfactual_explanation)                # initialize the counterfactual state
     counterfactual_explanation.target_encoded = encode_target(counterfactual_explanation)       # encode the target variable
@@ -97,7 +97,7 @@ function CounterfactualExplanation(;
     counterfactual_explanation.search = Dict(
         :iteration_count => 0,
         :times_changed_features =>
-            zeros(size(decode_state(counterfactual_explanation))),
+            zeros(size(map_from_latent(counterfactual_explanation))),
         :path => [counterfactual_explanation.s′],
         :terminated =>
             threshold_reached(counterfactual_explanation, counterfactual_explanation.x),
@@ -122,9 +122,19 @@ A convenience method that computes the output dimension of the predictive model.
 output_dim(counterfactual_explanation::CounterfactualExplanation) =
     size(Models.probs(counterfactual_explanation.M, counterfactual_explanation.x))[1]
 
-function reshape!(counterfactual_explanation::CounterfactualExplanation)
-    # Dimensionality:
-    x = counterfactual_explanation.x
+"""
+    adjust_shape(
+        counterfactual_explanation::CounterfactualExplanation, 
+        x::AbstractArray
+    )
+
+A convenience method that adjust the dimensions of `x`.
+"""
+function adjust_shape(
+    counterfactual_explanation::CounterfactualExplanation, 
+    x::AbstractArray
+)
+
     size_ =
         Int.(
             vcat(
@@ -132,13 +142,28 @@ function reshape!(counterfactual_explanation::CounterfactualExplanation)
                 counterfactual_explanation.num_counterfactuals,
             )
         )
-    s′ = copy(x)                                        # start from factual
-    s′ = repeat(x, outer = size_)                         # augment to account for specified number of counterfactuals
+    s′ = copy(x)                    
+    s′ = repeat(x, outer = size_) 
+
+    return s′ 
+
+end
+
+"""
+    adjust_shape!(counterfactual_explanation::CounterfactualExplanation)
+
+A convenience method that adjusts the dimensions of the counterfactual state and related fields.
+"""
+function adjust_shape!(counterfactual_explanation::CounterfactualExplanation)
+
+    # Dimensionality:
+    x = deepcopy(counterfactual_explanation.x)
+    s′ = adjust_shape(counterfactual_explanation, x)      # augment to account for specified number of counterfactuals
     counterfactual_explanation.s′ = s′
 
     # Parameters:
     params = counterfactual_explanation.params
-    params[:mutability] = repeat(params[:mutability], outer = size_)
+    params[:mutability] = adjust_shape(counterfactual_explanation, params[:mutability])      # augment to account for specified number of counterfactuals
     params[:initial_mutability] = params[:mutability]
     counterfactual_explanation.params = params
 end
@@ -150,29 +175,47 @@ Encodes counterfactual.
 """
 function encode_state(counterfactual_explanation::CounterfactualExplanation)
 
-    s′ = counterfactual_explanation.s′
[…]
