+++ src/generators/gradient_based/GravitationalGenerator.jl
@@ -15,7 +15,7 @@ end
 using Parameters, Flux
 @with_kw struct GravitationalGeneratorParams
     opt::Any = Flux.Optimise.Descent()
-    τ::AbstractFloat = 1e-5
+    τ::AbstractFloat = 1e-3
     K::Int = 50
     centroid::Union{Nothing,AbstractArray} = nothing
 end
