+++ src/generators/gradient_based/REVISEGenerator.jl
@@ -14,7 +14,7 @@ end
 # API streamlining:
 @with_kw struct REVISEGeneratorParams
     opt::Any = Flux.Optimise.Descent()
-    τ::AbstractFloat = 1e-5
+    τ::AbstractFloat = 1e-3
 end
 
 """
