+++ src/generators/gradient_based/DICEGenerator.jl
@@ -14,7 +14,7 @@ end
 # API streamlining:
 @with_kw struct DiCEGeneratorParams
     opt::Any = Flux.Optimise.Descent(1.0)
-    τ::AbstractFloat = 1e-5
+    τ::AbstractFloat = 1e-3
 end
 
 """
