+++ src/models/differentiable/simple.jl
@@ -1,204 +1,0 @@
-using MLUtils
-using SliceMap
-
-"""
-    LogisticModel <: AbstractDifferentiableJuliaModel
-
-Constructs a logistic classifier based on arrays containing coefficients `w` and constant terms `b`.
-
-# Examples
-
-```julia-repl
-w = [1.0 -2.0] # estimated coefficients
-b = [0] # estimated constant
-M = CounterfactualExplanations.Models.LogisticModel(w, b);
-```
-
-See also: 
-- [`logits(M::LogisticModel, X::AbstractArray)`](@ref)
-- [`probs(M::LogisticModel, X::AbstractArray)`](@ref)
-"""
-struct LogisticModel <: AbstractDifferentiableJuliaModel
-    W::Matrix
-    b::AbstractArray
-    likelihood::Symbol
-end
-
-LogisticModel(W, b; likelihood = :classification_binary) = LogisticModel(W, b, likelihood)
-
-# What follows are the two required outer methods:
-"""
-    logits(M::LogisticModel, X::AbstractArray)
-
-Computes logits as `WX+b`.
-
-# Examples
-
-```julia-repl
-using CounterfactualExplanations.Models
-w = [1.0 -2.0] # estimated coefficients
-b = [0] # estimated constant
-M = LogisticModel(w, b);
-x = [1,1]
-logits(M, x)
-```
-
-See also [`LogisticModel <: AbstractDifferentiableJuliaModel`](@ref).
-"""
-function logits(M::LogisticModel, X::AbstractArray)
-    if ndims(X) == 3
-        n = size(X, 3)
-        reshape(map(x -> (M.W*x.+M.b)[1], [X[:, :, i] for i ∈ 1:n]), 1, 1, n)
-        # SliceMap.slicemap(x -> M.W*x .+ M.b, X, dims=(1,2)) 
-    else
-        M.W * X .+ M.b
-    end
-end
-
-"""
-    probs(M::LogisticModel, X::AbstractArray)
-
-Computes predictive probabilities from logits as `σ(WX+b)` where 'σ' is the [sigmoid function](https://en.wikipedia.org/wiki/Sigmoid_function). 
-
-# Examples
-
-```julia-repl
-using CounterfactualExplanations.Models
-w = [1.0 -2.0] # estimated coefficients
-b = [0] # estimated constant
-M = LogisticModel(w, b);
-x = [1,1]
-probs(M, x)
-```
-
-See also [`LogisticModel <: AbstractDifferentiableJuliaModel`](@ref).
-"""
-probs(M::LogisticModel, X::AbstractArray) = Flux.σ.(logits(M, X))
-
-
-"""
-    BayesianLogisticModel <: AbstractDifferentiableJuliaModel
-
-Constructs a Bayesian logistic classifier based on maximum a posteriori (MAP) estimates `μ` (coefficients including constant term(s)) and `Σ` (covariance matrix). 
-
-# Examples
-
-```julia-repl
-using Random, LinearAlgebra
-Random.seed!(1234)
-μ = [0 1.0 -2.0] # MAP coefficients
-Σ = Symmetric(reshape(randn(9),3,3).*0.1 + UniformScaling(1.0)) # MAP covariance matrix
-M = CounterfactualExplanations.Models.BayesianLogisticModel(μ, Σ);
-```
-
-See also:
-- [`logits(M::BayesianLogisticModel, X::AbstractArray)`](@ref)
-- [`probs(M::BayesianLogisticModel, X::AbstractArray)`](@ref)
-"""
-struct BayesianLogisticModel <: AbstractDifferentiableJuliaModel
-    μ::Matrix
-    Σ::Matrix
-    likelihood::Symbol
[…]
