+++ src/models/differentiable/differentiable.jl
@@ -9,6 +9,5 @@ Base type for differentiable models written in pure Julia.
 """
 abstract type AbstractDifferentiableJuliaModel <: AbstractDifferentiableModel end
 
-include("simple.jl")
 include("flux.jl")
 include("laplace-redux.jl")
