### 2022-12-20 adb7298
basic support for categorical variables
