+++ docs/setup_docs.jl
@@ -4,17 +4,22 @@ setup_docs = quote
     Pkg.activate("docs")
 
     using CounterfactualExplanations
+    using CounterfactualExplanations: counterfactual
     using Flux
     using Flux.Optimise: update!, Adam
     using LinearAlgebra
     using MLJBase
+    using MLJModels: OneHotEncoder
     using Plots
     using Random
+    using StatsBase
 
     theme(:wong)
     www_path = "docs/src/generators/www/"
     Random.seed!(2023)
 
     include("docs/src/utils.jl")
+
+    synthetic = CounterfactualExplanations.Data.load_synthetic()
 
 end
