+++ src/data/functions.jl
@@ -1,5 +1,5 @@
 using BSON
-using CounterfactualExplanations # is this bad?
+using CounterfactualExplanations 
 using Flux
 using LazyArtifacts
 using Random
