+++ docs/Project.toml
@@ -22,6 +22,7 @@ LaplaceRedux = "c52c1a26-f7c5-402b-80be-ba1e638ad478"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 MLDatasets = "eb30cadb-4394-5ae3-aed4-317e484a6458"
 MLJ = "add582a8-e3ab-11e8-2d5e-e98b27df1bc7"
+MLJModels = "d491faf4-2d78-11e9-2867-c94bc002c0b7"
 MLUtils = "f1d291b0-491e-4a28-83b9-f70985020b54"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 OhMyREPL = "5fb14364-9ced-5910-84b2-373655c76a03"
