+++ test/plotting.jl
@@ -1,16 +1,17 @@
 using MLUtils
 using Plots
 
-M = LogisticModel([1 1], [0])
+M = synthetic[:classification_binary][:models][:flux][:model]
+
 generator = generator_catalog[:generic]()
 
 @testset "Two-dimensional" begin
 
-    X, ys = toy_data_linear()
-    X = MLUtils.stack(X, dims = 2)
+    xs, ys = (synthetic[:classification_binary][:data][:xs], synthetic[:classification_binary][:data][:ys])
+    X = MLUtils.stack(xs, dims = 2)
+    counterfactual_data = CounterfactualData(X, ys')
 
     # Model:
-    counterfactual_data = CounterfactualData(X, ys')
     plt = plot(M, counterfactual_data)
 
     # Counterfactual:
@@ -26,12 +27,13 @@ end
 
 @testset "Multi-dimensional" begin
 
-    # Model:
-    X, ys = toy_data_linear()
-    X = MLUtils.stack(X, dims = 2)
-    X = vcat(X, randn(1, size(X, 2)))
-    counterfactual_data = CounterfactualData(X, ys')
-    M = LogisticModel([1 1 0], [0])
-    plt = plot(M, counterfactual_data)
+    # # Model:
+    # xs, ys = (synthetic[:classification_binary][:data][:xs], synthetic[:classification_binary][:data][:ys])
+    # X = MLUtils.stack(xs, dims = 2)
+    # X = vcat(X, randn(1, size(X, 2)))
+    # counterfactual_data = CounterfactualData(X, ys')
+
+    # M = LogisticModel([1 1 0], [0])
+    # plt = plot(M, counterfactual_data)
 
 end
