+++ test/generative_models.jl
@@ -1,11 +1,12 @@
 using MLUtils
 
 # Setup
-M = LogisticModel([1 1], [0])
+M = synthetic[:classification_binary][:models][:flux][:model]
 generator = generator_catalog[:revise]()
-X, ys = toy_data_linear()
-X = MLUtils.stack(X, dims = 2)
+xs, ys = (synthetic[:classification_binary][:data][:xs], synthetic[:classification_binary][:data][:ys])
+X = MLUtils.stack(xs, dims = 2)
 counterfactual_data = CounterfactualData(X, ys')
+
 # Coutnerfactual search
 x = select_factual(counterfactual_data, rand(1:size(X, 2)))
 p_ = probs(M, x)
