### 2022-12-20 959a1d9
this seems to work well, but need to think about mutability constraints
