### 2022-12-21 #103 @pat-alt
Will do, thanks. Need to finish some work on a (related) research paper before the holidays and would then turn to this first thing in the new year. Does that work?
