### 2022-12-21 fd0e730
small bug wrt decoding
