+++ docs/src/generators/gradient_based/latent_space_generator.qmd
@@ -103,10 +103,9 @@ We then draw a random factual and generate two counterfactuals - one using gener
 ```{julia}
 counterfactual_data = CounterfactualData(X,ys)
 using Random
-# Random.seed!(123)
 x = select_factual(counterfactual_data, rand(1:size(X)[2])) 
 y = round(probs(M, x)[1])
-target = ifelse(y==1.0,1.0,0.0) # opposite label as target
+target = ifelse(y==1.0,0.0,1.0) # opposite label as target
 ```
 
 ```{julia}
@@ -114,9 +113,14 @@ target = ifelse(y==1.0,1.0,0.0) # opposite label as target
 generator = GenericGenerator()
 # Generate recourse:
 ce = generate_counterfactual(x, target, counterfactual_data, M, generator, latent_space=false)
-# ce_latent = generate_counterfactual(x, target, counterfactual_data, M, generator, latent_space=true; generative_model_params=(latent_dim=4, epochs=200))
-ce_latent = generate_counterfactual(x, target, counterfactual_data, M, generator, latent_space=true)
+ce_latent = generate_counterfactual(x, target, counterfactual_data, M, generator, latent_space=true; generative_model_params=(latent_dim=5, epochs=200))
 ```
+
+```{julia}
+generator = REVISEGenerator()
+ce_latent = generate_counterfactual(x, target, counterfactual_data, M, generator; generative_model_params=(latent_dim=5, epochs=200))
+```
+
 
 ```{julia}
 #| echo: false
