+++ src/generators/gradient_based/ClapROARGenerator.jl
@@ -5,14 +5,14 @@ mutable struct ClapROARGenerator <: AbstractGradientBasedGenerator
     complexity::Function # complexity function
     λ::Union{AbstractFloat,AbstractVector} # strength of penalty
     decision_threshold::Union{Nothing,AbstractFloat}
-    opt::Any # optimizer
+    opt::Flux.Optimise.AbstractOptimiser # optimizer
     τ::AbstractFloat # tolerance for convergence
 end
 
 # API streamlining:
 using Parameters, Flux
 @with_kw struct ClapROARGeneratorParams
-    opt::Any = Flux.Optimise.Descent()
+    opt::Flux.Optimise.AbstractOptimiser = Descent()
     τ::AbstractFloat = 1e-3
 end
 
@@ -22,7 +22,7 @@ end
         loss::Symbol=:logitbinarycrossentropy,
         complexity::Function=norm,
         λ::AbstractFloat=0.1,
-        opt::Any=Flux.Optimise.Descent(),
+        opt::Flux.Optimise.AbstractOptimiser=Flux.Optimise.Descent(),
         τ::AbstractFloat=1e-5
     )
 
