+++ v0.1.5/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"resources/resources.commonmark/#Further-Resources","page":"-","title":"Further Resources","text":"","category":"section"},{"location":"resources/resources.commonmark/#JuliaCon-2022","page":"-","title":"JuliaCon 2022","text":"","category":"section"},{"location":"resources/resources.commonmark/","page":"-","title":"-","text":"Slides: link","category":"page"},{"location":"resources/resources.commonmark/","page":"-","title":"-","text":"@raw html <iframe style=\"width:560px;height:315px\" src=\"https://www.paltmeyer.com/CounterfactualExplanations.jl/dev/resources/juliacon22/presentation.html\" title=\"YouTube video player\" frameborder=\"0\" allow=\"accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture\" allowfullscreen></iframe>","category":"page"},{"location":"tutorials/generators/","page":"Custom generators","title":"Custom generators","text":"CurrentModule = CounterfactualExplanations ","category":"page"},{"location":"tutorials/generators/#Custom-generators","page":"Custom generators","title":"Custom generators","text":"","category":"section"},{"location":"tutorials/generators/","page":"Custom generators","title":"Custom generators","text":"As we will see in this short tutorial, building custom counterfactual generators is straight-forward. We hope that this will facilitate contributions through the community.","category":"page"},{"location":"tutorials/generators/#Generic-generator-with-dropout","page":"Custom generators","title":"Generic generator with dropout","text":"","category":"section"},{"location":"tutorials/generators/","page":"Custom generators","title":"Custom generators","text":"To illustrate how custom generators can be implemented we will consider a simple example of a generator that extends the functionality of our GenericGenerator. We have noted elsewhere that the effectiveness of ounterfactual explanations depends to some degree on the quality of the fitted model. Another, perhaps trivial, thing to note is that counterfactual explanations are not unique: there are potentially many valid counterfactual paths. One interesting (or silly) idea following these two observations might be to introduce some form of regularization in the counterfactual search. For example, we could use dropout to randomly switch features on and off in each iteration. Without dwelling further on the usefulness of this idea, let us see how it can be implemented.","category":"page"},{"location":"tutorials/generators/","page":"Custom generators","title":"Custom generators","text":"The first code chunk below implements two important steps: 1) create an abstract subtype of the AbstractGradientBasedGenerator and 2) create a constructor similar to the GenericConstructor, but with one additional field for the probability of dropout.","category":"page"},{"location":"tutorials/generators/","page":"Custom generators","title":"Custom generators","text":"# Abstract suptype:\nabstract type AbstractDropoutGenerator <: AbstractGradientBasedGenerator end\n\n# Constructor:\nstruct DropoutGenerator <: AbstractDropoutGenerator\n    loss::Symbol # loss function\n    complexity::Function # complexity function\n    mutability::Union{Nothing,Vector{Symbol}} # mutibility constraints \n    λ::AbstractFloat # strength of penalty\n    ϵ::AbstractFloat # step size\n    τ::AbstractFloat # tolerance for convergence\n    p_dropout::AbstractFloat # dropout rate\nend\n\n# Instantiate:\nusing LinearAlgebra\ngenerator = DropoutGenerator(\n    :logitbinarycrossentropy,\n    norm,\n    nothing,\n    0.1,\n    0.1,\n    1e-5,\n    0.5\n)","c
[…]
