+++ dev/logo/logo.jl
@@ -1,0 +1,176 @@
+using Pkg;
+Pkg.activate("dev");
+
+using Colors
+using CounterfactualExplanations
+using Flux
+using Luxor
+using MLJBase
+using StatsBase: sample
+using Random
+
+const julia_colors = Dict(
+    :blue => Luxor.julia_blue,
+    :red => Luxor.julia_red,
+    :green => Luxor.julia_green,
+    :purple => Luxor.julia_purple,
+)
+
+function get_data(N=500; center_box=(-5 => 5), cluster_std=0.5)
+
+    # Data:
+    x, y = make_blobs(N, 2; centers=2, as_table=false, center_box=center_box, cluster_std=cluster_std)
+    y .= y .== 2
+    x = permutedims(x)
+    xs = Flux.unstack(x, 2)
+    data = zip(xs, y)
+    counterfactual_data = CounterfactualData(x, permutedims(y))
+
+    # Model:
+    nn = Chain(Dense(2, 1))
+    opt = Adam()
+    epochs = 100
+    loss(x, y) = Flux.Losses.logitbinarycrossentropy(nn(x), y)
+    avg_loss(data) = mean(map(d -> loss(d[1], d[2]), data))
+    for epoch = 1:epochs
+        for d in data
+            gs = gradient(Flux.params(nn)) do
+                l = loss(d...)
+            end
+            Flux.update!(opt, Flux.params(nn), gs)
+        end
+    end
+    M = FluxModel(nn)
+
+    return x, y, counterfactual_data, M
+end
+
+function logo_picture(;
+    ndots=3,
+    frame_size=500,
+    ms=frame_size // 10,
+    mcolor=(:red, :green, :purple),
+    margin=0.1,
+    fun=f(x) = x * cos(x),
+    xmax=2.5,
+    noise=0.5,
+    ged_data=get_data,
+    ntrue=50,
+    gt_color=julia_colors[:blue],
+    gt_stroke_size=5,
+    interval_color=julia_colors[:blue],
+    interval_alpha=0.2,
+    seed=2022
+)
+
+    # Setup
+    n_mcolor = length(mcolor)
+    mcolor = getindex.(Ref(julia_colors), mcolor)
+    Random.seed!(seed)
+
+    # Data
+    x, y, counterfactual_data, M = get_data()
+    factual = select_factual(counterfactual_data, rand(1:size(x)[2]))
+    factual_label = round(probs(M, factual)[1])
+    target = ifelse(factual_label == 1.0, 0.0, 1.0) # opposite label as target
+
+    # Counterfactual:
+    generator = GenericGenerator()
+    ce = generate_counterfactual(factual, target, counterfactual_data, M, generator)
+
+    _scale = (frame_size / (2 * maximum(x))) * (1 - margin)
+
+    # Decision Boundary:
+    setline(gt_stroke_size)
+    sethue(gt_color)
+    w = collect(Flux.params(M.model))[1]
+    b = collect(Flux.params(M.model))[2][1]
+    a = -w[2]/w[1]
+    _ymin =  _scale .* (b + a * (-xmax))
+    _ymax =  _scale .* (b + a * (xmax))
+    pmin = Point(_scale * (-xmax), _ymin)
+    pmax = Point(_scale * xmax, _ymax)
+    # line(pmin, pmax, action=:stroke)
+
+    # Data
+    # data_plot = zip(xplot, yplot)
+    # for i = 1:length(data_plot)
+    #     _x, _y = _scale .* collect(data_plot)[i]
+    #     color_idx = i % n_mcolor == 0 ? n_mcolor : i % n_mcolor
+    #     sethue(mcolor[color_idx]...)
[…]
