+++ dev/Project.toml
@@ -1,5 +1,6 @@
 [deps]
 BSON = "fbb218c0-5317-5bc6-957e-2ee96dd4b1f0"
+Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
 CounterfactualExplanations = "2f13d31b-18db-44c1-bc43-ebaf2cff0be0"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
@@ -8,7 +9,9 @@ GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
 GraphRecipes = "bd48cda9-67a9-57be-86fa-5b3c104eda73"
 Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
 LibGit2 = "76f85450-5226-5b5a-8eaa-529ad045b433"
+Luxor = "ae8d54c2-7ccd-5906-9d76-62fc9837b5bc"
 MLDatasets = "eb30cadb-4394-5ae3-aed4-317e484a6458"
+MLJBase = "a7f614a8-145f-11e9-1d2a-a57a1082229d"
 MLUtils = "f1d291b0-491e-4a28-83b9-f70985020b54"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
@@ -18,6 +21,8 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 RCall = "6f49c342-dc21-5d91-9882-a32aef131414"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 ghr_jll = "07c12ed4-43bc-5495-8a2a-d5838ef8d533"
 
 [compat]
