### 2023-01-18 ce5c773
basic functions for data loading
