+++ dev/artifacts/setup.jl
@@ -1,0 +1,12 @@
+setup = quote
+
+    using Pkg; Pkg.activate("dev/artifacts")
+
+    # Deps:
+    using CounterfactualExplanations
+    using CSV
+
+    # Utils
+    include("dev/artifacts/utils.jl")
+
+end
