+++ src/data/tabular.jl
@@ -29,19 +29,22 @@ function load_credit_default()
     y = df.target
 
     # Categorical encoding:
+    df.SEX = categorical(df.SEX)
+    df.EDUCATION = categorical(df.EDUCATION)
+    df.MARRIAGE = categorical(df.MARRIAGE)
     transformer = ContinuousEncoder()
     mach = MLJBase.fit!(machine(transformer, df[:, Not(:target)]))
     X = MLJBase.transform(mach, df[:, Not(:target)])
     X = permutedims(Matrix(X))
     features_categorical = [
-        [2,3],
-        collect(4:10),
-        collect(11:14)
+        [2, 3],             # SEX
+        collect(4:10),      # EDUCATION
+        collect(11:14)      # MARRIAGE
     ]
 
     counterfactual_data = CounterfactualData(
         X, y;
-        features_categorical = features_categorical
+        features_categorical=features_categorical
     )
 
     return counterfactual_data
