+++ src/data/Data.jl
@@ -22,6 +22,7 @@ const data_catalogue = Dict(
         :overlapping => load_overlapping,
         :moons => load_moons,
         :circles => load_circles,
+        :multi_class => load_multi_class,
     ),
     :tabular => Dict(
         :california_housing => load_california_housing,
@@ -34,7 +35,7 @@ const data_catalogue = Dict(
 )
 
 export data_catalogue
-export load_linearly_separable, load_overlapping, load_circles, load_moons
+export load_linearly_separable, load_overlapping, load_circles, load_moons, load_multi_class, load_synthetic_data
 export load_california_housing, load_credit_default, load_gmsc
 export load_mnist
 
