+++ src/models/Models.jl
@@ -36,7 +36,7 @@ include("plotting.jl")
 A dictionary containing all trainable machine learning models.
 """
 const model_catalogue = Dict(
-    :LogisticRegression => LogisticRegression,
+    :Linear => Linear,
     :MLP => FluxModel,
     :DeepEnsemble => FluxEnsemble,
 )
