+++ test/models.jl
@@ -5,21 +5,17 @@ using LinearAlgebra
 using MLUtils
 using Random
 
-### Models for synthetic data
-synthetic = CounterfactualExplanations.Data.load_synthetic([:flux])
-
 @testset "Models for synthetic data" begin
     for (key, value) ∈ synthetic
         name = string(key)
         @testset "$name" begin
-            X = value[:data][:xs]
-            X = MLUtils.stack(X, dims = 2)
+            X = value[:data].X
             for (likelihood, model) ∈ value[:models]
                 name = string(likelihood)
                 @testset "$name" begin
                     @testset "Matrix of inputs" begin
-                        @test size(logits(model[:model], X))[2] == size(X)[2]
-                        @test size(probs(model[:model], X))[2] == size(X)[2]
+                        @test size(logits(model[:model], X))[2] == size(X,2)
+                        @test size(probs(model[:model], X))[2] == size(X,2)
                     end
                     @testset "Vector of inputs" begin
                         @test size(logits(model[:model], X[:, 1]), 2) == 1
