+++ src/data_preprocessing/DataPreprocessing.jl
@@ -5,7 +5,7 @@ include("functions.jl")
 export CounterfactualData, select_factual, apply_domain_constraints
 
 include("plotting.jl")
-
+include("utils.jl")
 
 
 end
