### 2023-01-21 2808118
basically done, also largely done with support for supplying y as categorical vector
