### 2023-01-24 08495a5
more progress on this. Work on allowing strings and symbols as target input
