+++ src/counterfactuals/plotting.jl
@@ -125,11 +125,10 @@ function set_up_plots(
     plot_proba,
     kwargs...,
 )
-    target_idx = get_target_index(counterfactual_explanation.data.y_levels, counterfactual_explanation.target) 
     p1 = Models.plot(
         counterfactual_explanation.M,
         counterfactual_explanation.data;
-        target = target_idx,
+        target = counterfactual_explanation.target_encoded,
         alpha = alpha,
         kwargs...,
     )
