+++ src/data_preprocessing/DataPreprocessing.jl
@@ -2,7 +2,7 @@ module DataPreprocessing
 
 using ..GenerativeModels
 include("functions.jl")
-export CounterfactualData, select_factual, apply_domain_constraints
+export CounterfactualData, select_factual, apply_domain_constraints, OutputEncoder, _transform
 
 include("plotting.jl")
 include("utils.jl")

[…]
