+++ src/data/Data.jl
@@ -11,7 +11,6 @@ export load_synthetic, toy_data_linear, toy_data_multi, toy_data_non_linear
 export mnist_data, mnist_ensemble, mnist_model, mnist_vae
 export cats_dogs_data, cats_dogs_model
 
-include("functions.jl")
 include("synthetic.jl")
 include("tabular.jl")
 include("vision.jl")
