+++ dev/logo/logo.jl
@@ -18,16 +18,16 @@ const julia_colors = Dict(
     :purple => Luxor.julia_purple,
 )
 
-function get_data(N=500; center_box=(-5 => 5), cluster_std=0.5)
+function get_data(N=500; seed=42)
 
-    counterfactual_data = load_linearly_separable(N)
+    counterfactual_data = load_linearly_separable(N; seed=seed)
     M = fit_model(counterfactual_data, :Linear)
 
     return counterfactual_data, M
 end
 
 function logo_picture(;
-    ndots=3,
+    ndots=20,
     frame_size=500,
     ms=frame_size // 10,
     mcolor=(:red, :green, :purple),
@@ -50,8 +50,11 @@ function logo_picture(;
     Random.seed!(seed)
 
     # Data
-    counterfactual_data, M = get_data()
-    factual = select_factual(counterfactual_data, rand(1:size(counterfactual_data.X,2)))
+    counterfactual_data, M = get_data(seed=seed)
+    X = counterfactual_data.X
+    x = X[1,:]
+    y = X[2,:]
+    factual = select_factual(counterfactual_data, rand(1:size(counterfactual_data.X, 2)))
     factual_label = predict_label(M, counterfactual_data, factual)[1]
     target = ifelse(factual_label == 1.0, 2.0, 1.0) # opposite label as target
 
@@ -59,36 +62,43 @@ function logo_picture(;
     generator = GenericGenerator()
     ce = generate_counterfactual(factual, target, counterfactual_data, M, generator)
 
-    _scale = (frame_size / (2 * maximum(counterfactual_data.X))) * (1 - margin)
+    # Dots:
+    idx = sample(1:size(X,2), ndots, replace=false)
+    xplot, yplot = (x[idx], y[idx])
+    _scale = (frame_size / (5 * maximum(counterfactual_data.X))) * (1 - margin)
 
     # Decision Boundary:
     setline(gt_stroke_size)
     sethue(gt_color)
     w = collect(Flux.params(M.model))[1]
-    b = 0
     a = -w[1] / w[2]
-    _ymin = _scale .* (b + a * (-xmax))
-    _ymax = _scale .* (b + a * (xmax))
-    pmin = Point(_scale * (-xmax), _ymin)
-    pmax = Point(_scale * xmax, _ymax)
+    rule(Point(0, 0), atan(a))
+
+    # _ymin = _scale .* (b + a * (-xmax))
+    # _ymax = _scale .* (b + a * (xmax))
+    # pmin = Point(_scale * (-xmax), _ymin)
+    # pmax = Point(_scale * xmax, _ymax)
     # line(pmin, pmax, action=:stroke)
 
     # Data
-    # data_plot = zip(xplot, yplot)
-    # for i = 1:length(data_plot)
-    #     _x, _y = _scale .* collect(data_plot)[i]
-    #     color_idx = i % n_mcolor == 0 ? n_mcolor : i % n_mcolor
-    #     sethue(mcolor[color_idx]...)
-    #     circle(Point(_x, _y), ms, action=:fill)
-    # end
+    data_plot = zip(xplot, yplot)
+    for i = 1:length(data_plot)
+        _point = collect(data_plot)[i]
+        _lab = predict_label(M, counterfactual_data, [_point[1], _point[2]])
+        print(_lab)
+        color_idx = get_target_index(counterfactual_data.y_levels, _lab)
+        _x, _y = _scale .* _point
+        sethue(mcolor[color_idx]...)
+        circle(Point(_x, _y), ms, action=:fill)
+    end
 
 end
 
-function draw_small_logo(filename="docs/src/assets/logo.svg"; width=500)
[…]
