### 2023-01-24 fc678b8
progress on logo
