### 2023-01-25 653ca43
CompatHelper: add new compat entry for MLJModels at version 0.16, (keep existing compat)
