+++ Project.toml
@@ -33,6 +33,7 @@ DataFrames = "1"
 Flux = "0.12, 0.13"
 LaplaceRedux = "0.1"
 MLJBase = "0.21.3"
+MLJModels = "0.16"
 MLUtils = "0.2, 0.3, 0.4"
 MultivariateStats = "0.9, 0.10"
 NearestNeighborModels = "0.2"
