### 2023-01-25 7d6a18f
CompatHelper: add new compat entry for CSV at version 0.10, (keep existing compat)
