### 2023-01-25 2c31656
CompatHelper: add new compat entry for Tables at version 1, (keep existing compat)
