+++ Project.toml
@@ -40,6 +40,7 @@ Parameters = "0.12"
 Plots = "1.38.2"
 SliceMap = "0.2"
 StatsBase = "0.33"
+Tables = "1"
 UMAP = "0.1"
 julia = "1.7, 1.8"
 
