### 2023-01-25 82010ed
CompatHelper: add new compat entry for CategoricalArrays at version 0.10, (keep existing compat)
