+++ Project.toml
@@ -32,6 +32,7 @@ CUDA = "3"
 DataFrames = "1"
 Flux = "0.12, 0.13"
 LaplaceRedux = "0.1"
+MLDatasets = "0.7"
 MLJBase = "0.21.3"
 MLUtils = "0.2, 0.3, 0.4"
 MultivariateStats = "0.9, 0.10"
