+++ dev/logo/logo.jl
@@ -18,30 +18,28 @@ julia_colors = Dict(
     :purple => Luxor.julia_purple,
 )
 
-function get_data(N=500; seed=42)
+function get_data(N=500; seed=1234, centers=2, center_box=(-2 => 2), cluster_std=0.3)
 
-    counterfactual_data = load_linearly_separable(N; seed=seed)
+    counterfactual_data = load_blobs(N; seed=seed, centers=centers, center_box=center_box, cluster_std=cluster_std)
     M = fit_model(counterfactual_data, :Linear)
 
     return counterfactual_data, M
 end
 
 function logo_picture(;
-    ndots=20,
+    ndots=100,
     frame_size=500,
-    ms=frame_size // 30,
-    mcolor=(:red, :green, :purple),
-    margin=0.5,
-    fun=f(x) = x * cos(x),
-    xmax=2.5,
-    noise=0.5,
-    ged_data=get_data,
-    ntrue=50,
-    gt_color=julia_colors[:blue],
-    gt_stroke_size=5,
-    interval_color=julia_colors[:blue],
-    interval_alpha=0.2,
-    seed=2022
+    ms=frame_size // 50,
+    mcolor=(:red, :green),
+    margin=-0.1,
+    db_color=julia_colors[:purple],
+    db_stroke_size=frame_size // 50,
+    ce_color=julia_colors[:purple],
+    m_alpha=0.2,
+    seed=2022,
+    cluster_std=0.3,
+    γ=0.9,
+    η=0.02
 )
 
     # Setup
@@ -50,30 +48,35 @@ function logo_picture(;
     Random.seed!(seed)
 
     # Data
-    counterfactual_data, M = get_data(seed=seed)
+    counterfactual_data, M = get_data(seed=seed, cluster_std=cluster_std)
     X = counterfactual_data.X
-    x = X[1,:]
-    y = X[2,:]
+    x = X[1, :]
+    y = X[2, :]
     factual = select_factual(counterfactual_data, rand(1:size(counterfactual_data.X, 2)))
     factual_label = predict_label(M, counterfactual_data, factual)[1]
     target = ifelse(factual_label == 1.0, 2.0, 1.0) # opposite label as target
 
     # Counterfactual:
-    generator = GenericGenerator(opt=Flux.Descent(0.01))
+    generator = GravitationalGenerator(
+        opt=Flux.Descent(η),
+        decision_threshold=γ
+    )
     ce = generate_counterfactual(factual, target, counterfactual_data, M, generator)
 
     # Dots:
-    idx = sample(1:size(X,2), ndots, replace=false)
+    idx = sample(1:size(X, 2), ndots, replace=false)
     xplot, yplot = (x[idx], y[idx])
     _scale = (frame_size / (2 * maximum(abs.(counterfactual_data.X)))) * (1 - margin)
 
     # Decision Boundary:
-    setline(gt_stroke_size)
-    sethue(gt_color)
+    setline(db_stroke_size)
+    sethue(db_color)
     w = collect(Flux.params(M.model))[1]
     a = -w[1] / w[2]
-    b = collect(Flux.params(M.model))[2][1]
-    rule(Point(0, b), atan(a))
+    _bias = collect(Flux.params(M.model))[2][1]
+    b = -_bias / w[2]
+    _intercept = _scale .* (0, b)
+    rule(Point(_intercept[1], _intercept[2]), atan(a))
 
     # Data
     data_plot = zip(xplot, yplot)
@@ -82,36 +85,44 @@ function logo_picture(;
         _lab = predict_label(M, counterfactual_data, [_point[1], _point[2]])
         color_idx = get_target_index(counterfactual_data.y_levels, _lab[1])
         _x, _y = _scale .* _point
[…]
