+++ docs/src/_intro.qmd
@@ -1,0 +1,128 @@
+`CounterfactualExplanations.jl` is a package for generating Counterfactual Explanations (CE) and Algorithmic Recourse (AR) for black-box algorithms. Both CE and AR are related tools for explainable artificial intelligence (XAI). While the package is written purely in Julia, it can be used to explain machine learning algorithms developed and trained in other popular programming languages like Python and R. See below for a short introduction and other resources or dive straight into the [docs](https://pat-alt.github.io/CounterfactualExplanations.jl/dev).
+
+## Installation 🚩
+
+You can install the stable release from [Julia's General Registry](https://github.com/JuliaRegistries/General) as follows:
+
+```julia
+using Pkg
+Pkg.add("CounterfactualExplanations")
+```
+
+`CounterfactualExplanations.jl` is under active development. To install the development version of the package you can run the following command:
+
+```julia
+using Pkg
+Pkg.add(url="https://github.com/pat-alt/CounterfactualExplanations.jl")
+```
+
+## Background and Motivation 
+
+#### The Need for Explainability ⬛
+
+Machine learning models like deep neural networks have become so complex, opaque and underspecified in the data that they are generally considered Black Boxes. Nonetheless, such models often play a key role in data-driven decision-making systems. This often creates the following problem: human operators in charge of such systems have to rely on them blindly, while those individuals subject to them generally have no way of challenging an undesirable outcome:
+
+> “You cannot appeal to (algorithms). They do not listen. Nor do they bend.”
+>
+> — Cathy O’Neil in [*Weapons of Math Destruction*](https://en.wikipedia.org/wiki/Weapons_of_Math_Destruction), 2016
+
+#### Enter: Counterfactual Explanations 🔮
+
+Counterfactual Explanations can help human stakeholders make sense of the systems they develop, use or endure: they explain how inputs into a system need to change for it to produce different decisions. Explainability benefits internal as well as external quality assurance. The figure below, for example, shows various counterfactuals generated through different approaches that all turn the predicted label of some classifier from a 9 into a 4. 
+
+
+![Realistic counterfactual explanations for MNIST data: turning a 4 into a 9.](https://raw.githubusercontent.com/pat-alt/CounterfactualExplanations.jl/main/docs/src/examples/image/www/MNIST_9to4.png)
+
+Explanations that involve realistic and actionable changes can be used for Algorithmic Recourse (AR): they offer human stakeholders a way to not only understand the system's behaviour but also react to it or adjust it. The figure below illustrates the point of AR through a toy example: it shows the counterfactual path of one sad cat 🐱 that would like to be grouped with her cool dog friends. Unfortunately, based on her tail length and height she was classified as a cat by a black-box classifier. The recourse algorithm perturbs her features in such a way that she ends up crossing the decision boundary into a dense region inside the target class.
+
+![A sad 🐱 on its counterfactual path to its cool dog friends.](https://raw.githubusercontent.com/pat-alt/CounterfactualExplanations.jl/main/docs/src/www/recourse_laplace.gif)
+
+Counterfactual Explanations have certain advantages over related tools for explainable artificial intelligence (XAI) like surrogate explainers (LIME and SHAP). These include:
+
+- Full fidelity to the black-box model, since no proxy is involved. 
+- No need for (reasonably) interpretable features as opposed to LIME and SHAP.
+- Clear link to Causal Inference and Bayesian Machine Learning.
+- Less susceptible to adversarial attacks than LIME and SHAP.
+
+## Usage example 🔍
+
+Generating counterfactuals will typically look like follows:
+
+```{julia}
[…]
