+++ docs/src/index.qmd
@@ -6,4 +6,4 @@ CurrentModule = CounterfactualExplanations
 
 Documentation for [CounterfactualExplanations.jl](https://github.com/pat-alt/CounterfactualExplanations.jl).
 
-{{< include intro.qmd >}}
+{{< include _intro.qmd >}}
