+++ paper/paper.tex
@@ -27,7 +27,7 @@ data-driven decision-making systems in real-world applications.
 
 As others have pointed out, this scenario gives rise to an undesirable
 principal-agent problem involving a group of principals - i.e.~human
-stakeholders - that fail to understand the behavior of their agent -
+stakeholders - that fail to understand the behaviour of their agent -
 i.e.~the black-box system \cite{borch2022machine}. The group of
 principals may include programmers, product managers and other
 decision-makers who develop and operate the system as well as those
@@ -47,19 +47,19 @@ stakeholders make sense of the systems they develop, use or endure: they
 explain how inputs into a system need to change for it to produce
 different decisions. Explainability benefits internal as well as
 external quality assurance. Explanations that involve realistic and
-actionable changes can be used for the purpose of Algorithmic Recourse
-(AR): they offer the group of principals a way to not only understand
-their agent's behavior, but also adjust or react to it.
+actionable changes can be used for Algorithmic Recourse (AR): they offer
+the group of principals a way to not only understand their agent's
+behaviour, but also adjust or react to it.
 
-The availability of open-source software for the purpose of explaining
-black-box models through counterfactuals is still limited. Most existing
-implementations are specific to particular methodologies. They are also
-exclusively built in Python and for Python models. The only existing
-unifying software approach, for example, is tailored to models built in
-the two most popular Python libraries for deep learning. The Julia
-ecosystem has so far lacked an open-source implementation of CE.
+The availability of open-source software to explain black-box models
+through counterfactuals is still limited. Most existing implementations
+are specific to particular methodologies. They are also exclusively
+built in Python and for Python models. The only existing unifying
+software approach, for example, is tailored to models built in the two
+most popular Python libraries for deep learning. The Julia ecosystem has
+so far lacked an open-source implementation of CE.
 
-Through the work presented here we aim to close that gap and thereby
+Through the work presented here, we aim to close that gap and thereby
 contribute to broader community efforts towards explainable AI. We
 envision this package to one day be the go-to place for Counterfactual
 Explanations in Julia. Thanks to Julia's unique support for
@@ -72,11 +72,12 @@ Counterfactual Explanations for differentiable classification models
 trained in Julia. It comes with detailed documentation involving various
 illustrative example datasets, linear and deep learning classifiers and
 counterfactual generators for binary and multi-class prediction tasks. A
-carefully designed package architecture allows for seamless extension of
-the package functionality through custom generators and models. Through
-simple examples we also demonstrate how to use our package to explain
-models that were built and trained in \texttt{Python} and \texttt{R},
-although at the time of writing this feature is still experimental.
+carefully designed package architecture allows for a seamless extension
+of the package functionality through custom generators and models.
+Through simple examples, we also demonstrate how to use our package to
+explain models that were built and trained in \texttt{Python} and
+\texttt{R}, although at the time of writing this feature is still
+experimental.
 
 The remainder of this article is structured as follows:
 Section~\ref{sec-related} presents related work on explainable AI as
@@ -124,7 +125,7 @@ often used interchangeably, but this can lead to confusion. We find the
 distinction made in \cite{rudin2019stop} useful: interpretable AI
 involves models that are inherently interpretable and transparent such
 as general additive models (GAM), decision trees and rule-based models;
-explainable AI may involve models that are not inherently interpretable,
[…]
