+++ src/models/plotting.jl
@@ -19,7 +19,7 @@ function Plots.plot(
     colorbar = true,
     title = "",
     length_out = 50,
-    zoom = -1,
+    zoom = -0.1,
     xlims = nothing,
     ylims = nothing,
     linewidth = 0.1,
@@ -39,6 +39,7 @@ function Plots.plot(
     X, y, multi_dim = DataPreprocessing.prepare_for_plotting(data; dim_red = dim_red)
 
     # Surface range:
+    zoom = zoom * maximum(abs.(X))
     if isnothing(xlims)
         xlims = (minimum(X[:, 1]), maximum(X[:, 1])) .+ (zoom, -zoom)
     else
