+++ paper/_paper.tex
@@ -145,7 +145,7 @@
 \date{}
 
 \begin{document}
-\ifdefined\Shaded\renewenvironment{Shaded}{\begin{tcolorbox}[breakable, interior hidden, enhanced, sharp corners, boxrule=0pt, borderline west={3pt}{0pt}{shadecolor}, frame hidden]}{\end{tcolorbox}}\fi
+\ifdefined\Shaded\renewenvironment{Shaded}{\begin{tcolorbox}[interior hidden, boxrule=0pt, enhanced, borderline west={3pt}{0pt}{shadecolor}, breakable, frame hidden, sharp corners]}{\end{tcolorbox}}\fi
 
 \hypertarget{sec-intro}{%
 \section{Introduction}\label{sec-intro}}
@@ -222,13 +222,13 @@ advanced usage examples. In Section~\ref{sec-custom} we demonstrate how
 the package functionality can be customized and extended. To provide a
 flavour of its practical use, we use the package to explain models
 trained on MNIST data in Section~\ref{sec-emp}. Finally, we also discuss
-current limitations of our package, as well as its future outlook in
+the current limitations of our package, as well as its future outlook in
 Section~\ref{sec-outlook}. Section~\ref{sec-conclude} concludes.
 
 \hypertarget{sec-related}{%
 \section{Background and related work}\label{sec-related}}
 
-In this section we first briefly introduce the broad field of
+In this section, we first briefly introduce the broad field of
 explainable artificial intelligence (XAI), before narrowing it down to
 Counterfactual Explanations. We introduce the methodological framework
 and finally point to existing open-source software.
@@ -477,10 +477,12 @@ counterfactual generators have been implemented in the package:
 \item Greedy \cite{schut2021generating}
 \item DiCE \cite{mothilal2020explaining}
 \item Latent Space Search as in REVISE \cite{joshi2019realistic} and CLUE \cite{antoran2020getting}
+\item ClaPROAR \cite{altmeyer2023endogenous}
+\item Gravitational \cite{altmeyer2023endogenous}
 \end{itemize}
 
-The package currently offers native support for deep learning models
-built in \href{https://fluxml.ai/}{Flux}.
+The package currently offers native support for models built and trained
+in \href{https://fluxml.ai/}{Flux}.
 
 \begin{figure}
 
