+++ paper/_paper.qmd
@@ -38,11 +38,11 @@ Through the work presented here, we aim to close that gap and thereby contribute
 
 Our package provides a simple and intuitive interface to generate Counterfactual Explanations for differentiable classification models trained in Julia. It comes with detailed documentation involving various illustrative example datasets, linear and deep learning classifiers and counterfactual generators for binary and multi-class prediction tasks. A carefully designed package architecture allows for a seamless extension of the package functionality through custom generators and models. Through simple examples, we also demonstrate how to use our package to explain models that were built and trained in `Python` and `R`, although at the time of writing this feature is still experimental.
 
-The remainder of this article is structured as follows: @sec-related presents related work on explainable AI as well as a brief overview of the methodological framework underlying CE. @sec-arch introduces the Julia package and its high-level architecture. @sec-use then presents a number of basic and advanced usage examples. In @sec-custom we demonstrate how the package functionality can be customized and extended. To provide a flavour of its practical use, we use the package to explain models trained on MNIST data in @sec-emp. Finally, we also discuss current limitations of our package, as well as its future outlook in @sec-outlook. @sec-conclude concludes.
+The remainder of this article is structured as follows: @sec-related presents related work on explainable AI as well as a brief overview of the methodological framework underlying CE. @sec-arch introduces the Julia package and its high-level architecture. @sec-use then presents a number of basic and advanced usage examples. In @sec-custom we demonstrate how the package functionality can be customized and extended. To provide a flavour of its practical use, we use the package to explain models trained on MNIST data in @sec-emp. Finally, we also discuss the current limitations of our package, as well as its future outlook in @sec-outlook. @sec-conclude concludes.
 
 # Background and related work {#sec-related}
 
-In this section we first briefly introduce the broad field of explainable artificial intelligence (XAI), before narrowing it down to Counterfactual Explanations. We introduce the methodological framework and finally point to existing open-source software.
+In this section, we first briefly introduce the broad field of explainable artificial intelligence (XAI), before narrowing it down to Counterfactual Explanations. We introduce the methodological framework and finally point to existing open-source software.
 
 ## Literature on explainable AI
 
@@ -87,9 +87,11 @@ Generally speaking, software development in the space of XAI has largely focused
 \item Greedy \cite{schut2021generating}
 \item DiCE \cite{mothilal2020explaining}
 \item Latent Space Search as in REVISE \cite{joshi2019realistic} and CLUE \cite{antoran2020getting}
+\item ClaPROAR \cite{altmeyer2023endogenous}
+\item Gravitational \cite{altmeyer2023endogenous}
 \end{itemize}
 
-The package currently offers native support for deep learning models built in [Flux](https://fluxml.ai/). 
+The package currently offers native support for models built and trained in [Flux](https://fluxml.ai/). 
 
 ![High-level schematic overview of package architecture. Modules are shown in red, structs in green and functions in blue.](www/pkg_architecture.png){#fig-arch width=20pc height=12pc}
 
@@ -117,7 +119,7 @@ x = select_factual(counterfactual_data, chosen)
 
 # Counterfactual search:
 generator = GenericGenerator()
-counterfactual = generate_counterfactual(
+ce = generate_counterfactual(
     x, target, counterfactual_data, M, generator)
[…]
