### 2023-01-25 899bcbc
for MNIST should still probably work with artifacts, just use serialize instead
