+++ src/data/vision.jl
@@ -12,6 +12,6 @@ function load_mnist()
     X, y = MNIST(:train)[:]
     X = Flux.flatten(X)
     y = categorical(y)
-    counterfactual_data = CounterfactualData(X, y)
+    counterfactual_data = CounterfactualData(X, y; domain=(0,1))
     return counterfactual_data
 end
