+++ dev/artifacts/setup.jl
@@ -1,9 +1,12 @@
 setup = quote
 
-    using Pkg; Pkg.activate("dev/artifacts")
+    using Pkg
+    Pkg.activate("dev/artifacts")
 
     # Deps:
     using CounterfactualExplanations
+    using CounterfactualExplanations.Models
+    using CounterfactualExplanations.Data
     using CSV
 
     # Utils
