+++ src/models/model_utils.jl
@@ -6,16 +6,14 @@ using MLJBase
 
 Prepares counterfactual data for training in Flux.
 """
-function data_loader(data::CounterfactualData)
+function data_loader(data::CounterfactualData; batchsize=1)
     X, y = CounterfactualExplanations.DataPreprocessing.unpack_data(data)
-    xs = MLUtils.unstack(X, dims=2)
     output_dim = length(unique(y))
     if output_dim > 2
-        y = Flux.onehotbatch(y, sort(unique(y)))
-        y = Flux.unstack(y; dims=3)
+        y = data.output_encoder.y   # get raw outputs
+        y = Flux.onehotbatch(y, data.y_levels)
     end
-    data = zip(xs, y)
-    return data
+    return DataLoader((X, y), batchsize=batchsize)
 end
 
 """
@@ -24,18 +22,10 @@ end
 Helper function to compute F-Score for `AbstractFittedModel` on a (test) data set.
 """
 function model_evaluation(M::AbstractFittedModel, test_data::CounterfactualData)
-    X, y = CounterfactualExplanations.DataPreprocessing.unpack_data(test_data)
+    y = test_data.output_encoder.y
     m = MulticlassFScore()
-    binary = M.likelihood == :classification_binary
-    if binary
-        proba = reduce(hcat, map(x -> binary ? [1 - x, x] : x, probs(M, X)))
-        ŷ = Flux.onecold(proba, 0:1)
-    else
-        y = Flux.onecold(y, 1:size(y, 1))
-        ŷ = Flux.onecold(probs(M, X), sort(unique(y)))
-    end
+    ŷ = predict_label(M, test_data)
     fscore = m(ŷ, vec(y))
-
     return fscore
 end
 
