+++ src/models/Models.jl
@@ -66,6 +66,6 @@ function fit_model(
     return M
 end
 
-export model_catalogue, fit_model, model_evaluation, predict_label
+export model_catalogue, fit_model, model_evaluation, predict_label, reset!
 
 end
