+++ src/data/vision.jl
@@ -12,6 +12,19 @@ function load_mnist()
     X, y = MNIST(:train)[:]
     X = Flux.flatten(X)
     y = categorical(y)
-    counterfactual_data = CounterfactualData(X, y; domain=(0,1))
+    counterfactual_data = CounterfactualData(X, y; domain=(0, 1))
+    return counterfactual_data
+end
+
+"""
+    load_mnist_test()
+
+Loads and prepares MNIST test data.
+"""
+function load_mnist_test()
+    X, y = MNIST(:test)[:]
+    X = Flux.flatten(X)
+    y = categorical(y)
+    counterfactual_data = CounterfactualData(X, y; domain=(0, 1))
     return counterfactual_data
 end
