### 2023-01-26 6cc8e5d
latent space stuff
