+++ src/global_utils.jl
@@ -1,4 +1,6 @@
 using CategoricalArrays
+using Flux
+using Parameters
 
 # Constants:
 """
@@ -85,4 +87,37 @@ Utility that returns the index of `target` in `y_levels`.
 function get_target_index(y_levels, target)
     @assert target in y_levels "Specified `target` variable does not match any values of `y`."
     findall(y_levels .== target)[1]
+end
+
+"""
+    FluxModelParams
+
+Default MLP training parameters.
+"""
+@with_kw mutable struct FluxModelParams
+    loss::Symbol = :logitbinarycrossentropy
+    opt::Symbol = :Adam
+    n_epochs::Int = 100
+    batchsize::Int = 1
+    verbose::Bool = false
+end
+
+"""
+    flux_training_params
+
+The default training parameter for `FluxModels` etc.
+"""
+const flux_training_params = FluxModelParams()
+
+"""
+    reset!(flux_training_params::FluxModelParams)
+
+Restores the default parameter values.
+"""
+function reset!(flux_training_params::FluxModelParams)
+    default_flux_training_params = FluxModelParams()
+    for _name in fieldnames(typeof(flux_training_params))
+        setfield!(flux_training_params, _name, getfield(default_flux_training_params, _name))
+    end
+    return flux_training_params
 end
