### 2023-01-27 1600e99
going to try to refactor #102
