+++ paper/_paper.qmd
@@ -581,7 +581,7 @@ models = Dict("MLP" => M, "Ensemble" => M_ens)
 using Flux: onecold
 
 # Specific image:
-function from_digit_to_digit(from::AbstractArray, to::Number, generator, model; γ=0.95, x=X, y=ys, seed=1234, T=1000)
+function from_digit_to_digit(from::AbstractArray, to::Number, generator, model; γ=0.8, x=X, y=ys, seed=1234, T=100)
     
     x = from
     target = to + 1
@@ -599,7 +599,7 @@ function from_digit_to_digit(from::AbstractArray, to::Number, generator, model; 
 end
 
 # Specific digit:
-function from_digit_to_digit(from::Number, to::Number, generator::Dict, model::Dict; γ=0.95, x=X, y=ys, seed=1234, T=1000)
+function from_digit_to_digit(from::Number, to::Number, generator::Dict, model::Dict; γ=0.8, x=X, y=ys, seed=1234, T=100)
 
     Random.seed!(seed)
 
@@ -623,7 +623,7 @@ We begin by turning the nine in @fig-mnist-orig into a four. @fig-mnist-9to4 sho
 
 ```{julia}
 to = 4
-counterfactuals = from_digit_to_digit(x,to,generators,models)
+counterfactuals = from_digit_to_digit(x,to,generators,models,T=5)
 plts =  first(values(counterfactuals)).x |> x -> plot(convert2image(MNIST, reshape(x,Int(√(input_dim)),Int(√(input_dim)))),title="Original")
 plts = vcat(plts, [plot(convert2image(MNIST, reshape(CounterfactualExplanations.counterfactual(v),Int(√(input_dim)),Int(√(input_dim)))),title=k) for (k,v) in counterfactuals])
 plt = plot(plts...,layout=(1,length(plts)),axis=nothing, size=(1200,300))

[…]
