### 2023-01-27 ad2c6e0
set up a notebook for a whistle-stop tour, which will also serve as tempalte for Pluto notebook
