+++ src/generators/gradient_based/GreedyGenerator.jl
@@ -22,11 +22,13 @@ end
 end
 
 """
-    GreedyGenerator(
-        ;
-        loss::Symbol=:logitbinarycrossentropy,
-        ϵ::Union{AbstractFloat,Nothing}=nothing,
-        n::Union{Int,Nothing}=nothing
+    GreedyGenerator(;
+        loss::Union{Nothing,Symbol}=nothing,
+        complexity::Function=LinearAlgebra.norm,
+        λ::AbstractFloat=0.0,
+        decision_threshold=0.5,
+        opt::Union{Nothing,Flux.Optimise.AbstractOptimiser}, # learning rate
+        kwargs...
     )
 
 An outer constructor method that instantiates a greedy generator.
@@ -38,12 +40,18 @@ generator = GreedyGenerator()
 ```
 """
 function GreedyGenerator(;
-    loss::Union{Nothing,Symbol} = nothing,
-    complexity::Function = LinearAlgebra.norm,
-    λ::AbstractFloat = 0.0,
-    decision_threshold = 0.5,
-    kwargs...,
+    loss::Union{Nothing,Symbol}=nothing,
+    complexity::Function=LinearAlgebra.norm,
+    λ::AbstractFloat=0.0,
+    decision_threshold=0.5,
+    opt::Union{Nothing,Flux.Optimise.AbstractOptimiser}, # learning rate
+    kwargs...
 )
+
+    if !isnothing(opt)
+        @warn "The `GreedyGenerator` does not not work with a `Flux` optimiser. Argument `opt` will be ignored."
+        opt = nothing
+    end
 
     # Load hyperparameters:
     params = GreedyGeneratorParams(; kwargs...)
@@ -95,7 +103,7 @@ function generate_perturbations(
     𝐠ₜ = ∇(generator, counterfactual_explanation.M, counterfactual_explanation) # gradient
     𝐠ₜ[counterfactual_explanation.params[:mutability].==:none] .= 0
     function choose_most_salient(x)
-        s = -((abs.(x) .== maximum(abs.(x), dims = 1)) .* generator.ϵ .* sign.(x))
+        s = -((abs.(x) .== maximum(abs.(x), dims=1)) .* generator.ϵ .* sign.(x))
         non_zero_elements = findall(vec(s) .!= 0)
         # If more than one equal, randomise:
         if length(non_zero_elements) > 1
@@ -106,7 +114,7 @@ function generate_perturbations(
         end
         return s
     end
-    Δs′ = SliceMap.slicemap(x -> choose_most_salient(x), 𝐠ₜ, dims = 1) # choose most salient feature
+    Δs′ = SliceMap.slicemap(x -> choose_most_salient(x), 𝐠ₜ, dims=1) # choose most salient feature
     return Δs′
 end
 
