+++ src/data/synthetic.jl
@@ -3,16 +3,32 @@ using MLJBase
 using Random
 
 """
+    load_blobs(n=100; seed=data_seed, kwrgs...)
+
+Loads overlapping synthtetic data.
+"""
+function load_blobs(n=100; seed=data_seed, kwrgs...)
+
+    Random.seed!(seed)
+
+    X, y = make_blobs(n, 2; kwrgs...)
+    counterfactual_data = CounterfactualData(X, y)
+
+    return counterfactual_data
+
+end
+
+"""
     load_linearly_separable(n=100; seed=data_seed)
 
 Loads linearly separable synthtetic data.
 """
 function load_linearly_separable(n=100; seed=data_seed)
 
-    Random.seed!(seed)
-
-    X, y = make_blobs(n, 2; centers=2, center_box=(-2 => 2), cluster_std=0.1)
-    counterfactual_data = CounterfactualData(X, y)
+    counterfactual_data = load_blobs(
+        n;
+        seed=seed, centers=2, center_box=(-2 => 2), cluster_std=0.1
+    )
 
     return counterfactual_data
 
@@ -25,26 +41,26 @@ Loads overlapping synthtetic data.
 """
 function load_overlapping(n=100; seed=data_seed)
 
-    Random.seed!(seed)
-
-    X, y = make_blobs(n, 2; centers=2, center_box=(-2 => 2), cluster_std=0.5)
-    counterfactual_data = CounterfactualData(X, y)
+    counterfactual_data = load_blobs(
+        n;
+        seed=seed, centers=2, center_box=(-2 => 2), cluster_std=0.5
+    )
 
     return counterfactual_data
 
 end
 
 """
-    load_blobs(n=100; seed=data_seed, kwrgs...)
+    load_multi_class(n=100; seed=data_seed)
 
-Loads overlapping synthtetic data.
+Loads multi-class synthtetic data.
 """
-function load_blobs(n=100; seed=data_seed, kwrgs...)
+function load_multi_class(n=100; seed=data_seed, centers=4)
 
-    Random.seed!(seed)
-
-    X, y = make_blobs(n, 2; kwrgs...)
-    counterfactual_data = CounterfactualData(X, y)
+    counterfactual_data = load_blobs(
+        n;
+        seed=seed, centers=centers, center_box=(-2 => 2), cluster_std=0.1
+    )
 
     return counterfactual_data
 
@@ -83,22 +99,6 @@ function load_moons(n=100; seed=data_seed, kwrgs...)
 end
 
 """
-    load_multi_class(n=100; seed=data_seed)
-
-Loads synthtetic data with multiple classes.
-"""
-function load_multi_class(n=100, n_classes=4; seed=data_seed)
-
-    Random.seed!(seed)
-
-    X, y = make_blobs(n, 2; centers=n_classes, center_box=(-2 => 2), cluster_std=0.1)
-    counterfactual_data = CounterfactualData(X, y)
-
-    return counterfactual_data
-
-end
-
-"""
     load_synthetic_data(n=100; seed=data_seed)
 
 Loads all synthetic datasets and wraps them in a dictionary.
@@ -107,9 +107,9 @@ function load_synthetic_data(n=100; seed=data_seed)
     data = Dict(
         :linearly_separable => load_linearly_separable(n; seed=seed),
         :overlapping => load_overlapping(n; seed=seed),
+        :multi_class => load_multi_class(n; seed=seed),
         :circles => load_circles(n; seed=seed),
         :moons => load_moons(n; seed=seed),
-        :multi_class => load_multi_class(n; seed=seed),
     )
     return data
 end
