### 2023-01-27 6748164
progress on encoding stuff
