+++ test/generative_models.jl
@@ -11,7 +11,7 @@ generator = generator_catalog[:revise]()
 x = select_factual(counterfactual_data, rand(1:size(X, 2)))
 y = predict_label(M, counterfactual_data, x)
 target = get_target(counterfactual_data, y[1])
-counterfactual = generate_counterfactual(x, target, counterfactual_data, M, generator)
+ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
 
 using CounterfactualExplanations.GenerativeModels: retrain!
 CounterfactualExplanations.GenerativeModels.retrain!(
