### 2023-01-28 6c0e174
REVISE definitely working for MNIST pre-trained VAE
