### 2023-01-28 e420b32
vae decoder just returning mean of the data
