+++ src/generative_models/vae.jl
@@ -37,7 +37,7 @@ struct Encoder
 end
 @functor Encoder
 
-Encoder(input_dim::Int, latent_dim::Int, hidden_dim::Int; activation = relu) = Encoder(
+Encoder(input_dim::Int, latent_dim::Int, hidden_dim::Int; activation = sigmoid) = Encoder(
     Dense(input_dim, hidden_dim, activation),   # linear
     Dense(hidden_dim, latent_dim),        # μ
     Dense(hidden_dim, latent_dim),        # logσ
@@ -73,7 +73,7 @@ end
 
 The default decoder architecture is just a Flux Chain with one hidden layer and a linear output layer. 
 """
-Decoder(input_dim::Int, latent_dim::Int, hidden_dim::Int; activation = relu) =
+Decoder(input_dim::Int, latent_dim::Int, hidden_dim::Int; activation = tanh) =
     Chain(Dense(latent_dim, hidden_dim, activation), Dense(hidden_dim, input_dim))
 
 """
@@ -161,11 +161,11 @@ function model_loss(generative_model::VAE, λ, x, device)
     # KL-divergence
     kl_q_p = 0.5f0 * sum(@. (exp(2.0f0 * logσ) + μ^2 - 1.0f0 - 2.0f0 * logσ)) / len
     # Negative log-likelihood: - log(p(x|z))
-    nll_x_z = generative_model.params.nll(z, x, agg = sum) / len
+    nll_x_z = - generative_model.params.nll(z, x, agg = sum) / len
     # Weight regularization:
     reg = λ * sum(x -> sum(x .^ 2), Flux.params(generative_model.decoder))
 
-    elbo = nll_x_z + kl_q_p + reg
+    elbo = - nll_x_z + kl_q_p + reg
 
     return elbo
 end
@@ -189,7 +189,6 @@ function train!(generative_model::VAE, X::AbstractArray, y::AbstractArray; kws..
     end
 
     # training
-    train_steps = 0
     for epoch = 1:args.epochs
 
         avg_loss = []
@@ -202,8 +201,6 @@ function train!(generative_model::VAE, X::AbstractArray, y::AbstractArray; kws..
             avg_loss = vcat(avg_loss, loss)
             grad = back(1.0f0)
             Flux.Optimise.update!(args.opt, ps, grad)
-
-            train_steps += 1
         end
 
         avg_loss = mean(avg_loss)
@@ -214,7 +211,7 @@ function train!(generative_model::VAE, X::AbstractArray, y::AbstractArray; kws..
     end
 
     # Set training status to true:
-    generative_model.trained = true
+    generative_model.trained = true;
 
 end
 
@@ -229,6 +226,12 @@ function retrain!(generative_model::VAE, X::AbstractArray, y::AbstractArray; n_e
 
     # parameters
     ps = Flux.params(generative_model)
+
+    # Verbosity
+    if flux_training_params.verbose
+        @info "Begin training VAE"
+        p_epoch = Progress(args.epochs; desc="Progress on epochs:", showspeed=true, color=:green)
+    end
 
     # training
     train_steps = 0
