+++ docs/src/generators/gradient_based/latent_space_generator.qmd
@@ -32,7 +32,7 @@ We then draw a random factual and generate two counterfactuals - one using gener
 
 ```{julia}
 # Data and Classifier:
-counterfactual_data = load_linearly_separable()
+counterfactual_data = load_linearly_separable(500)
 M = fit_model(counterfactual_data, :Linear)
 
 # Randomly selected factual:
