### 2023-01-28 b89138f
search for MNIST taking quite long without latent space search
