+++ dev/logo/logo.jl
@@ -5,7 +5,10 @@ using Colors
 using CounterfactualExplanations
 using CounterfactualExplanations.Data
 using CounterfactualExplanations.Models
+using Distributions
+using EasyFit
 using Flux
+using LinearAlgebra
 using Luxor
 using MLJBase
 using StatsBase: sample
@@ -19,21 +22,76 @@ julia_colors = Dict(
 )
 bg_color = RGBA(julia_colors[:blue]..., 0.25)
 
-function get_data(N=500; seed=1234, centers=2, center_box=(-2 => 2), cluster_std=0.3)
+function (fit::EasyFit.Cubic)(x::Real)
+    a = fit.a
+    b = fit.b
+    c = fit.c
+    d = fit.d
+    return a * x^3 + b * x^2 + c * x + d
+end
 
-    counterfactual_data = load_blobs(N; seed=seed, centers=centers, center_box=center_box, cluster_std=cluster_std)
-    M = fit_model(counterfactual_data, :Linear)
+function get_data(
+    N=600;
+    seed=2023,
+    cluster_std=0.3,
+    generator=GenericGenerator(),
+    model=:Linear,
+    n_steps=nothing
+)
 
-    return counterfactual_data, M
+
+    Random.seed!(seed)
+    
+    n_per = Int.(round(N / 3))
+    d1 = MvNormal([0, 1], UniformScaling(cluster_std)(2))
+    x1 = rand(d1, n_per)
+    d2 = MvNormal([-1, -1], UniformScaling(cluster_std)(2))
+    x2 = rand(d2, n_per)
+    d3 = MvNormal([1, -1], UniformScaling(cluster_std)(2))
+    x3 = rand(d3, n_per)
+    X = hcat(x1, x2, x3)
+    y = categorical(reduce(vcat, permutedims(repeat([1, 2, 3], 1, n_per))))
+    counterfactual_data = CounterfactualData(X, y)
+    M = fit_model(counterfactual_data, model)
+    X = counterfactual_data.X
+    x = X[1, :]
+    y = X[2, :]
+    factual = select_factual(counterfactual_data, rand(1:size(counterfactual_data.X, 2)))
+    factual_label = predict_label(M, counterfactual_data, factual)[1]
+    target = counterfactual_data.y_levels[counterfactual_data.y_levels.!=factual_label][1]
+
+    # Decision boundary:
+    n_range = 100
+    xlims, ylims = extrema(X, dims=2)
+    _zoom = 0.5 * maximum(abs.(X))
+    xrange = range(xlims[1] - _zoom, xlims[2] + _zoom, length=n_range)
+    yrange = range(ylims[1] - _zoom, ylims[2] + _zoom, length=n_range)
+    target_idx = get_target_index(counterfactual_data.y_levels, target)
+    yhat = [MLJBase.predict(M, counterfactual_data, [x, y])[target_idx] for x in xrange, y in yrange]
+    db_points = findall(0.0 .< yhat .- 0.5 .< 0.1)
+    db_points = map(p -> [xrange[p[1]], yrange[p[2]]], db_points)
+    _X = reduce(hcat, db_points)
+    x1 = _X[1, :]
+    x2 = _X[2, :]
+    fit = fitcubic(x1, x2)
+    _yhat = fit.(xrange)
+    db_points = map(x -> [x[1], x[2]], zip(collect(xrange), _yhat))
+
+
+    # Counterfactual:
+    T = isnothing(n_steps) ? 100 : n_steps
+    ce = generate_counterfactual(factual, target, counterfactual_data, M, generator; T=T)
+
+    return x, y, M, ce, db_points
 end
 
 function logo_picture(;
     ndots=100,
     frame_size=500,
     ms=frame_size // 50,
-    mcolor=(:red, :purple),
[…]
