+++ bib.bib
@@ -1,0 +1,2406 @@
+@InProceedings{altmeyer2023endogenous,
+  author    = {Altmeyer, Patrick and Angela, Giovan and Buszydlik, Aleksander and Dobiczek, Karol and van Deursen, Arie and Liem, Cynthia},
+  booktitle = {First {IEEE} {Conference} on {Secure} and {Trustworthy} {Machine} {Learning}},
+  date      = {2023},
+  title     = {Endogenous {Macrodynamics} in {Algorithmic} {Recourse}},
+  file      = {:altmeyerendogenous - Endogenous Macrodynamics in Algorithmic Recourse.pdf:PDF},
+}
+
+%% This BibTeX bibliography file was created using BibDesk.
+%% https://bibdesk.sourceforge.io/
+
+%% Created for Patrick Altmeyer at 2022-12-13 12:58:22 +0100 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
+@article{abadie2002instrumental,
+	author = {Abadie, Alberto and Angrist, Joshua and Imbens, Guido},
+	date = {2002},
+	date-added = {2022-12-13 12:58:01 +0100},
+	date-modified = {2022-12-13 12:58:01 +0100},
+	journaltitle = {Econometrica : journal of the Econometric Society},
+	number = {1},
+	pages = {91--117},
+	shortjournal = {Econometrica},
+	title = {Instrumental Variables Estimates of the Effect of Subsidized Training on the Quantiles of Trainee Earnings},
+	volume = {70}}
+
+@article{abadie2003economic,
+	author = {Abadie, Alberto and Gardeazabal, Javier},
+	date = {2003},
+	date-added = {2022-12-13 12:58:01 +0100},
+	date-modified = {2022-12-13 12:58:01 +0100},
+	journaltitle = {American economic review},
+	number = {1},
+	pages = {113--132},
+	title = {The Economic Costs of Conflict: {{A}} Case Study of the {{Basque Country}}},
+	volume = {93}}
+
+@inproceedings{ackerman2021machine,
+	author = {Ackerman, Samuel and Dube, Parijat and Farchi, Eitan and Raz, Orna and Zalmanovici, Marcel},
+	booktitle = {2021 {{IEEE}}/{{ACM Third International Workshop}} on {{Deep Learning}} for {{Testing}} and {{Testing}} for {{Deep Learning}} ({{DeepTest}})},
+	date = {2021},
+	date-added = {2022-12-13 12:58:01 +0100},
+	date-modified = {2022-12-13 12:58:01 +0100},
+	pages = {1--8},
+	publisher = {{IEEE}},
+	title = {Machine {{Learning Model Drift Detection Via Weak Data Slices}}}}
+
+@article{allen2017referencedependent,
+	author = {Allen, Eric J and Dechow, Patricia M and Pope, Devin G and Wu, George},
+	date = {2017},
+	date-added = {2022-12-13 12:58:01 +0100},
+	date-modified = {2022-12-13 12:58:01 +0100},
+	journaltitle = {Management Science},
+	number = {6},
+	pages = {1657--1672},
+	title = {Reference-Dependent Preferences: {{Evidence}} from Marathon Runners},
+	volume = {63}}
+
+@article{altmeyer2018option,
+	author = {Altmeyer, Patrick and Grapendal, Jacob Daniel and Pravosud, Makar and Quintana, Gand Derry},
+	date = {2018},
+	date-added = {2022-12-13 12:58:01 +0100},
+	date-modified = {2022-12-13 12:58:01 +0100},
+	title = {Option Pricing in the {{Heston}} Stochastic Volatility Model: An Empirical Evaluation}}
+
+@article{altmeyer2021deep,
[…]
