+++ Artifacts.toml
@@ -79,11 +79,11 @@ lazy = true
     url = "https://github.com/pat-alt/CounterfactualExplanations.jl/releases/download/data/mnist_vae.tar.gz"
 
 [model-vision]
-git-tree-sha1 = "f659c61b99c237bfb2a2bb4c2a52851410eb0753"
+git-tree-sha1 = "c394601e5b2fcda2ab1a26a2a70d5732c99f1541"
 lazy = true
 
     [[model-vision.download]]
-    sha256 = "729ad856b0ce56b467ee9fcf59eab8b5c53727dfb5f07a451d097fc5534dd08f"
+    sha256 = "5a13360fb5c3ec39cf22f25cfbc30f3b760acf51f29fe10f191b448930d7df01"
     url = "https://github.com/pat-alt/CounterfactualExplanations.jl/releases/download/artifacts/model-vision.tar.gz"
 
 [synthetic_data]
