+++ src/models/model_utils.jl
@@ -25,13 +25,14 @@ function model_evaluation(M::AbstractFittedModel, test_data::CounterfactualData)
 end
 
 """
-    MLJBase.predict(M::AbstractFittedModel, counterfactual_data::CounterfactualData, X::Union{Nothing,AbstractArray})
+    predict_proba(M::AbstractFittedModel, counterfactual_data::CounterfactualData, X::Union{Nothing,AbstractArray})
 
 Returns the predicted output probabilities for a given model `M`, data set `counterfactual_data` and input data `X`.
 """
-function MLJBase.predict(M::AbstractFittedModel, counterfactual_data::CounterfactualData, X::Union{Nothing,AbstractArray})
+function predict_proba(M::AbstractFittedModel, counterfactual_data::CounterfactualData, X::Union{Nothing,AbstractArray})
     X = isnothing(X) ? counterfactual_data.X : X
     p = probs(M, X)
+    # println(p)
     binary = M.likelihood == :classification_binary
     function binary_to_onehot(p)
         nobs = size(p, 2)
@@ -49,9 +50,8 @@ end
 Returns the predicted output label for a given model `M`, data set `counterfactual_data` and input data `X`.
 """
 function predict_label(M::AbstractFittedModel, counterfactual_data::CounterfactualData, X::AbstractArray)
-    p = probs(M, X)
     y_levels = counterfactual_data.y_levels
-    p = MLJBase.predict(M, counterfactual_data, X)
+    p = predict_proba(M, counterfactual_data, X)
     y = Flux.onecold(p, y_levels)
     return y
 end
