+++ paper/paper.tex
@@ -41,7 +41,7 @@ bend.'' \cite{oneil2016weapons}
 \end{quote}
 
 In light of all this, a quickly growing body of literature on
-explainable artificial intelligence (XAI) has emerged. Counterfactual
+Explainable Artificial Intelligence (XAI) has emerged. Counterfactual
 Explanations (CE) fall into this broader category. They can help human
 stakeholders make sense of the systems they develop, use or endure: they
 explain how inputs into a system need to change for it to produce
@@ -80,7 +80,7 @@ explain models that were built and trained in \texttt{Python} and
 experimental.
 
 The remainder of this article is structured as follows:
-Section~\ref{sec-related} presents related work on explainable AI as
+Section~\ref{sec-related} presents related work on Explainable AI as
 well as a brief overview of the methodological framework underlying CE.
 Section~\ref{sec-arch} introduces the Julia package and its high-level
 architecture. Section~\ref{sec-use} then presents a number of basic and
@@ -95,15 +95,15 @@ Section~\ref{sec-outlook}. Section~\ref{sec-conclude} concludes.
 \section{Background and related work}\label{sec-related}}
 
 In this section, we first briefly introduce the broad field of
-explainable artificial intelligence (XAI), before narrowing it down to
+Explainable Artificial Intelligence (XAI), before narrowing it down to
 Counterfactual Explanations. We introduce the methodological framework
 and finally point to existing open-source software.
 
 \hypertarget{literature-on-explainable-ai}{%
-\subsection{Literature on explainable
+\subsection{Literature on Explainable
 AI}\label{literature-on-explainable-ai}}
 
-The field of explainable AI is still relatively young and made up of a
+The field of Explainable AI is still relatively young and made up of a
 variety of subdomains, definitions, concepts and taxonomies. Covering
 all of these is beyond the scope of this article, so we will focus only
 on high-level concepts. The following literature surveys provide more
@@ -313,7 +313,7 @@ explainable AI in 2022.\footnote{For details, see the Google Summer of
 With our work on Counterfactual Explanations we hope to contribute to
 these efforts. We think that because of its unique transparency the
 Julia language naturally lends itself towards building a greater degree
-of trust in machine learning and artificial intelligence.
+of trust in Machine Learning and Artificial Intelligence.
 
 \hypertarget{sec-arch}{%
 \section{\texorpdfstring{Introducing:
@@ -424,7 +424,7 @@ x = select_factual(counterfactual_data, chosen) @\label{line:simple-x}@
 
 # Counterfactual search:
 generator = GenericGenerator() @\label{line:simple-gen}@
-counterfactual = generate_counterfactual(
+ce = generate_counterfactual(
     x, target, counterfactual_data, M, generator) @\label{line:simple-search}@
 \end{lstlisting}
 
@@ -453,10 +453,10 @@ the classifier's sensitivity to regions with high aleatoric uncertainty.
 Consider, for example, the scenario illustrated in
 Figure~\ref{fig-binary-wrong}, which involves the same logistic
 classifier, but a massively overfitted version of it. In this case,
-generic search may yield an unrealistic counterfactual that is well into
-the yellow region and yet far away from all other samples (red marker)
-or an ambiguous counterfactual near the decision boundary (purple
-marker).
+generic search may yield an entirely ambiguous counterfactual near the
+decision boundary (purple marker) or an unrealistic counterfactual that
+has moved well into the target domain, but remains far away from all
+other samples (red marker).
 
 \begin{figure}
 
@@ -495,37 +495,38 @@ explanations may be suitable depending on the practical context.
 
[…]
