+++ paper/_paper.qmd
@@ -521,18 +521,33 @@ Now that we have explained the basic functionality of `CounterfactualExplanation
 
 ## Give Me Some Credit
 
-The **Give Me Some Credit** dataset is one of the tabular real-world datasets that ship with the package \cite{kaggle2011give}^[It is publicly available from [Kaggle](https://www.kaggle.com/competitions/GiveMeSomeCredit/leaderboard)]. It can be used to train a binary classifier to predict whether a borrower is likely to experience financial difficulties in the next two years. In particular, we have an output variable $y \in \{0=\text{no stress},1=\text{stress}\}$ and a number of demographic and credit-related features $X$.
-Using the DiCE generator we will generate counterfactuals for a randomly chosen individual in the non-target class. With respect to the mutability of features, we only impose that the `age` cannot be decreased. 
+The **Give Me Some Credit** dataset is one of the tabular real-world datasets that ship with the package \cite{kaggle2011give}^[It is publicly available from [Kaggle](https://www.kaggle.com/competitions/GiveMeSomeCredit/leaderboard)]. It can be used to train a binary classifier to predict whether a borrower is likely to experience financial difficulties in the next two years. In particular, we have an output variable $y \in \{0=\texttt{no stress},1=\texttt{stress}\}$ and a feature matrix $X$ that includes socio-demographic variables like `age` and `income` $X$. A retail bank might use such a classifier to determine if potential borrowers should receive credit or not.
 
 ```{julia}
-counterfactual_data = load_gmsc()
-M = fit_model(counterfactual_data, :MLP)
-ŷ = predict_label(M, counterfactual_data)
+counterfactual_data = load_gmsc(10000)
+train_data, test_data = CounterfactualExplanations.DataPreprocessing.train_test_split(counterfactual_data)
+flux_training_params.batchsize = 50
+flux_training_params.verbose = true
+M = fit_model(train_data, :MLP; dropout=true, n_hidden=64, n_layers=4)
+ŷ = predict_label(M, train_data)
 # Randomly selected factual:
 target_class = 1
 non_target_class = 0
-x = select_factual(counterfactual_data,rand(findall(vec(ŷ).==non_target_class)))
+x = select_factual(train_data,rand(findall(vec(ŷ).==non_target_class),10))
+CounterfactualExplanations.reset!(flux_training_params)
 ```
+
+For the classification task we use a Multi-Layer Perceptron with dropout regularization. 
+
+```{julia}
+#| output: true
+using Markdown
+Markdown.parse("""
+"The F1-score of our pre-trained classifier on test data is: $(round(model_evaluation(M, test_data), digits=2))"
+""")
+```
+
+Using the generic generator (Wachter) we will generate counterfactuals for ten randomly chosen individuals that would be denied credit based on our pre-trained model. With respect to the mutability of features, we only impose that the `age` cannot be decreased. 
 
 ```{julia}
 input_names = Symbol.([
@@ -550,64 +565,89 @@ input_names = Symbol.([
 _mut = [:both for i in 1:length(input_names)]
 _mut[2] = :increase
 counterfactual_data.mutability = _mut
+x1 = :income   # Amount of given credit
+x2 = :age
 ```
 
 ```{julia}
-generator = DiCEGenerator(λ=[0.1, 5.0])
-counterfactual = generate_counterfactual(
-  x, target_class, counterfactual_data, model, generator;
-  num_counterfactuals = 5
+generator = GenericGenerator(
+    decision_threshold = 0.8,
+    opt = Descent(0.1)
+)
+counterfactuals = generate_counterfactual(
[…]

[…]
