+++ paper/paper.tex
@@ -922,7 +922,7 @@ recommend using Latent Space search with care.
 
 \raisebox{-\height}{
 
-\includegraphics{www/mnist_factual.png}
+\includegraphics[width=3.33333in,height=\textheight]{www/mnist_factual.png}
 
 }
 
@@ -937,7 +937,7 @@ recommend using Latent Space search with care.
 
 \raisebox{-\height}{
 
-\includegraphics{www/mnist_vae_strong.png}
+\includegraphics[width=3.33333in,height=\textheight]{www/mnist_vae_strong.png}
 
 }
 
@@ -952,7 +952,7 @@ recommend using Latent Space search with care.
 
 \raisebox{-\height}{
 
-\includegraphics{www/mnist_vae_weak.png}
+\includegraphics[width=3.33333in,height=\textheight]{www/mnist_vae_weak.png}
 
 }
 
