+++ paper/_paper.tex
@@ -145,7 +145,7 @@
 \date{}
 
 \begin{document}
-\ifdefined\Shaded\renewenvironment{Shaded}{\begin{tcolorbox}[enhanced, boxrule=0pt, borderline west={3pt}{0pt}{shadecolor}, frame hidden, interior hidden, sharp corners, breakable]}{\end{tcolorbox}}\fi
+\ifdefined\Shaded\renewenvironment{Shaded}{\begin{tcolorbox}[breakable, interior hidden, sharp corners, frame hidden, enhanced, borderline west={3pt}{0pt}{shadecolor}, boxrule=0pt]}{\end{tcolorbox}}\fi
 
 \hypertarget{sec-intro}{%
 \section{Introduction}\label{sec-intro}}
@@ -1056,7 +1056,7 @@ recommend using Latent Space search with care.
 
 \raisebox{-\height}{
 
-\includegraphics{www/mnist_factual.png}
+\includegraphics[width=3.33333in,height=\textheight]{www/mnist_factual.png}
 
 }
 
@@ -1071,7 +1071,7 @@ recommend using Latent Space search with care.
 
 \raisebox{-\height}{
 
-\includegraphics{www/mnist_vae_strong.png}
+\includegraphics[width=3.33333in,height=\textheight]{www/mnist_vae_strong.png}
 
 }
 
@@ -1086,7 +1086,7 @@ recommend using Latent Space search with care.
 
 \raisebox{-\height}{
 
-\includegraphics{www/mnist_vae_weak.png}
+\includegraphics[width=3.33333in,height=\textheight]{www/mnist_vae_weak.png}
 
 }
 
