+++ paper/paper.tex
@@ -9,7 +9,7 @@
 
 \begin{abstract}
 
-Machine learning models like deep neural networks have become so complex and opaque over recent years that they are generally considered as black boxes. Nonetheless, such models often play a key role in modern automated decision-making systems. Counterfactual Explanations can help human stakeholders make sense of the systems they build, use and endure: they explain how inputs into a system need to change for it to produce different decisions. Explanations that involve realistic and actionable changes can be used for the purpose of Algorithmic Recourse: they offer humans a way to not only understand the behaviour of a system, but also to adjust and react to it. In this article we discuss the usefulness of Counterfactual Explanations for explainable machine learning and demonstrate its implementation in Julia using the \verb|CounterfactualExplanations.jl| package. The package is straight-forward to use and designed with a focus on customization and extensibility. We envision it to one day be the go-to place for explaining arbitrary predictive models in Julia and beyond through a diverse suite of counterfactual generators.
+Machine Learning models like Deep Neural Networks have become so complex and opaque over recent years that they are generally considered Black Boxes. Nonetheless, such models often play a key role in modern automated decision-making systems. Counterfactual Explanations can help human stakeholders make sense of the systems they build, use and endure: they explain how inputs into a system need to change for it to produce different decisions. Explanations that involve realistic and actionable changes can be used for the purpose of Algorithmic Recourse: they offer humans a way to not only understand the behaviour of a system, but also to adjust and react to it. In this article we discuss the usefulness of Counterfactual Explanations for Explainable Machine Learning and demonstrate its implementation in Julia using the \verb|CounterfactualExplanations.jl| package. The package is straightforward to use and designed with a focus on customization and extensibility. We envision it to one day be the go-to place for explaining arbitrary predictive models in Julia through a diverse suite of counterfactual generators.
 
 \end{abstract}
 
@@ -49,7 +49,7 @@ different decisions. Explainability benefits internal as well as
 external quality assurance. Explanations that involve realistic and
 actionable changes can be used for Algorithmic Recourse (AR): they offer
 the group of principals a way to not only understand their agent's
-behaviour, but also adjust or react to it.
+behaviour but also adjust or react to it.
 
 The availability of open-source software to explain black-box models
 through counterfactuals is still limited. Most existing implementations
@@ -134,7 +134,7 @@ models altogether and instead focus solely on interpretable AI
 be geared towards interpretable models, avoiding black boxes altogether
 would entail missed opportunities and anyway is probably not very
 realistic at this point. For that reason, we expect the need for
-explainable AI to persist in the near future. Explainable AI can further
+explainable AI to persist in the medium term. Explainable AI can further
 be broadly divided into \textbf{global} and \textbf{local}
 explainability: the former is concerned with explaining the average
 behaviour of a model, while the latter involves explanations for
@@ -247,22 +247,22 @@ in terms of the Manhattan distance between factual and counterfactual
 feature values. While this is an intuitive choice, it is too simple to
 address many of the desirable properties of effective Counterfactual
 Explanations that have been set out. These desiderata include:
[…]
