+++ paper/_paper.tex
@@ -145,7 +145,7 @@
 \date{}
 
 \begin{document}
-\ifdefined\Shaded\renewenvironment{Shaded}{\begin{tcolorbox}[breakable, interior hidden, sharp corners, frame hidden, enhanced, borderline west={3pt}{0pt}{shadecolor}, boxrule=0pt]}{\end{tcolorbox}}\fi
+\ifdefined\Shaded\renewenvironment{Shaded}{\begin{tcolorbox}[borderline west={3pt}{0pt}{shadecolor}, frame hidden, boxrule=0pt, sharp corners, interior hidden, breakable, enhanced]}{\end{tcolorbox}}\fi
 
 \hypertarget{sec-intro}{%
 \section{Introduction}\label{sec-intro}}
@@ -183,7 +183,7 @@ different decisions. Explainability benefits internal as well as
 external quality assurance. Explanations that involve realistic and
 actionable changes can be used for Algorithmic Recourse (AR): they offer
 the group of principals a way to not only understand their agent's
-behaviour, but also adjust or react to it.
+behaviour but also adjust or react to it.
 
 The availability of open-source software to explain black-box models
 through counterfactuals is still limited. Most existing implementations
@@ -268,7 +268,7 @@ models altogether and instead focus solely on interpretable AI
 be geared towards interpretable models, avoiding black boxes altogether
 would entail missed opportunities and anyway is probably not very
 realistic at this point. For that reason, we expect the need for
-explainable AI to persist in the near future. Explainable AI can further
+explainable AI to persist in the medium term. Explainable AI can further
 be broadly divided into \textbf{global} and \textbf{local}
 explainability: the former is concerned with explaining the average
 behaviour of a model, while the latter involves explanations for
@@ -381,22 +381,22 @@ in terms of the Manhattan distance between factual and counterfactual
 feature values. While this is an intuitive choice, it is too simple to
 address many of the desirable properties of effective Counterfactual
 Explanations that have been set out. These desiderata include:
-\textbf{closeness} - the average distance between factual and
+\textbf{closeness} --- the average distance between factual and
 counterfactual features should be small
-\cite{wachter2017counterfactual}; \textbf{actionability} - the proposed
-feature perturbation should actually be actionable
+\cite{wachter2017counterfactual}; \textbf{actionability} --- the
+proposed feature perturbation should actually be actionable
 (\cite{ustun2019actionable}, \cite{poyiadzi2020face});
-\textbf{plausibility} - the counterfactual explanation should be
+\textbf{plausibility} --- the counterfactual explanation should be
 realistic plausible to a human (\cite{joshi2019realistic},
-\cite{schut2021generating}); \textbf{unambiguity} - a human should have
-no trouble assigning a label to the counterfactual
-\cite{schut2021generating}; \textbf{sparsity} - the counterfactual
+\cite{schut2021generating}); \textbf{unambiguity} --- a human should
+have no trouble assigning a label to the counterfactual
+\cite{schut2021generating}; \textbf{sparsity} --- the counterfactual
 explanation should involve as few individual feature changes as possible
-\cite{schut2021generating}; \textbf{robustness} - the counterfactual
+\cite{schut2021generating}; \textbf{robustness} --- the counterfactual
 explanation should be robust to domain and model shifts
-\cite{upadhyay2021robust}; \textbf{diversity} - ideally multiple diverse
-Counterfactual Explanations should be provided
-\cite{mothilal2020explaining}; and \textbf{causality} - Counterfactual
[…]
