+++ paper/_paper.qmd
@@ -4,6 +4,7 @@ bibliography: ref.bib
 execute:
     eval: false
     echo: false
+    output: false
 jupyter: julia-1.8
 ---
 
@@ -21,10 +22,12 @@ using MLDatasets: convert2image
 using MLUtils
 using Plots
 using Random
+using Serialization
 Random.seed!(42)
 theme(:wong)
 default(size=(500, 375))
 www_path = "paper/www"
+output_path = "paper/output"
 ```
 
 # Introduction {#sec-intro}
@@ -35,7 +38,7 @@ As others have pointed out, this scenario gives rise to an undesirable principal
 
 > “You cannot appeal to (algorithms). They do not listen. Nor do they bend.” \cite{oneil2016weapons}
 
-In light of all this, a quickly growing body of literature on Explainable Artificial Intelligence (XAI) has emerged. Counterfactual Explanations (CE) fall into this broader category. They can help human stakeholders make sense of the systems they develop, use or endure: they explain how inputs into a system need to change for it to produce different decisions. Explainability benefits internal as well as external quality assurance. Explanations that involve realistic and actionable changes can be used for Algorithmic Recourse (AR): they offer the group of principals a way to not only understand their agent's behaviour, but also adjust or react to it.  
+In light of all this, a quickly growing body of literature on Explainable Artificial Intelligence (XAI) has emerged. Counterfactual Explanations (CE) fall into this broader category. They can help human stakeholders make sense of the systems they develop, use or endure: they explain how inputs into a system need to change for it to produce different decisions. Explainability benefits internal as well as external quality assurance. Explanations that involve realistic and actionable changes can be used for Algorithmic Recourse (AR): they offer the group of principals a way to not only understand their agent's behaviour but also adjust or react to it.  
 
 The availability of open-source software to explain black-box models through counterfactuals is still limited. Most existing implementations are specific to particular methodologies. They are also exclusively built in Python and for Python models. The only existing unifying software approach, for example, is tailored to models built in the two most popular Python libraries for deep learning. The Julia ecosystem has so far lacked an open-source implementation of CE. 
 
@@ -53,7 +56,7 @@ In this section, we first briefly introduce the broad field of Explainable Artif
 
 The field of Explainable AI is still relatively young and made up of a variety of subdomains, definitions, concepts and taxonomies. Covering all of these is beyond the scope of this article, so we will focus only on high-level concepts. The following literature surveys provide more detail: Arrieta et al. (2020) provide a broad overview of XAI \cite{arrieta2020explainable}; Fan et al. (2020) focus on explainability in the context of deep learning \cite{fan2020interpretability}; and finally, Karimi et al. (2020) \cite{karimi2020survey} and Verma et al. (2020) \cite{verma2020counterfactual} offer detailed reviews of the literature on Counterfactual Explanations and Algorithmic Recourse.^[Readers who prefer a text-book approach may also want to consider \cite{molnar2020interpretable} and \cite{varshney2022trustworthy}] Finally, Miller (2019) explicitly discusses the concept of explainability from the perspective of a social scientist \cite{miller2019explanation}.
 
[…]

[…]
