+++ dev/tutorials/whistle_stop.qmd
@@ -1,0 +1,64 @@
+```@meta
+CurrentModule = CounterfactualExplanations 
+```
+
+```{julia}
+#| echo: false
+include("docs/setup_docs.jl")
+eval(setup_docs)
+```
+
+# Whistle-Stop Tour
+
+## Full Illustrative Example
+
+```{julia}
+n = 200
+data_name = :multi_class
+model_name = :MLP
+counterfactual_data = data_catalogue[:synthetic][data_name](n)
+M = fit_model(counterfactual_data, model_name)
+```
+
+```{julia}
+# Factual and target:
+x = select_factual(counterfactual_data,rand(1:size(counterfactual_data.X,2)))
+y = predict_label(M, counterfactual_data, x)[1]
+target = counterfactual_data.y_levels[counterfactual_data.y_levels .!= y][1]
+# Search params:
+opt = Descent(0.01)
+γ = 0.95
+ces = Dict()
+plts = []
+# Search:
+for (key, Generator) in generator_catalog
+    generator = Generator(; 
+        opt=opt, 
+        decision_threshold=γ
+    )
+    ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
+    ces[key] = ce
+    plts = [plts..., plot(ce; title=key, colorbar=false)]
+end
+```
+
+```{julia}
+plt_height = 300
+_size = length(plts)/2 * plt_height
+_ncol = ceil(sqrt(length(plts)))
+_nrow = ceil(length(plts)/_ncol)
+asp_ratio = _ncol/_nrow
+plot(
+    plts..., 
+    ncol=_ncol, size=(asp_ratio * _size,_size),
+    plot_title="Factual: $(y) → Target: $(target)"
+)
+```
+
+## Complete Example
+
+
+
+
+
+
