+++ src/generators/gradient_based/DICEGenerator.jl
@@ -37,7 +37,7 @@ generator = DiCEGenerator()
 function DiCEGenerator(;
     loss::Union{Nothing,Symbol} = nothing,
     complexity::Function = norm,
-    λ::Union{AbstractFloat,AbstractVector} = [0.1, 1.0],
+    λ::Union{AbstractFloat,AbstractVector} = [0.1, 0.5],
     decision_threshold = 0.5,
     kwargs...,
 )
