+++ docs/setup_docs.jl
@@ -10,6 +10,7 @@ setup_docs = quote
     using Flux
     using Flux.Optimise: update!, Adam
     using LinearAlgebra
+    using Markdown
     using MLJBase
     using MLJModels: OneHotEncoder
     using Plots
