+++ docs/src/tutorials/whistle_stop.md
@@ -1,0 +1,66 @@
+
+``` @meta
+CurrentModule = CounterfactualExplanations 
+```
+
+# Whistle-Stop Tour
+
+In this tutorial, we will go through a simple example involving synthetic data. We will generate Counterfactual Explanations using different generators and visualize the results.
+
+## Data and Classifier
+
+``` julia
+counterfactual_data = load_blobs(n_samples; k=n_dim, centers=n_classes)
+M = fit_model(counterfactual_data, model_name)
+```
+
+The chart below visualizes our data along with the model predictions. In particular, the contour indicates the predicted probabilities generated by our classifier. By default, these are the predicted probabilities for *y* = 1, the first label. For multi-dimensional input data is compressed into two dimensions and the decision boundary is approximated using Nearest Neighbors (this is still somewhat experimental).
+
+``` julia
+plot(M, counterfactual_data)
+```
+
+![](whistle_stop_files/figure-commonmark/cell-6-output-1.svg)
+
+## Counterfactual Explanation
+
+Next, we begin by specifying our target and factual label. We then draw a random sample from the non-target (factual) class.
+
+``` julia
+# Factual and target:
+target = 2
+factual = 4
+chosen = rand(findall(predict_label(M, counterfactual_data) .== factual))
+x = select_factual(counterfactual_data,chosen)
+```
+
+Counterfactual generators accept several default parameters that can be used to adjust the counterfactual search at a high level: for example, a `Flux.jl` optimizer can be supplied to define how exactly gradient steps are performed. Importantly, one can also define the threshold probability at which the counterfactual search will converge. This relates to the probability predicted by the underlying black-box model, that the counterfactual belongs to the target class. A higher decision threshold typically prolongs the counterfactual search.
+
+``` julia
+# Search params:
+opt = Descent()
+decision_threshold = 0.75
+num_counterfactuals = 3
+```
+
+The code below runs the counterfactual search for each generator available in the `generator_catalogue`. In each case, we also call the generic `plot()` method on the generated instance of type `CounterfactualExplanation`. This generates a simple plot that visualizes the entire counterfactual path. The chart below shows the results for all counterfactual generators.
+
+``` julia
+ces = Dict()
+plts = []
+# Search:
+for (key, Generator) in generator_catalog
+    generator = Generator(; 
+        opt=opt, 
+        decision_threshold=decision_threshold
+    )
+    ce = generate_counterfactual(
+        x, target, counterfactual_data, M, generator;
+        num_counterfactuals = num_counterfactuals
+    )
+    ces[key] = ce
+    plts = [plts..., plot(ce; title=key, colorbar=false)]
+end
+```
+
+![](whistle_stop_files/figure-commonmark/cell-11-output-1.svg)
