+++ docs/src/contributing/interop.md
@@ -1,266 +1,0 @@
-
-``` @meta
-CurrentModule = CounterfactualExplanations 
-```
-
-# Interoperability
-
-    !!! info "Contributor's Guide" 
-        Our work on language interoperability is still experimental. In this tutorial we demonstrate our early attempts to extend the package to accommodate `torch` models trained in R and Python. The goal here is to provide a template and/or starting point for contributors that would like to add support for other models trained in foreign programming languages. If that contributor could be you, please see the related [issue](https://github.com/pat-alt/CounterfactualExplanations.jl/issues/67) and work in the linked [branch](https://github.com/pat-alt/CounterfactualExplanations.jl/tree/67-interoperability-broken-after-moving-to-julia-18).
-
-    !!! info "Native support currently broken and removed" 
-        After bumping Julia version compatibility to `1.8`, interoperability broke for some reason (it was admittedly always experimental and hacky). The code that was previously used for interoperability was removed from the main branch, but still lives [here](https://github.com/pat-alt/CounterfactualExplanations.jl/tree/67-interoperability-broken-after-moving-to-julia-18).
-
-The Julia language offers unique support for programming language interoperability. For example, calling Python and R is made remarkably easy through `PyCall.jl` (potentially `PythonCall.jl` is a better option) and `RCall.jl`. In this tutorial we will see how `CounterfactualExplanations.jl` can leverage this functionality.
-
-To get started we will first load some two-dimensional toy data:
-
-``` julia
-using Random
-# Some random data:
-Random.seed!(1234);
-N = 100
-using CounterfactualExplanations
-using CounterfactualExplanations.Data
-xs, ys = Data.toy_data_non_linear(N)
-X = hcat(xs...)
-counterfactual_data = CounterfactualData(X,ys')
-```
-
-## `torch` model trained in R
-
-The code below builds a simple MLP in R:
-
-``` julia
-using RCall
-R"""
-# Data
-library(torch)
-X <- torch_tensor(t($X))
-ys <- torch_tensor($ys)
-
-# Model:
-mlp <- nn_module(
-  initialize = function() {
-    self$layer1 <- nn_linear(2, 32)
-    self$layer2 <- nn_linear(32, 1)
-  },
-  forward = function(input) {
-    input <- self$layer1(input)
-    input <- nnf_sigmoid(input)
-    input <- self$layer2(input)
-    input
-  }
-)
-model <- mlp()
-optimizer <- optim_adam(model$parameters, lr = 0.1)
-loss_fun <- nnf_binary_cross_entropy_with_logits
-"""
-```
-
-The following code trains the MLP for the binary prediction task at hand:
-
-``` julia
-R"""
-for (epoch in 1:100) {
-
-  model$train()  
-
-  # Compute prediction and loss:
-  output <- model(X)[,1]
-  loss <- loss_fun(output, ys)
-
-  # Backpropagation:
-  optimizer$zero_grad()
-  loss$backward()
-  optimizer$step()
-  
-  cat(sprintf("Loss at epoch %d: %7f\n", epoch, loss$item()))
-}
-"""
-```
-
-### Making the model compatible
-
[…]

[…]
