+++ docs/make.jl
@@ -23,25 +23,31 @@ makedocs(;
         "🫣 Tutorials" => [
             "Overview" => "tutorials/_index.md",
             "Whiste-Stop Tour" => "tutorials/whistle-stop.md",
-            "Datasets" => "tutorials/data.md",
-            "Models" => "tutorials/models.md",
-            "Data Pre-Processing" => "tutorials/data_preprocessing.md",
-        ],
-        "🫡 How-To ..." => [
-            "Overview" => "how_to_guides/_index.md",
-            "... explain an image classifer" => "how_to_guides/mnist.qmd",
-            "... add custom models" => "how_to_guides/custom_models.md",
-            "... add custom generators" => "how_to_guides/custom_generators.md",
-            "... explain R/Python models" => "how_to_guides/interop.md",
+            "Handling Data" => "tutorials/data_preprocessing.md",
+            "Data Catalogue" => "tutorials/data_catalogue.md",
+            "Handling Models" => "tutorials/models.md",
+            "Model Catalogue" => "tutorials/model_catalogue.md",
         ],
         "🤓 Explanation" => [
             "Overview" => "explanation/_index.md",
             "Package Architecture" => "explanation/architecture.md",
             "Generators" => [
-                "REVISE" => "explanation/generators/latent_space_generator.md",
+                "Generic" => "explanation/generators/generic.md", 
+                "Gravitational" => "explanation/generators/gravitational.md", 
+                "REVISE" => "explanation/generators/revise.md",
                 "DiCE" => "explanation/generators/dice.md",
+                "ClaPROAR" => "explanation/generators/clap_roar.md", 
+                "Greedy" => "explanation/generators/greedy.md", 
             ],
-            "Loss functions" => "contributing/loss.md",
+            "Categorical Features" => "explanation/categorical.md",
+            "Loss functions" => "explanation/loss.md",
+        ],
+        "🫡 How-To ..." => [
+            "Overview" => "how_to_guides/_index.md",
+            "... explain an image classifer" => "how_to_guides/mnist.md",
+            "... add custom models" => "how_to_guides/custom_models.md",
+            "... add custom generators" => "how_to_guides/custom_generators.md",
+            "... explain R/Python models" => "how_to_guides/interop.md",
         ],
         "🧐 Reference" => "_reference.md",
         "🛠 Contribute" => "contribute.md",
