+++ docs/src/tutorials/models.qmd
@@ -1,0 +1,12 @@
+```@meta
+CurrentModule = CounterfactualExplanations 
+```
+
+```{julia}
+#| echo: false
+include("docs/setup_docs.jl")
+eval(setup_docs)
+```
+
+# Handling Models
+
