+++ docs/src/tutorials/data_preprocessing.qmd
@@ -1,221 +1,134 @@
----
-output: true
----
-
-## Data Preprocessing
+```@meta
+CurrentModule = CounterfactualExplanations 
+```
 
 ```{julia}
 #| echo: false
 include("docs/setup_docs.jl")
-eval(setup_docs);
+eval(setup_docs)
 ```
 
-To illustrate how data is preprocessed, we consider a simple toy dataset with three categorical features (`name`, `grade` and `sex`) and one continuous feature (`age`):
+# Handling Data
+
+## `CounterfactualData`
+
+## Categorical Features
+
+## Immutable Features
+
+In practice, features usually cannot be perturbed arbitrarily. Suppose, for example, that one of the features used by a bank to predict the credit worthiness of its clients is *gender*. If a counterfactual explanation for the prediction model indicates that female clients should change their gender to improve their credit worthiness, then this is an interesting insight (it reveals gender bias), but it is not usually an actionable transformation in practice. In such cases we may want to constrain the mutability of features to ensure actionable and realistic recourse. To illustrate how this can be implemented in `CounterfactualExplanations.jl` we will look at the linearly separable toy dataset again.
 
 ```{julia}
-X = (
-    name=categorical(["Danesh", "Lee", "Mary", "John"]),
-    grade=categorical(["A", "B", "A", "C"], ordered=true),
-    sex=categorical(["male","female","male","male"]),
-    height=[1.85, 1.67, 1.5, 1.67],
-)
-schema(X)
+# Data:
+using CounterfactualExplanations.Data
+xs, ys = Data.toy_data_linear()
+X = hcat(xs...)
+
+# Model
+w = [1.0 1.0]# true coefficients
+b = 0
+# Logit model:
+M = LogisticModel(w, [b])
 ```
 
-Categorical features are expected to be one-hot or dummy encoded. To this end, we could use `MLJ`, for example:
+## Mutability
+
+Mutability of features can be defined in terms of four different options: 1) the feature is mutable in both directions, 2) the feature can only increase (e.g. age), 3) the feature can only decrease (e.g. time left until your next deadline) and 4) the feature is not mutable (e.g. skin colour, ethnicity, ...). To specify which category a feature belongs to, you can pass a vector of symbols containing the mutability constraints at the pre-processing stage. For each feature you can choose from these four options: `:both` (mutable in both directions), `:increase` (only up), `:decrease` (only down) and `:none` (immutable). By default, `nothing` is passed to that keyword argument and it is assumed that all features are mutable in both directions.
+
+Below we impose that the second feature is immutable. The resulting counterfactual path is shown in @fig-mutability below. Since only the first feature can be perturbed, the sample can only move along the horizontal axis.
 
 ```{julia}
-hot = OneHotEncoder()
-mach = fit!(machine(hot, X))
-W = transform(mach, X)
-schema(W)
-```
-
-The matrix that will actually be perturbed during the counterfactual search looks as follows:
-
-```{julia}
-X = permutedims(MLJBase.matrix(W))
-```
-
[…]
