+++ src/data_preprocessing/functions.jl
@@ -2,6 +2,7 @@ using CategoricalArrays
 using CounterfactualExplanations
 using Flux
 using MultivariateStats
+using SliceMap
 using StatsBase
 using Tables
 using UMAP
@@ -102,7 +103,6 @@ counterfactual_data = CounterfactualData(X,y')
 function CounterfactualData(
     X::AbstractMatrix,
     y::RawOutputArrayType;
-    likelihood::Union{Nothing, Symbol}=nothing,
     mutability::Union{Vector{Symbol},Nothing}=nothing,
     domain::Union{Any,Nothing}=nothing,
     features_categorical::Union{Vector{Vector{Int}},Nothing}=nothing,
@@ -257,16 +257,20 @@ A subroutine that is used to apply the predetermined domain constraints.
 """
 function apply_domain_constraints(counterfactual_data::CounterfactualData, x::AbstractArray)
 
-    # Continuous variables:
-    if !isnothing(counterfactual_data.domain)
-        for i in counterfactual_data.features_continuous
-            x[i] = clamp(
-                x[i],
-                counterfactual_data.domain[i][1],
-                counterfactual_data.domain[i][2],
-            )
+    SliceMap.slicemap(x, dims=(1, 2)) do _x
+        # Continuous variables:    
+        if !isnothing(counterfactual_data.domain)
+            for i in counterfactual_data.features_continuous
+                _x_i = clamp(
+                    _x[i,:][1],
+                    counterfactual_data.domain[i][1],
+                    counterfactual_data.domain[i][2],
+                )
+                _x[i,:] = _x_i
+            end
         end
     end
+    
 
     return x
 
