+++ docs/src/tutorials/data_preprocessing.md
@@ -1,0 +1,274 @@
+
+``` @meta
+CurrentModule = CounterfactualExplanations 
+```
+
+# Handling Data
+
+The package works with custom data containers that contain the input and output data as well as information about the type and mutability of features. In this tutorial, we will see how data can be prepared for use with the package.
+
+## Basic Functionality
+
+To demonstrate the basic way to prepare data, let’s look at a standard benchmark dataset: Fisher’s classic iris dataset. We can use `MLDatasets` to load this data.
+
+``` julia
+dataset = Iris()
+```
+
+Our data constructor `CounterfactualData` needs at least two inputs: features `X` and targets `y`.
+
+``` julia
+X = dataset.features
+y = dataset.targets
+```
+
+Next, we convert the input data to a `Tables.MatrixTable` (following `MLJ.jl`) convention. Concerning the target variable, we just assign grab the first column of the data frame.
+
+``` julia
+X = table(Tables.matrix(X))
+y = y[:,1]
+```
+
+Now we can feed these two ingredients to our constructor:
+
+``` julia
+counterfactual_data = CounterfactualData(X, y)
+```
+
+Under the hood, the constructor performs basic preprocessing steps. For example, the output variable `y` is automatically one-hot encoded:
+
+``` julia
+counterfactual_data.y
+```
+
+    3×150 Matrix{Bool}:
+     1  1  1  1  1  1  1  1  1  1  1  1  1  …  0  0  0  0  0  0  0  0  0  0  0  0
+     0  0  0  0  0  0  0  0  0  0  0  0  0     0  0  0  0  0  0  0  0  0  0  0  0
+     0  0  0  0  0  0  0  0  0  0  0  0  0     1  1  1  1  1  1  1  1  1  1  1  1
+
+Similarly, a transformer used to scale continuous input features is automatically fitted:
+
+``` julia
+counterfactual_data.dt
+```
+
+    ZScoreTransform{Float64, Vector{Float64}}(4, 2, [5.843333333333335, 3.0540000000000007, 3.7586666666666693, 1.1986666666666672], [0.8280661279778629, 0.4335943113621737, 1.7644204199522617, 0.7631607417008414])
+
+## Categorical Features
+
+For the counterfactual search, it is important to distinguish between continuous and categorical features. This is because categorical features cannot be perturbed arbitrarily: they can take specific discrete values, but not just any value on the real line.
+
+Consider the following example:
+
+``` julia
+y = rand([1,0],4)
+X = (
+    name=categorical(["Danesh", "Lee", "Mary", "John"]),
+    grade=categorical(["A", "B", "A", "C"], ordered=true),
+    sex=categorical(["male","female","male","male"]),
+    height=[1.85, 1.67, 1.5, 1.67],
+)
+schema(X)
+```
+
+    ┌────────┬──────────────────┬──────────────────────────────────┐
+    │ names  │ scitypes         │ types                            │
+    ├────────┼──────────────────┼──────────────────────────────────┤
+    │ name   │ Multiclass{4}    │ CategoricalValue{String, UInt32} │
+    │ grade  │ OrderedFactor{3} │ CategoricalValue{String, UInt32} │
+    │ sex    │ Multiclass{2}    │ CategoricalValue{String, UInt32} │
+    │ height │ Continuous       │ Float64                          │
+    └────────┴──────────────────┴──────────────────────────────────┘
+
+Typically, in the context of Unserpervised Learning, categorical features are one-hot or dummy encoded. To this end, we could use `MLJ`, for example:
+
+``` julia
+hot = OneHotEncoder()
+mach = MLJBase.fit!(machine(hot, X))
+W = MLJBase.transform(mach, X)
+X = permutedims(MLJBase.matrix(W))
+```
+
[…]
