### 2023-02-01 4105cc0
CompatHelper: add new compat entry for ProgressMeter at version 1, (keep existing compat)
