+++ Project.toml
@@ -45,6 +45,7 @@ MultivariateStats = "0.9, 0.10"
 NearestNeighborModels = "0.2"
 Parameters = "0.12"
 Plots = "1.38.2"
+ProgressMeter = "1"
 SliceMap = "0.2"
 StatsBase = "0.33"
 Tables = "1"
