### 2023-02-01 d41c159
CompatHelper: add new compat entry for PkgTemplates at version 0.7, (keep existing compat)
