+++ Project.toml
@@ -44,6 +44,7 @@ MLUtils = "0.2, 0.3, 0.4"
 MultivariateStats = "0.9, 0.10"
 NearestNeighborModels = "0.2"
 Parameters = "0.12"
+PkgTemplates = "0.7"
 Plots = "1.38.2"
 SliceMap = "0.2"
 StatsBase = "0.33"
