+++ docs/setup_docs.jl
@@ -9,9 +9,11 @@ setup_docs = quote
     using CounterfactualExplanations.Data
     using Flux
     using Flux.Optimise: update!, Adam
+    using Images
     using LinearAlgebra
     using Markdown
     using MLDatasets
+    using MLDatasets: convert2image
     using MLJBase
     using MLJModels: OneHotEncoder
     using Plots
@@ -25,5 +27,13 @@ setup_docs = quote
     www_path = "docs/src/www"
     include("docs/src/utils.jl")
     synthetic = CounterfactualExplanations.Data.load_synthetic_data()
+
+    # Counteractual data and model:
+    counterfactual_data = load_linearly_separable()
+    M = fit_model(counterfactual_data, :Linear)
+    target = 2
+    factual = 1
+    chosen = rand(findall(predict_label(M, counterfactual_data) .== factual))
+    x = select_factual(counterfactual_data, chosen)
 
 end;
