+++ docs/src/explanation/generators/binary.qmd
@@ -1,140 +1,0 @@
-```@meta
-CurrentModule = CounterfactualExplanations 
-```
-
-# Binary classification
-
-```{julia}
-#| echo: false
-using Pkg; Pkg.activate("docs")
-using Flux, Random, Plots, PlotThemes, CounterfactualExplanations
-theme(:wong)
-using Logging
-disable_logging(Logging.Info)
-include("dev/utils.jl") # some plotting functions
-www_path = "docs/src/tutorials/www"
-```
-
-To understand the core functionality of `CounterfactualExplanations.jl` we will look at two example use cases of the `generate_counterfactual` function. This function takes a structure of type `AbstractGenerator` as its main argument. You can utilize one of the [default generators](#default-generators): `GenericGenerator <: AbstractGenerator`, `GreedyGenerator <: AbstractGenerator`. Alternatively, you can also create your own custom generators as we will see in own of the following sections. 
-
-## Default generators
-
-### `GenericGenerator`
-
-Let $t\in\{0,1\}$ denote the target label, $M$ the model (classifier) and x′ ∈ ℝᴰ the vector of counterfactual features. In order to generate recourse the `GenericGenerator` optimizes the following objective function through steepest descent
-
-```math
-x\prime = \arg \min_{x\prime}  \ell(M(x\prime),t) + \lambda h(x\prime)
-```
-
-where $\ell$ denotes some loss function targeting the deviation between the target label and the predicted label and $h(\cdot)$ is a complexity penalty generally addressing the *realism* or *cost* of the proposed counterfactual. 
-
-Let's generate some toy data:
-
-```{julia}
-# Some random data:
-using CounterfactualExplanations.Data
-Random.seed!(1234)
-N = 25
-w = [1.01 1.0]# true coefficients
-b = 0
-xs, ys = Data.toy_data_linear(N)
-X = hcat(xs...)
-counterfactual_data = CounterfactualData(X,ys')
-plt = plot()
-plt = scatter!(counterfactual_data)
-savefig(plt, joinpath(www_path, "binary_samples.png"))
-```
-
-![](www/binary_samples.png)
-
-For this toy data we will now generate counterfactual explanations as follows:
-
-- Use the coefficients `w` and `b` (assumed to be known or estimated) to define our model using `CounterfactualExplanations.Models.LogisticModel(w, b)`. (The first figure below shows the posterior predictive surface for this plugin estimator.)
-- Define our `GenericGenerator`.
-- Generate counterfactual.
-
-```{julia}
-# Logit model:
-using CounterfactualExplanations.Models: LogisticModel, probs 
-M = LogisticModel(w, [b])
-# Randomly selected factual:
-Random.seed!(123)
-x = select_factual(counterfactual_data,rand(1:size(X)[2]))
-y = round(probs(M, x)[1])
-target = ifelse(y==1.0,0.0,1.0) # opposite label as target
-
-# Define generator:
-generator = GenericGenerator()
-
-# Generate explanations:
-counterfactual = generate_counterfactual(x, target, counterfactual_data, M, generator)
-```
-
-```{julia}
-#| echo: false
-plt = plot(M,counterfactual_data;title="Posterior predictive - Plugin")
-savefig(plt, joinpath(www_path, "binary_contour.png"))
-```
-
-![](www/binary_contour.png)
-
[…]
