+++ src/generators/gradient_based/DICEGenerator.jl
@@ -37,8 +37,8 @@ generator = DiCEGenerator()
 function DiCEGenerator(;
     loss::Union{Nothing,Symbol} = nothing,
     complexity::Function = norm,
-    λ::Union{AbstractFloat,AbstractVector} = [0.1, 0.5],
-    decision_threshold = 0.5,
+    λ::Union{AbstractFloat,AbstractVector} = [0.05, 1.0],
+    decision_threshold = nothing,
     kwargs...,
 )
     params = DiCEGeneratorParams(; kwargs...)
