+++ docs/src/explanation/generators/greedy.qmd
@@ -1,101 +1,33 @@
----
-format: 
-  commonmark:
-    variant: -raw_html
-    wrap: none
-    self-contained: true
-crossref:
-  fig-prefix: Figure
-  tbl-prefix: Table
-bibliography: https://raw.githubusercontent.com/pat-alt/bib/main/bib.bib
-output: asis
-execute: 
-  eval: false
-  echo: true
-jupyter: julia-1.7
----
-
 ```@meta
 CurrentModule = CounterfactualExplanations 
 ```
 
 ```{julia}
 #| echo: false
-using Pkg; Pkg.activate("docs")
-using Flux, Random, Plots, PlotThemes, CounterfactualExplanations
-theme(:wong)
-www_path = "docs/src/generators/www/"
-include("docs/src/utils.jl")
+include("docs/setup_docs.jl")
+eval(setup_docs)
 ```
 
 # `GreedyGenerator`
 
-Next we will repeat the exercise above, but instead use the `GreedyGenerator` in the context of a Bayesian classifier. This generator is greedy in the sense that it simply chooses the most salient feature {x′}ᵈ where 
+We use the term **greedy** to describe the counterfactual generator introduced by @schut2021generating.
 
-```math
-d=\arg\max_{d \in [1,D]} \nabla_{x\prime} \ell(M(x\prime),t)
-```
+## Description
 
-and perturbs it by a fixed amount $\delta$. In other words, optimization is penalty-free. This is possible in the Bayesian context, because maximizing the predictive probability corresponds to minimizing the predictive uncertainty: by construction the generated counterfactual will therefore be *realistic* (low epistemic uncertainty) and *unambiguous* (low aleotoric uncertainty).
+The Greedy generator works under the premise of generating realistic counterfactuals by minimizing predictive uncertainty. @schut2021generating show that for models that incorporates predictive uncertainty in their predictions, maximizing the predictive probability corresponds to minimizing the predictive uncertainty: by construction, the generated counterfactual will therefore be *realistic* (low epistemic uncertainty) and *unambiguous* (low aleatoric uncertainty). 
 
-```{julia}
-using MLJ
-N = 1000
-X, ys = make_blobs(N, 2; centers=2, as_table=false, center_box=(-2 => 2), cluster_std=0.1)
-ys .= ys.==2
-X = X'
-xs = Flux.unstack(X,2)
-data = zip(xs,ys)
-counterfactual_data = CounterfactualData(X,ys')
-```
+For the counterfactual search @schut2021generating propose using a Jacobian-based Saliency Map Attack(JSMA). It is **greedy** in the sense that it is an "iterative algorithm that updates the most salient feature, i.e. the feature that has the largest influence on the classification, by $\delta$ at each step" [@schut2021generating].
 
+## Usage
 
+The approach can be used in our package as follows:
 
 ```{julia}
-using Flux
-nn = Chain(Dense(2,1))
-using Flux.Optimise: update!, Adam
-opt = Adam()
-epochs = 100
-loss(x, y) = Flux.Losses.logitbinarycrossentropy(nn(x), y)
-avg_loss(data) = mean(map(d -> loss(d[1],d[2]), data))
-show_every = epochs/10
-for epoch = 1:epochs
-  for d in data
-    gs = gradient(Flux.params(nn)) do
-      l = loss(d...)
-    end
-    update!(opt, Flux.params(nn), gs)
-  end
-  if epoch % show_every == 0
-    println("Epoch " * string(epoch))
-    @show avg_loss(data)
-  end
-end
-```
-
-```{julia}
-using LaplaceRedux
[…]

[…]
