+++ docs/src/assets/_resources.md
@@ -8,3 +8,7 @@ Slides: [link](juliacon22/presentation.html)
 <iframe style="width:560px;height:315px" src="https://www.paltmeyer.com/CounterfactualExplanations.jl/dev/resources/juliacon22/presentation.html" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
 ```
 
+### JuliaCon Proceedings Paper
+
+TBD
+
