+++ docs/src/how_to_guides/interop.qmd
@@ -1,63 +1,26 @@
----
-format: 
-  commonmark:
-    variant: -raw_html
-    wrap: none
-    self-contained: true
-crossref:
-  fig-prefix: Figure
-  tbl-prefix: Table
-bibliography: https://raw.githubusercontent.com/pat-alt/bib/main/bib.bib
-output: asis
-execute: 
-  echo: true
-  eval: false
-jupyter: julia-1.7
----
-
 ```@meta
 CurrentModule = CounterfactualExplanations 
 ```
 
-# Interoperability
-
-```
-!!! info "Contributor's Guide" 
-    Our work on language interoperability is still experimental. In this tutorial we demonstrate our early attempts to extend the package to accommodate `torch` models trained in R and Python. The goal here is to provide a template and/or starting point for contributors that would like to add support for other models trained in foreign programming languages. If that contributor could be you, please see the related [issue](https://github.com/pat-alt/CounterfactualExplanations.jl/issues/67) and work in the linked [branch](https://github.com/pat-alt/CounterfactualExplanations.jl/tree/67-interoperability-broken-after-moving-to-julia-18).
-```
-
-```
-!!! info "Native support currently broken and removed" 
-    After bumping Julia version compatibility to `1.8`, interoperability broke for some reason (it was admittedly always experimental and hacky). The code that was previously used for interoperability was removed from the main branch, but still lives [here](https://github.com/pat-alt/CounterfactualExplanations.jl/tree/67-interoperability-broken-after-moving-to-julia-18).
-```
-
-The Julia language offers unique support for programming language interoperability. For example, calling Python and R is made remarkably easy through `PyCall.jl` (potentially `PythonCall.jl` is a better option) and `RCall.jl`. In this tutorial we will see how `CounterfactualExplanations.jl` can leverage this functionality. 
-
 ```{julia}
 #| echo: false
-using Plots, PlotThemes
-theme(:wong)
-using Logging
-disable_logging(Logging.Info)
-include("dev/utils.jl") # some plotting functions
-www_path = "docs/src/contributing/www"
+include("docs/setup_docs.jl")
+eval(setup_docs)
 ```
 
-To get started we will first load some two-dimensional toy data:
+# How to explain R/Python models
 
-```{julia}
-using Random
-# Some random data:
-Random.seed!(1234);
-N = 100
-using CounterfactualExplanations
-using CounterfactualExplanations.Data
-xs, ys = Data.toy_data_non_linear(N)
-X = hcat(xs...)
-counterfactual_data = CounterfactualData(X,ys')
-```
+The Julia language offers unique support for programming language interoperability. For example, calling Python and R is made remarkably easy through `PyCall.jl` (or `PythonCall.jl`) and `RCall.jl`. In this tutorial, we will see how `CounterfactualExplanations.jl` can leverage this functionality. 
 
 ## `torch` model trained in R
+
+First off, let's load some data:
+
+```{julia}
+counterfactual_data = load_circles()
+X = counterfactual_data.X
+y = vec(float.(counterfactual_data.y))
+```
 
 The code below builds a simple MLP in R:
 
@@ -67,7 +30,7 @@ R"""
 # Data
 library(torch)
 X <- torch_tensor(t($X))
-ys <- torch_tensor($ys)
+y <- torch_tensor($y)
 
 # Model:
 mlp <- nn_module(
[…]
