+++ docs/src/how_to_guides/custom_models.md
@@ -1,11 +1,6 @@
-```@meta
-CurrentModule = CounterfactualExplanations 
-```
 
-```{julia}
-#| echo: false
-include("docs/setup_docs.jl")
-eval(setup_docs)
+``` @meta
+CurrentModule = CounterfactualExplanations 
 ```
 
 # How to add Custom Models
@@ -16,14 +11,14 @@ Adding custom models is possible and relatively straightforward, as we will demo
 
 Apart from the default models you can use any arbitrary (differentiable) model and generate recourse in the same way as before. Only two steps are necessary to make your own Julia model compatible with this package:
 
-1. The model needs to be declared as a subtype of `<:CounterfactualExplanations.Models.AbstractFittedModel`.
-2. You need to extend the functions `CounterfactualExplanations.Models.logits` and `CounterfactualExplanations.Models.probs` for your custom model.
+1.  The model needs to be declared as a subtype of `<:CounterfactualExplanations.Models.AbstractFittedModel`.
+2.  You need to extend the functions `CounterfactualExplanations.Models.logits` and `CounterfactualExplanations.Models.probs` for your custom model.
 
 ### How `FluxModel` was added
 
-To demonstrate how this can be done in practice, we will reiterate here how native support for [`Flux.jl`](https://fluxml.ai/) models was enabled [@innes2018flux]. Once again we use synthetic data for an illustrative example. The code below loads the data and builds a simple model architecture that can be used for a multi-class prediction task. Note how outputs from the final layer are not passed through a softmax activation function, since the counterfactual loss is evaluated with respect to logits. The model is trained with dropout.
+To demonstrate how this can be done in practice, we will reiterate here how native support for [`Flux.jl`](https://fluxml.ai/) models was enabled (Innes 2018). Once again we use synthetic data for an illustrative example. The code below loads the data and builds a simple model architecture that can be used for a multi-class prediction task. Note how outputs from the final layer are not passed through a softmax activation function, since the counterfactual loss is evaluated with respect to logits. The model is trained with dropout.
 
-```{julia}
+``` julia
 # Data:
 N = 200
 counterfactual_data = load_blobs(N; centers=4, cluster_std=0.5)
@@ -60,7 +55,7 @@ end
 
 The code below implements the two steps that were necessary to make Flux models compatible with the package. We first declare our new struct as a subtype of `<:AbstractDifferentiableModel`, which itself is an abstract subtype of `<:AbstractFittedModel`. Computing logits amounts to just calling the model on inputs. Predicted probabilities for labels can in this case be computed by passing predicted logits through the softmax function. Finally, we just instantiate our model in the same way as always.
 
-```{julia}
+``` julia
 # Step 1)
 struct MyFluxModel <: AbstractDifferentiableModel
     model::Any
@@ -78,9 +73,7 @@ M = MyFluxModel(model, :classification_multi)
 
 The code below implements the counterfactual search and plots the results:
 
-```{julia}
-#| output: true
-
+``` julia
 factual_label = 4
 target = 2
 chosen = rand(findall(predict_label(M, counterfactual_data) .== factual_label))
@@ -92,5 +85,8 @@ ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
 plot(ce)
 ```
 
+![](custom_models_files/figure-commonmark/cell-5-output-1.svg)
+
 ## References
 
+Innes, Mike. 2018. “Flux: Elegant Machine Learning with Julia.” *Journal of Open Source Software* 3 (25): 602.
