### 2023-02-01 e71e5c6
fixed small bug in Claproar implementation (arose due to recent breaking changes)
