+++ docs/src/explanation/generators/gravitational.qmd
@@ -76,4 +76,40 @@ ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
 plot(ce)
 ```
 
+### Comparison to `GenericGenerator`
+
+The figure below compares the outcome for the `GenericGenerator` with different decision thresholds and the `GravitationalGenerator`. 
+
+```{julia}
+#| echo: false
+counterfactual_data = load_linearly_separable(1000)
+M = fit_model(counterfactual_data, :Linear)
+target = 1
+factual_label = 2
+chosen = rand(findall(predict_label(M, counterfactual_data) .== factual_label))
+x = select_factual(counterfactual_data, chosen)
+opt = Flux.Descent(0.01)
+```
+
+```{julia}
+#| output: true
+#| echo: false
+# Generators:
+generators = Dict(
+    "Generic (γ=0.5)" => GenericGenerator(opt = opt, decision_threshold=0.5),
+    "Generic (γ=0.9)" => GenericGenerator(opt = opt, decision_threshold=0.9),
+    "Gravitational" => GravitationalGenerator(opt = opt)
+)
+
+
+counterfactuals = Dict([name => generate_counterfactual(x, target, counterfactual_data, M, gen;) for (name, gen) in generators])
+# Plots:
+plts = []
+for (name,ce) ∈ counterfactuals
+    plt = plot(ce; title=name, colorbar=false, ticks = false, legend=false, zoom=0)
+    plts = vcat(plts..., plt)
+end
+plot(plts..., size=(550,200), layout=(1,3))
+```
+
 ## References
