+++ docs/src/index.md
@@ -7,33 +7,29 @@ CurrentModule = CounterfactualExplanations
 
 Documentation for [CounterfactualExplanations.jl](https://github.com/pat-alt/CounterfactualExplanations.jl).
 
-`CounterfactualExplanations.jl` is a package for generating Counterfactual Explanations (CE) and Algorithmic Recourse (AR) for black-box algorithms. Both CE and AR are related tools for explainable artificial intelligence (XAI). While the package is written purely in Julia, it can be used to explain machine learning algorithms developed and trained in other popular programming languages like Python and R. See below for short introduction and other resources or dive straight into the [docs](https://pat-alt.github.io/CounterfactualExplanations.jl/dev).
-
-## 🔁 Status
-
-This package is ready for research, but not for production. So far [@pat-alt](https://github.com/pat-alt) has developed this package entirely on his own and focussed on using it for his PhD research. The package is therefore still subject to some serious limitations (see below).
+`CounterfactualExplanations.jl` is a package for generating Counterfactual Explanations (CE) and Algorithmic Recourse (AR) for black-box algorithms. Both CE and AR are related tools for explainable artificial intelligence (XAI). While the package is written purely in Julia, it can be used to explain machine learning algorithms developed and trained in other popular programming languages like Python and R. See below for a short introduction and other resources or dive straight into the [docs](https://pat-alt.github.io/CounterfactualExplanations.jl/dev).
 
 ## Installation 🚩
 
-The first release of this package is now on [Julia’s General Registry](https://github.com/JuliaRegistries/General) and can be installed as follows:
+You can install the stable release from [Julia’s General Registry](https://github.com/JuliaRegistries/General) as follows:
 
 ``` julia
 using Pkg
 Pkg.add("CounterfactualExplanations")
 ```
 
-CounterfactualExplanations.jl is currently under active development. To install the development version of the package you can run the following command:
+`CounterfactualExplanations.jl` is under active development. To install the development version of the package you can run the following command:
 
 ``` julia
 using Pkg
 Pkg.add(url="https://github.com/pat-alt/CounterfactualExplanations.jl")
 ```
 
-## Background and motivation
+## Background and Motivation
 
 #### The Need for Explainability ⬛
 
-Machine learning models like deep neural networks have become so complex, opaque and underspecified in the data that they are generally considered as black boxes. Nonetheless, such models often play a key role in data-driven decision-making systems. This often creates the following problem: human operators in charge of such systems have to rely on them blindly, while those individuals subject to them generally have no way of challenging an undesirable outcome:
+Machine learning models like deep neural networks have become so complex, opaque and underspecified in the data that they are generally considered Black Boxes. Nonetheless, such models often play a key role in data-driven decision-making systems. This often creates the following problem: human operators in charge of such systems have to rely on them blindly, while those individuals subject to them generally have no way of challenging an undesirable outcome:
 
 > “You cannot appeal to (algorithms). They do not listen. Nor do they bend.”
 >
@@ -41,15 +37,15 @@ Machine learning models like deep neural networks have become so complex, opaque
 
 #### Enter: Counterfactual Explanations 🔮
 
[…]
