+++ docs/src/tutorials/whistle_stop.md
@@ -14,7 +14,7 @@ counterfactual_data = load_blobs(n_samples; k=n_dim, centers=n_classes)
 M = fit_model(counterfactual_data, model_name)
 ```
 
-The chart below visualizes our data along with the model predictions. In particular, the contour indicates the predicted probabilities generated by our classifier. By default, these are the predicted probabilities for *y* = 1, the first label. For multi-dimensional input data is compressed into two dimensions and the decision boundary is approximated using Nearest Neighbors (this is still somewhat experimental).
+The chart below visualizes our data along with the model predictions. In particular, the contour indicates the predicted probabilities generated by our classifier. By default, these are the predicted probabilities for $y=1$, the first label. For multi-dimensional input data is compressed into two dimensions and the decision boundary is approximated using Nearest Neighbors (this is still somewhat experimental).
 
 ``` julia
 plot(M, counterfactual_data)
