+++ docs/src/explanation/generators/dice.qmd
@@ -53,7 +53,7 @@ end
 gif(anim, joinpath(www_path, "dice_intro.gif"))
 ```
 
-![](../www/dice_intro.gif){#fig-intro}
+![](../../www/dice_intro.gif){#fig-intro}
 
 ## Usage
 
