+++ dev/examples/image/MNIST.qmd
@@ -1,217 +1,0 @@
-```@meta
-CurrentModule = CounterfactualExplanations 
-```
-
-# MNIST
-
-In this example we will see how different counterfactual generators can be used to explain deep learning models for image classification. In particular, we will look at MNIST data and visually inspect how the different generators perturb images of handwritten digits in order to change the predicted label to a target label. @fig-samples shows a random sample of handwritten digits.
-
-```{julia}
-#| echo: false
-using Pkg; Pkg.activate("docs")
-using CounterfactualExplanations, Plots, MLDatasets
-using MLDatasets: convert2image
-using BSON: @save, @load
-theme(:wong)
-using Logging
-disable_logging(Logging.Info)
-include("dev/utils.jl") # some plotting functions
-www_path = "docs/src/examples/image/www"
-```
-
-```{julia}
-#| echo: false
-data_train = MNIST(:train)
-train_x, train_y = data_train[:]
-input_dim = prod(size(train_x[:,:,1]))
-using Images, Random, StatsBase
-Random.seed!(1)
-n_samples = 10
-samples = train_x[:,:,sample(1:end, n_samples, replace=false)]
-mosaic = mosaicview([convert2image(MNIST, samples[:,:,i]) for i ∈ 1:n_samples]...,ncol=Int(n_samples/2))
-plt = plot(mosaic, size=(500,260), axis=nothing, background=:transparent)
-savefig(plt, joinpath(www_path, "mnist_samples.png"))
-```
-
-![A few random handwritten digits.](www/mnist_samples.png){#fig-samples}
-
-## Pre-trained classifiers
-
-Next we will load two pre-trained deep-learning classifiers:
-
-1. Simple MLP - `model`
-2. Deep ensemble - `ensemble`
-
-```{julia}
-using Flux
-using CounterfactualExplanations.Data: mnist_data, mnist_model, mnist_ensemble
-X, ys = mnist_data()
-model = mnist_model()
-ensemble = mnist_ensemble()
-```
-
-The following code just prepares the models to be used with `CounterfactualExplanations.jl`:
-
-```{julia}
-using CounterfactualExplanations, CounterfactualExplanations.Models
-import CounterfactualExplanations.Models: logits, probs # import functions in order to extend
-
-M = FluxModel(model, likelihood=:classification_multi)
-M_ensemble=FluxEnsemble(ensemble, likelihood=:classification_multi)
-```
-
-## Generating counterfactuals
-
-We will look at four different approaches here: 
-
-1. Generic approach for the MLP [@wachter2017counterfactual].
-2. Greedy approach for the MLP.
-3. Generic approach for the deep ensemble.
-4. Greedy approach for the deep ensemble [@schut2021generating].
-
-They can be implemented using the `GenericGenerator` and the `GreedyGenerator`.
-
-### Turning a 9 into a 4
-
-We will start with an example that should yield intuitive results: the process of turning a handwritten 9 in @fig-nine into a 4 is straight-forward for a human - just erase the top part. Let's see how the different algorithmic approaches perform. First, we preprocess the data below, where we impose that the features (pixel values) are constrained to the follwoing domain: $\mathcal{X}=[0,1]\subset\mathbb{R}$.
-
-```{julia}
-counterfactual_data = CounterfactualData(X,ys;domain=(0,1))
-```
-
[…]
