+++ dev/contributing/loss.qmd
@@ -1,358 +1,0 @@
-```@meta
-CurrentModule = CounterfactualExplanations 
-```
-
-# Loss functions
-
-For the computation of loss functions and their gradients we leverage the functionality already implmented in [Flux](https://fluxml.ai/). All of the loss functions from Flux have been imported: 
-
-```julia-repl
-julia> names(CounterfactualExplanations.Losses)
-19-element Vector{Symbol}:
- :Losses
- :binary_focal_loss
- :binarycrossentropy
- :crossentropy
- :ctc_loss
- :dice_coeff_loss
- :focal_loss
- :hinge_loss
- ⋮
- :logitbinarycrossentropy
- :logitcrossentropy
- :mae
- :mse
- :msle
- :poisson_loss
- :squared_hinge_loss
- :tversky_loss
-```
-
-## Classification
-
-For most classification tasks the default `:logitbinarycrossentropy` (binary) and `:logitcrossentropy` should be sufficient. For both choices the package has been tested and works natively. When using other loss functions, some caution is recommended though:
-
-```
-!!! warning "External loss functions"
-    Some margin-based loss functions like hinge loss do not expect inputs in the domain $\mathcal{Y}=\{0,1\}$, but rather $\mathcal{Y}=\{-1,1\}$. In those case one needs to ensure that the training labels $y$ are encoded accordingly. In order to use distance-based loss functions like mean squared error (MSE) loss needs to be computed with respect to probibilities rather than logits. This is currently not supported and we genenerally recommend not to use distance-based loss functions in the classification setting (more on this below).
-```
-
-## Regression
-
-At this point `CounterfactualExplanations.jl` is designed to be used with classification models, since the overwhelming majority of the existing literature on counterfactual explanations is set in this context. By default margin-based loss functions are used and computed with respect to logits (more on this below). To produce counterfactual explanations for regression problems users currently need to binarize the problem: let $t$ denote some target value for the continuous dependent variable $y$ in the regression context, then we could respecify the dependent variable as $\tilde{t}=0$ for all $y<t$ and $\tilde{t}=1$ otherwise. In future work we want to add full support for regression problems.
-
-# Methodological background
-
-This is a short tutorial on loss functions and gradients typically involved in counterfactual search. It involves more maths than perhaps some of the other tutorials. 
-
-```{julia}
-#| echo: false
-using Pkg; Pkg.activate("docs")
-using Zygote, Plots, PlotThemes, LinearAlgebra, CounterfactualExplanations
-theme(:wong)
-using Logging
-disable_logging(Logging.Info)
-include("dev/utils.jl") # some plotting functions
-www_path = "docs/src/contributing/www"
-```
-
-## General setup
-
-We begin by restating the general setup for generic counterfactual search. Let $t\in\{0,1\}$ denote the target label, $M$ the model (classifier) and $x\prime\in\mathbb{R}^D$ the vector of counterfactual features (we will assume all features are continuous). Then the differentiable optimization problem in algorithmic recourse is generally of the following form
-
-```math
-x\prime = \arg \min_{x\prime}  \ell(M(x\prime),t) + \lambda h(x\prime)
-```
-
[…]
