+++ dev/_metadata.yml
@@ -1,5 +1,5 @@
 format: 
   commonmark:
-    variant: -raw_html
+    variant: -raw_html+tex_math_dollars
     wrap: none
-    self-contained: true
+    mermaid-format: png

[…]
