+++ docs/src/explanation/index.md
@@ -1,0 +1,14 @@
+
+``` @meta
+CurrentModule = CounterfactualExplanations
+```
+
+# Explanation
+
+In this section you will find detailed explanations about the methodology and code.
+
+> Explanation clarifies, deepens and broadens the reader’s understanding of a subject.
+>
+> — [Diátaxis](https://diataxis.fr/explanation/)
+
+In other words, you come here because you are interested in understanding how all of this actually works 🤓.
