+++ dev/index.qmd
@@ -2,7 +2,7 @@
 CurrentModule = CounterfactualExplanations
 ```
 
-# CounterfactualExplanations
+![](assets/wide_logo.png)
 
 Documentation for [CounterfactualExplanations.jl](https://github.com/pat-alt/CounterfactualExplanations.jl).
 
