+++ README.md
@@ -3,13 +3,9 @@
 
 [![Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://pat-alt.github.io/CounterfactualExplanations.jl/stable) [![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://pat-alt.github.io/CounterfactualExplanations.jl/dev) [![Build Status](https://github.com/pat-alt/CounterfactualExplanations.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/pat-alt/CounterfactualExplanations.jl/actions/workflows/CI.yml?query=branch%3Amain) [![Coverage](https://codecov.io/gh/pat-alt/CounterfactualExplanations.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/pat-alt/CounterfactualExplanations.jl) [![Code Style: Blue](https://img.shields.io/badge/code%20style-blue-4495d1.svg)](https://github.com/invenia/BlueStyle) [![ColPrac: Contributor’s Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-blueviolet.png)](https://github.com/SciML/ColPrac)
 
-``` @meta
-CurrentModule = CounterfactualExplanations 
-```
-
 `CounterfactualExplanations.jl` is a package for generating Counterfactual Explanations (CE) and Algorithmic Recourse (AR) for black-box algorithms. Both CE and AR are related tools for explainable artificial intelligence (XAI). While the package is written purely in Julia, it can be used to explain machine learning algorithms developed and trained in other popular programming languages like Python and R. See below for a short introduction and other resources or dive straight into the [docs](https://pat-alt.github.io/CounterfactualExplanations.jl/dev).
 
-## Installation 🚩
+## 🚩 Installation
 
 You can install the stable release from [Julia’s General Registry](https://github.com/JuliaRegistries/General) as follows:
 
@@ -118,7 +114,7 @@ The animation below shows the resulting counterfactual path:
 
 ![](README_files/intro.gif)
 
-## Implemented Counterfactual Generators:
+## ☑️ Implemented Counterfactual Generators
 
 Currently, the following counterfactual generators are implemented:
 
@@ -157,7 +153,7 @@ If any of the below applies to you, this might be the right open-source project 
 
 For more details on how to contribute see [here](https://www.paltmeyer.com/CounterfactualExplanations.jl/dev/contributing/). Please follow the [SciML ColPrac guide](https://github.com/SciML/ColPrac).
 
-## Citation 🎓
+## 🎓 Citation
 
 If you want to use this codebase, please consider citing:
 
@@ -168,7 +164,7 @@ If you want to use this codebase, please consider citing:
       year = {2022}
     }
 
-## References 📚
+## 📚 References
 
 Altmeyer, Patrick, Giovan Angela, Aleksander Buszydlik, Karol Dobiczek, Arie van Deursen, and Cynthia Liem. 2023. “Endogenous Macrodynamics in Algorithmic Recourse.” In *First IEEE Conference on Secure and Trustworthy Machine Learning*.
 
