+++ dev/_intro.qmd
@@ -1,7 +1,3 @@
-```@meta
-CurrentModule = CounterfactualExplanations 
-```
-
 ```{julia}
 #| echo: false
 include("docs/setup_docs.jl")
@@ -10,7 +6,7 @@ eval(setup_docs)
 
 `CounterfactualExplanations.jl` is a package for generating Counterfactual Explanations (CE) and Algorithmic Recourse (AR) for black-box algorithms. Both CE and AR are related tools for explainable artificial intelligence (XAI). While the package is written purely in Julia, it can be used to explain machine learning algorithms developed and trained in other popular programming languages like Python and R. See below for a short introduction and other resources or dive straight into the [docs](https://pat-alt.github.io/CounterfactualExplanations.jl/dev).
 
-## Installation 🚩
+## 🚩 Installation 
 
 You can install the stable release from [Julia's General Registry](https://github.com/JuliaRegistries/General) as follows:
 
@@ -302,7 +298,7 @@ gif(anim, "README_files/intro.gif")
 
 ![](README_files/intro.gif)
 
-## Implemented Counterfactual Generators:
+## ☑️ Implemented Counterfactual Generators
 
 Currently, the following counterfactual generators are implemented:
 
@@ -341,7 +337,7 @@ If any of the below applies to you, this might be the right open-source project 
 
 For more details on how to contribute see [here](https://www.paltmeyer.com/CounterfactualExplanations.jl/dev/contributing/). Please follow the [SciML ColPrac guide](https://github.com/SciML/ColPrac).
 
-## Citation 🎓
+## 🎓 Citation 
 
 If you want to use this codebase, please consider citing:
 
@@ -354,4 +350,4 @@ If you want to use this codebase, please consider citing:
 }
 ```
 
-## References 📚
+## 📚 References 
