### 2023-02-03 4da7284
fixed some errors in docs
