+++ docs/src/reference.md
@@ -1,0 +1,48 @@
+
+``` @meta
+CurrentModule = CounterfactualExplanations 
+```
+
+# Reference
+
+In this reference, you will find a detailed overview of the package API.
+
+> Reference guides are technical descriptions of the machinery and how to operate it. Reference material is information-oriented.
+>
+> — [Diátaxis](https://diataxis.fr/reference/)
+
+In other words, you come here because you want to take a very close look at the code 🧐.
+
+## Content
+
+``` @contents
+Pages = ["_reference.md"]
+```
+
+## Exported functions
+
+``` @autodocs
+Modules = [
+    CounterfactualExplanations, 
+    CounterfactualExplanations.Data,
+    CounterfactualExplanations.DataPreprocessing,
+    CounterfactualExplanations.Models, 
+    CounterfactualExplanations.Generators, 
+    CounterfactualExplanations.Losses
+]
+Private = false
+```
+
+## Internal functions
+
+``` @autodocs
+Modules = [
+    CounterfactualExplanations, 
+    CounterfactualExplanations.Data,
+    CounterfactualExplanations.DataPreprocessing,
+    CounterfactualExplanations.Models, 
+    CounterfactualExplanations.Generators, 
+    CounterfactualExplanations.Losses
+]
+Public = false
+```
