+++ docs/src/assets/resources.qmd
@@ -1,6 +1,6 @@
-## Further Resources
+# Further Resources
 
-### JuliaCon 2022 
+## JuliaCon 2022 
 
 Slides: [link](juliacon22/presentation.html)
 
@@ -8,7 +8,7 @@ Slides: [link](juliacon22/presentation.html)
 <iframe style="width:560px;height:315px" src="https://www.paltmeyer.com/CounterfactualExplanations.jl/dev/resources/juliacon22/presentation.html" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
 ```
 
-### JuliaCon Proceedings Paper
+## JuliaCon Proceedings Paper
 
 TBD
 
