+++ dev/how_to_guides/custom_generators.qmd
@@ -8,7 +8,7 @@ include("docs/setup_docs.jl")
 eval(setup_docs)
 ```
 
-# How to add Custom Models
+# How to add Custom Generators
 
 As we will see in this short tutorial, building custom counterfactual generators is straightforward. We hope that this will facilitate contributions through the community.
 
