### 2023-02-13 3e8e006
fixed small bug in target_probs function
