+++ src/counterfactuals/plotting.jl
@@ -3,12 +3,34 @@ using Plots
 using Parameters
 using SliceMap
 
+"""
+    Plots.plot(
+        counterfactual_explanation::CounterfactualExplanation;
+        alpha_ = 0.5,
+        plot_up_to::Union{Nothing,Int} = nothing,
+        plot_proba::Bool = false,
+        kwargs...,
+    )
+
+Calling `plot` on an instance of type `CounterfactualExplanation` returns a plot that visualises the entire counterfactual path. For multi-dimensional input data, the data is first compressed into two dimensions. The decision boundary is then approximated using using a Nearest Neighbour classifier. This is still somewhat experimental at the moment.
+
+
+# Examples
+
+```julia-repl
+# Search:
+generator = GenericGenerator()
+ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
+
+plot(ce)
+```
+"""
 function Plots.plot(
     counterfactual_explanation::CounterfactualExplanation;
-    alpha_ = 0.5,
-    plot_up_to::Union{Nothing,Int} = nothing,
-    plot_proba::Bool = false,
-    kwargs...,
+    alpha_=0.5,
+    plot_up_to::Union{Nothing,Int}=nothing,
+    plot_proba::Bool=false,
+    kwargs...
 )
 
     T = total_steps(counterfactual_explanation)
@@ -18,9 +40,9 @@ function Plots.plot(
     T += 1
     ingredients = set_up_plots(
         counterfactual_explanation;
-        alpha = alpha_,
-        plot_proba = plot_proba,
-        kwargs...,
+        alpha=alpha_,
+        plot_proba=plot_proba,
+        kwargs...
     )
 
     for t ∈ 1:T
@@ -39,15 +61,25 @@ end
 """
     animate_path(counterfactual_explanation::CounterfactualExplanation, path=tempdir(); plot_proba::Bool=false, kwargs...)
 
-Animate the counterfactual path.
+Returns and animation of the counterfactual path.
+
+# Examples
+
+```julia-repl
+# Search:
+generator = GenericGenerator()
+ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
+
+animate_path(ce)
+```
 """
 function animate_path(
     counterfactual_explanation::CounterfactualExplanation,
-    path = tempdir();
-    alpha_ = 0.5,
-    plot_up_to::Union{Nothing,Int} = nothing,
-    plot_proba::Bool = false,
-    kwargs...,
+    path=tempdir();
+    alpha_=0.5,
+    plot_up_to::Union{Nothing,Int}=nothing,
+    plot_proba::Bool=false,
+    kwargs...
 )
     T = total_steps(counterfactual_explanation)
     T =
@@ -56,9 +88,9 @@ function animate_path(
     T += 1
     ingredients = set_up_plots(
         counterfactual_explanation;
-        alpha = alpha_,
-        plot_proba = plot_proba,
-        kwargs...,
+        alpha=alpha_,
+        plot_proba=plot_proba,
+        kwargs...
     )
 
     anim = @animate for t ∈ 1:T
@@ -70,24 +102,33 @@ function animate_path(
     return anim
 end
 
+"""
+    plot_state(
+        counterfactual_explanation::CounterfactualExplanation,
+        t::Int,
+        final_sate::Bool;
+        kwargs...
+    )
 
+Helper function that plots a single step of the counterfactual path.
+"""
 function plot_state(
     counterfactual_explanation::CounterfactualExplanation,
     t::Int,
     final_sate::Bool;
-    kwargs...,
+    kwargs...
 )
     args = PlotIngredients(; kwargs...)
[…]
