+++ docs/src/reference.md
@@ -16,7 +16,8 @@ In other words, you come here because you want to take a very close look at the 
 ## Content
 
 ``` @contents
-Pages = ["_reference.md"]
+Pages = ["reference.md"]
+Depth = 3
 ```
 
 ## Exported functions
