+++ src/data_preprocessing/utils.jl
@@ -16,12 +16,23 @@ end
 
 Splits data into train and test split.
 """
-function train_test_split(data::CounterfactualData; test_size=0.2)
+function train_test_split(data::CounterfactualData; test_size = 0.2)
     N = size(data.X, 2)
     classes_ = data.y_levels
     n_per_class = round(N / length(classes_))
     y = data.output_encoder.y
-    test_idx = sort(reduce(vcat, [sample(findall(vec(y .== cls)), Int(floor(test_size * n_per_class)), replace=false) for cls in classes_]))
+    test_idx = sort(
+        reduce(
+            vcat,
+            [
+                sample(
+                    findall(vec(y .== cls)),
+                    Int(floor(test_size * n_per_class)),
+                    replace = false,
+                ) for cls in classes_
+            ],
+        ),
+    )
     train_idx = setdiff(1:N, test_idx)
     train_data = _subset(data, train_idx)
     test_data = _subset(data, test_idx)
@@ -41,7 +52,15 @@ function undersample(data::CounterfactualData, n::Int)
     n_per_class = Int(round(n / n_classes))
     y_cls = data.output_encoder.labels
 
-    idx = sort(reduce(vcat, [sample(findall(vec(y_cls .== cls)), n_per_class, replace=false) for cls in classes_]))
+    idx = sort(
+        reduce(
+            vcat,
+            [
+                sample(findall(vec(y_cls .== cls)), n_per_class, replace = false) for
+                cls in classes_
+            ],
+        ),
+    )
     data.X = X[:, idx]
     data.y = y[:, idx]
     data.output_encoder.y = data.output_encoder.y[idx]
