+++ src/models/Models.jl
@@ -38,11 +38,8 @@ include("pretrained.jl")
 
 A dictionary containing all trainable machine learning models.
 """
-const model_catalogue = Dict(
-    :Linear => Linear,
-    :MLP => FluxModel,
-    :DeepEnsemble => FluxEnsemble,
-)
+const model_catalogue =
+    Dict(:Linear => Linear, :MLP => FluxModel, :DeepEnsemble => FluxEnsemble)
 
 """
     fit_model(
@@ -52,10 +49,7 @@ const model_catalogue = Dict(
 
 Fits one of the available default models to the `counterfactual_data`. The `model` argument can be used to specify the desired model. The available values correspond to the keys of the [`model_catalogue`](@ref) dictionary.
 """
-function fit_model(
-    counterfactual_data::CounterfactualData, model::Symbol=:MLP;
-    kwrgs...
-)
+function fit_model(counterfactual_data::CounterfactualData, model::Symbol = :MLP; kwrgs...)
     @assert model in keys(model_catalogue) "Specified model does not match any of the models available in the `model_catalogue`."
 
     # Set up:
