+++ src/benchmark/functions.jl
@@ -6,7 +6,7 @@ using Statistics
 
 Checks of the counterfactual search has been successful. In case multiple counterfactuals were generated, the function returns the proportion of successful counterfactuals.
 """
-function success_rate(counterfactual_explanation::CounterfactualExplanation; agg=mean)
+function success_rate(counterfactual_explanation::CounterfactualExplanation; agg = mean)
     agg(
         CounterfactualExplanations.target_probs(counterfactual_explanation) .>=
         counterfactual_explanation.params[:γ],
@@ -23,7 +23,7 @@ Computes the proportion of successful counterfactuals acress a vector of counter
 """
 function success_rate(
     counterfactual_explanations::Vector{CounterfactualExplanation};
-    agg=mean
+    agg = mean,
 )
     agg(success_rate.(counterfactual_explanations))
 end
@@ -33,7 +33,7 @@ end
 
 Computes the Euclidean distance of the counterfactual to the original factual.
 """
-function distance(counterfactual_explanation::CounterfactualExplanation; agg=mean)
+function distance(counterfactual_explanation::CounterfactualExplanation; agg = mean)
     x = CounterfactualExplanations.factual(counterfactual_explanation)
     x′ = CounterfactualExplanations.counterfactual(counterfactual_explanation)
     return agg(LinearAlgebra.norm.(x .- x′))
@@ -49,7 +49,7 @@ Computes the average Euclidean distance of multiple counterfactuals from their c
 """
 function distance(
     counterfactual_explanations::Vector{CounterfactualExplanation};
-    agg=mean
+    agg = mean,
 )
     agg(distance.(counterfactual_explanations))
 end
@@ -59,9 +59,9 @@ end
 
 Computes the feature redundancy: that is, the number of features that remain unchanged from their original, factual values.
 """
-function redundancy(counterfactual_explanation::CounterfactualExplanation; agg=mean)
+function redundancy(counterfactual_explanation::CounterfactualExplanation; agg = mean)
     x′ = CounterfactualExplanations.counterfactual(counterfactual_explanation)
-    redundant_x = mapslices(x -> sum(x .== 0) / size(x, 1), x′, dims=[1, 2])
+    redundant_x = mapslices(x -> sum(x .== 0) / size(x, 1), x′, dims = [1, 2])
     return agg(redundant_x)
 end
 
@@ -75,7 +75,7 @@ Computes the average redundancy across multiple counterfactuals.
 """
 function redundancy(
     counterfactual_explanations::Vector{CounterfactualExplanation};
-    agg=mean
+    agg = mean,
 )
     agg(redundancy.(counterfactual_explanations))
 end
