### 2023-02-16 e85029b
moving to separate module for losses instead of relying just on Flux
