+++ docs/src/index.md
@@ -5,9 +5,9 @@ CurrentModule = CounterfactualExplanations
 
 ![](assets/wide_logo.png)
 
-Documentation for [CounterfactualExplanations.jl](https://github.com/pat-alt/CounterfactualExplanations.jl).
+Documentation for [CounterfactualExplanations.jl](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl).
 
-`CounterfactualExplanations.jl` is a package for generating Counterfactual Explanations (CE) and Algorithmic Recourse (AR) for black-box algorithms. Both CE and AR are related tools for explainable artificial intelligence (XAI). While the package is written purely in Julia, it can be used to explain machine learning algorithms developed and trained in other popular programming languages like Python and R. See below for a short introduction and other resources or dive straight into the [docs](https://pat-alt.github.io/CounterfactualExplanations.jl/dev).
+`CounterfactualExplanations.jl` is a package for generating Counterfactual Explanations (CE) and Algorithmic Recourse (AR) for black-box algorithms. Both CE and AR are related tools for explainable artificial intelligence (XAI). While the package is written purely in Julia, it can be used to explain machine learning algorithms developed and trained in other popular programming languages like Python and R. See below for a short introduction and other resources or dive straight into the [docs](https://juliatrustworthyai.github.io/CounterfactualExplanations.jl/dev).
 
 ## 🚩 Installation
 
@@ -22,7 +22,7 @@ Pkg.add("CounterfactualExplanations")
 
 ``` julia
 using Pkg
-Pkg.add(url="https://github.com/pat-alt/CounterfactualExplanations.jl")
+Pkg.add(url="https://github.com/juliatrustworthyai/CounterfactualExplanations.jl")
 ```
 
 ## 🤔 Background and Motivation
@@ -145,7 +145,7 @@ Additionally, our ambition is to enhance the package through the following featu
 
 ## 🛠 Contribute
 
-Contributions of any kind are very much welcome! Take a look at the [issue](https://github.com/pat-alt/CounterfactualExplanations.jl/issues) to see what things we are currently working on.
+Contributions of any kind are very much welcome! Take a look at the [issue](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues) to see what things we are currently working on.
 
 If any of the below applies to you, this might be the right open-source project for you:
 
@@ -153,7 +153,7 @@ If any of the below applies to you, this might be the right open-source project 
 - You’re experienced with Julia and are happy to help someone less experienced to up their game. Ideally, you are also curious about Trustworthy AI.
 - You’re new to Julia and open-source development and would like to start your learning journey by contributing to a recent and active development. Ideally, you are familiar with machine learning.
 
-[@pat-alt](https://github.com/pat-alt) here: I am still very much at the beginning of my Julia journey, so if you spot any issues or have any suggestions for design improvement, please just open [issue](https://github.com/pat-alt/CounterfactualExplanations.jl/issues) or start a [discussion](https://github.com/pat-alt/CounterfactualExplanations.jl/discussions). Our goal is to provide a go-to place for counterfactual explanations in Julia.
+[@pat-alt](https://github.com/pat-alt) here: I am still very much at the beginning of my Julia journey, so if you spot any issues or have any suggestions for design improvement, please just open [issue](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues) or start a [discussion](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/discussions). Our goal is to provide a go-to place for counterfactual explanations in Julia.
 
[…]
