+++ docs/src/index.qmd
@@ -4,6 +4,6 @@ CurrentModule = CounterfactualExplanations
 
 ![](assets/wide_logo.png)
 
-Documentation for [CounterfactualExplanations.jl](https://github.com/pat-alt/CounterfactualExplanations.jl).
+Documentation for [CounterfactualExplanations.jl](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl).
 
 {{< include _intro.qmd >}}

[…]
