### 2023-02-24 e420a16
minor update to counterfactual_label function
