+++ Project.toml
@@ -1,5 +1,6 @@
 name = "CounterfactualExplanations"
 uuid = "2f13d31b-18db-44c1-bc43-ebaf2cff0be0"
+repo = "https://github.com/JuliaTrustworthyAI/CounterfactualExplanations.jl"
 authors = ["Patrick Altmeyer <p.altmeyer@tudelft.nl>"]
 version = "0.1.7"
 
