+++ src/objectives/Objectives.jl
@@ -1,0 +1,10 @@
+module Objectives
+
+using ..CounterfactualExplanations
+
+include("loss_functions.jl")
+export logitbinarycrossentropy, logitcrossentropy, mse
+
+include("penalties.jl")
+
+end
