+++ src/global_utils.jl
@@ -3,6 +3,9 @@ using Flux
 using MLJBase
 using Parameters
 
+# Abstract Base Types
+
+
 # Constants:
 """
     RawTargetType
