+++ src/base_types.jl
@@ -1,0 +1,9 @@
+"Base type for counterfactual explanations."
+abstract type AbstractCounterfactualExplanation end
+
+"Base type for fitted models."
+abstract type AbstractFittedModel end
+
+"An abstract type that serves as the base type for counterfactual generators."
+abstract type AbstractGenerator end
+
