+++ docs/src/tutorials/benchmarking.qmd
@@ -1,0 +1,18 @@
+```@meta
+CurrentModule = CounterfactualExplanations 
+```
+
+```{julia}
+#| echo: false
+include("docs/setup_docs.jl")
+eval(setup_docs)
+```
+
+# Benchmarking
+
+Comparing the performance of different counterfactual generators is an important task. Researchers can use benchmarks to test new ideas they want to implement. Practitioners can find the right counterfactual generator for their specific use case through benchmarks. In this tutorial, we will see how to run benchmarks for counterfactual generators. 
+
+
+```{julia}
+counterfactual_data = load_linearly_separable()
+```
