+++ src/benchmark/Benchmark.jl
@@ -1,10 +1,45 @@
 module Benchmark
-using DataFrames
+
 using ..CounterfactualExplanations
+using DataFrames
+using Statistics
 
-export benchmark
+export benchmark, evaluate, default_measures
+export validity, distance, redundancy
 
-include("functions.jl")
+include("measures.jl")
+const default_measures = [
+    validity,
+    distance,
+    redundancy
+]
+
+"""
+    evaluate(
+        counterfactual_explanation::Union{
+            CounterfactualExplanation,
+            Vector{CounterfactualExplanation},
+        };
+        measure::Union{Function,Vector{Function}} = default_measures,
+    )
+
+Just computes evaluation `measures` for the counterfactual explanation.
+"""
+function evaluate(
+    counterfactual_explanation::Union{
+        CounterfactualExplanation,
+        Vector{CounterfactualExplanation},
+    };
+    measure::Union{Function,Vector{Function}}=default_measures,
+    agg::Function=mean,
+    report_each::Bool=false,
+)
+    if typeof(measure) <: Function
+        measure = [measure]
+    end
+    agg = report_each ? (x -> x) : agg
+    return [fun(counterfactual_explanation; agg=agg) for fun in measure]
+end
 
 """
     benchmark(
@@ -22,10 +57,10 @@ function benchmark(
         CounterfactualExplanation,
         Vector{CounterfactualExplanation},
     };
-    to_dataframe = true,
+    to_dataframe=true
 )
     bmk = Dict(
-        :success_rate => success_rate(counterfactual_explanation),
+        :validity => validity(counterfactual_explanation),
         :distance => distance(counterfactual_explanation),
         :redundancy => redundancy(counterfactual_explanation),
     )
