+++ docs/setup_docs.jl
@@ -5,9 +5,9 @@ setup_docs = quote
 
     using CounterfactualExplanations
     using CounterfactualExplanations: animate_path, counterfactual, counterfactual_label
-    using CounterfactualExplanations.Models
     using CounterfactualExplanations.Data
     using CounterfactualExplanations.DataPreprocessing: unpack_data
+    using CounterfactualExplanations.Models
     using Flux
     using Flux.Optimise: update!, Adam
     using Images
