+++ docs/src/reference.qmd
@@ -23,6 +23,7 @@ Depth = 3
 ```@autodocs
 Modules = [
     CounterfactualExplanations, 
+    CounterfactualExplanations.Benchmark,
     CounterfactualExplanations.Data,
     CounterfactualExplanations.DataPreprocessing,
     CounterfactualExplanations.Models, 
@@ -37,6 +38,7 @@ Private = false
 ```@autodocs
 Modules = [
     CounterfactualExplanations, 
+    CounterfactualExplanations.Benchmark,
     CounterfactualExplanations.Data,
     CounterfactualExplanations.DataPreprocessing,
     CounterfactualExplanations.Models, 
