+++ src/benchmark/measures.jl
@@ -11,16 +11,44 @@ function validity(counterfactual_explanation::CounterfactualExplanation)
 end
 
 """
-    distance(counterfactual_explanation::CounterfactualExplanation)
+    distance(counterfactual_explanation::CounterfactualExplanation, p::Real=2)
 
-Computes the Euclidean distance of the counterfactual to the original factual.
+Computes the distance of the counterfactual to the original factual.
 """
-function distance(counterfactual_explanation::CounterfactualExplanation)
+function distance(counterfactual_explanation::CounterfactualExplanation, p::Real=2)
     x = CounterfactualExplanations.factual(counterfactual_explanation)
     x′ = CounterfactualExplanations.counterfactual(counterfactual_explanation)
-    Δ = mapslices(_x -> LinearAlgebra.norm(_x .- x), x′, dims=[1, 2])
+    Δ = mapslices(_x -> LinearAlgebra.norm(_x .- x, p), x′, dims=[1, 2])
     return Δ
 end
+
+"""
+    distance_l0(counterfactual_explanaation::CounterfactualExplanation)
+
+Computes the L0 distance of the counterfactual to the original factual.
+"""
+distance_l0(counterfactual_explanation::CounterfactualExplanation) = distance(counterfactual_explanation, 0)
+
+"""
+    distance_l1(counterfactual_explanation::CounterfactualExplanation)
+
+Computes the L1 distance of the counterfactual to the original factual.
+"""
+distance_l1(counterfactual_explanation::CounterfactualExplanation) = distance(counterfactual_explanation, 1)
+
+"""
+    distance_l2(counterfactual_explanation::CounterfactualExplanation)
+
+Computes the L2 (Euclidean) distance of the counterfactual to the original factual.
+"""
+distance_l2(counterfactual_explanation::CounterfactualExplanation) = distance(counterfactual_explanation, 2)
+
+"""
+    distance_linf(counterfactual_explanation::CounterfactualExplanation)
+
+Computes the L-inf distance of the counterfactual to the original factual.
+"""
+distance_linf(counterfactual_explanation::CounterfactualExplanation) = distance(counterfactual_explanation, Inf)
 
 """
     redundancy(counterfactual_explanation::CounterfactualExplanation)
@@ -29,7 +57,7 @@ Computes the feature redundancy: that is, the number of features that remain unc
 """
 function redundancy(counterfactual_explanation::CounterfactualExplanation)
     x′ = CounterfactualExplanations.counterfactual(counterfactual_explanation)
-    redundant_x = mapslices(x -> sum(x .== 0) / size(x, 1), x′, dims = [1, 2])
+    redundant_x = mapslices(x -> sum(x .== 0) / size(x, 1), x′, dims=[1, 2])
     return redundant_x
 end
 
