+++ docs/src/tutorials/evaluation.qmd
@@ -14,7 +14,11 @@ Comparing the performance of different counterfactual generators is an important
 
 ## Default Measures
 
-Numerous evaluation measures for counterfactual explanations have been proposed. One of the most important measures is [`validity`](@ref), which simply determines whether or not a counterfactual explanation $x^{\prime}$ is valid in the sense that it yields the target prediction: $M(x^{\prime})=t$. We can evaluate the validity of a single counterfactual explanation `ce` using the [`Benchmark.evaluate`](@ref) function as follows:
+Numerous evaluation measures for counterfactual explanations have been proposed. 
+
+### Single Measure, Single Counterfactual
+
+One of the most important measures is [`validity`](@ref), which simply determines whether or not a counterfactual explanation $x^{\prime}$ is valid in the sense that it yields the target prediction: $M(x^{\prime})=t$. We can evaluate the validity of a single counterfactual explanation `ce` using the [`Benchmark.evaluate`](@ref) function as follows:
 
 ```{julia}
 using CounterfactualExplanations.Benchmark: evaluate, validity
@@ -28,11 +32,36 @@ using CounterfactualExplanations.Benchmark: distance
 evaluate(ce; measure=distance)
 ```
 
-If no `measure` is specified, the `evaluate` method will return all default measures:
+By default, `distance` computes the L2 (Euclidean) distance. 
+
+### Multiple Measures, Single Counterfactual
+
+You might be interested in computing not just the L2 distance, but various LP norms. This can be done by supplying a vector of functions to the `measure` key argument. For convenience, all default distance measures have already been collected in a vector:
+
+```{julia}
+using CounterfactualExplanations.Benchmark: distance_measures
+distance_measures
+```
+
+We can use this vector of evaluation measures as follows:
+
+```{julia}
+evaluate(ce; measure=distance_measures)
+```
+
+If no `measure` is specified, the `evaluate` method will return all default measures,
+
+```{julia}
+evaluate(ce)
+```
+
+which include:
 
 ```{julia}
 CounterfactualExplanations.Benchmark.default_measures
 ```
+
+### Multiple Measures and Counterfactuals
 
 We can also evaluate multiple counterfactual explanations at once:
 
@@ -42,7 +71,7 @@ ces = generate_counterfactual(x, target, counterfactual_data, M, generator; num_
 evaluate(ces)
 ```
 
-By default, the computed evaluation measures are aggregated across all counterfactual explanations. To return individual measures for each counterfactual explanation you can specify `report_each=true`
+By default, each evaluation measure is aggregated across all counterfactual explanations. To return individual measures for each counterfactual explanation you can specify `report_each=true`
 
 ```{julia}
 evaluate(ces; report_each=true)
