+++ src/evaluation/benchmark.jl
@@ -1,0 +1,3 @@
+struct Benchmark
+    counterfactual_explanations::Vector{CounterfactualExplanation}
+end
