+++ docs/src/tutorials/evaluation.md
@@ -1,22 +1,11 @@
----
-execute: 
-  output: true
----
 
-```@meta
+``` @meta
 CurrentModule = CounterfactualExplanations 
-```
-
-```{julia}
-#| echo: false
-#| output: false
-include("docs/setup_docs.jl")
-eval(setup_docs)
 ```
 
 # Performance Evaluation
 
-Now that we know how to generate counterfactual explanations in Julia, you may have a few follow-up questions: How do I know if the counterfactual search has been successful? How good is my counterfactual explanation? What does 'good' even mean in this context? In this tutorial, we will see how counterfactual explanations can be evaluated with respect to their performance.
+Now that we know how to generate counterfactual explanations in Julia, you may have a few follow-up questions: How do I know if the counterfactual search has been successful? How good is my counterfactual explanation? What does ‘good’ even mean in this context? In this tutorial, we will see how counterfactual explanations can be evaluated with respect to their performance.
 
 ## Default Measures
 
@@ -26,99 +15,130 @@ Numerous evaluation measures for counterfactual explanations have been proposed.
 
 One of the most important measures is [`validity`](@ref), which simply determines whether or not a counterfactual explanation $x^{\prime}$ is valid in the sense that it yields the target prediction: $M(x^{\prime})=t$. We can evaluate the validity of a single counterfactual explanation `ce` using the [`Evaluation.evaluate`](@ref) function as follows:
 
-```{julia}
+``` julia
 using CounterfactualExplanations.Evaluation: evaluate, validity
 evaluate(ce; measure=validity)
 ```
 
-For a single counterfactual explanation, this evaluation measure can only take two values: it is either equal to `1`, if the explanation is valid or `0` otherwise. Another important measure is [`distance`](@ref), which relates to the distance between the factual $x$ and the counterfactual $x^{\prime}$. In the context of Algorithmic Recourse, higher distances are typically associated with higher costs to individuals seeking recourse. 
+    1-element Vector{Vector{Float64}}:
+     [1.0]
 
-```{julia}
+For a single counterfactual explanation, this evaluation measure can only take two values: it is either equal to `1`, if the explanation is valid or `0` otherwise. Another important measure is [`distance`](@ref), which relates to the distance between the factual $x$ and the counterfactual $x^{\prime}$. In the context of Algorithmic Recourse, higher distances are typically associated with higher costs to individuals seeking recourse.
+
+``` julia
 using CounterfactualExplanations.Evaluation: distance
 evaluate(ce; measure=distance)
 ```
 
-By default, `distance` computes the L2 (Euclidean) distance. 
+    1-element Vector{Vector{Float64}}:
+     [0.7746574005639171]
+
+By default, `distance` computes the L2 (Euclidean) distance.
 
 ### Multiple Measures, Single Counterfactual
 
 You might be interested in computing not just the L2 distance, but various LP norms. This can be done by supplying a vector of functions to the `measure` key argument. For convenience, all default distance measures have already been collected in a vector:
 
-```{julia}
+``` julia
 using CounterfactualExplanations.Evaluation: distance_measures
 distance_measures
 ```
 
+    4-element Vector{Function}:
+     distance_l0 (generic function with 1 method)
+     distance_l1 (generic function with 1 method)
+     distance_l2 (generic function with 1 method)
+     distance_linf (generic function with 1 method)
+
 We can use this vector of evaluation measures as follows:
[…]
