+++ src/counterfactuals/functions.jl
@@ -45,11 +45,11 @@ function CounterfactualExplanation(;
     data::CounterfactualData,
     M::Models.AbstractFittedModel,
     generator::Generators.AbstractGenerator,
-    T::Int = 100,
-    latent_space::Union{Nothing,Bool} = nothing,
-    num_counterfactuals::Int = 1,
-    initialization::Symbol = :add_perturbation,
-    generative_model_params::NamedTuple = (;),
+    T::Int=100,
+    latent_space::Union{Nothing,Bool}=nothing,
+    num_counterfactuals::Int=1,
+    initialization::Symbol=:add_perturbation,
+    generative_model_params::NamedTuple=(;)
 )
 
     # Factual:
@@ -144,7 +144,7 @@ function adjust_shape(
             )
         )
     s′ = copy(x)
-    s′ = repeat(x, outer = size_)
+    s′ = repeat(x, outer=size_)
 
     return s′
 
@@ -187,7 +187,7 @@ Finally, it returns the encoded state variable.
 """
 function encode_state(
     counterfactual_explanation::CounterfactualExplanation,
-    x::Union{AbstractArray,Nothing} = nothing,
+    x::Union{AbstractArray,Nothing}=nothing,
 )
 
     # Unpack:
@@ -204,7 +204,7 @@ function encode_state(
     if !counterfactual_explanation.latent_space
         dt = data.dt
         idx = transformable_features(data)
-        SliceMap.slicemap(s′, dims = (1, 2)) do s
+        SliceMap.slicemap(s′, dims=(1, 2)) do s
             _s = s[idx, :]
             StatsBase.transform!(dt, _s)
             s[idx, :] = _s
@@ -255,7 +255,7 @@ Maps `x` from the feature space $\mathcal{X}$ to the latent space learned by the
 """
 function map_to_latent(
     counterfactual_explanation::CounterfactualExplanation,
-    x::Union{AbstractArray,Nothing} = nothing,
+    x::Union{AbstractArray,Nothing}=nothing,
 )
 
     # Unpack:
@@ -267,7 +267,7 @@ function map_to_latent(
         @info "Searching in latent space using generative model."
         generative_model = DataPreprocessing.get_generative_model(
             data;
-            counterfactual_explanation.generative_model_params...,
+            counterfactual_explanation.generative_model_params...
         )
         # map counterfactual to latent space: s′=z′∼p(z|x)
         s′, _, _ = GenerativeModels.rand(generative_model.encoder, s′)
@@ -293,7 +293,7 @@ Finally, the decoded counterfactual is returned.
 """
 function decode_state(
     counterfactual_explanation::CounterfactualExplanation,
-    x::Union{AbstractArray,Nothing} = nothing,
+    x::Union{AbstractArray,Nothing}=nothing,
 )
 
     # Unpack:
@@ -312,7 +312,7 @@ function decode_state(
 
         # Continuous:
         idx = transformable_features(data)
-        SliceMap.slicemap(s′, dims = (1, 2)) do s
+        SliceMap.slicemap(s′, dims=(1, 2)) do s
             _s = s[idx, :]
             StatsBase.reconstruct!(dt, _s)
             s[idx, :] = _s
@@ -337,7 +337,7 @@ Maps the state variable back from the latent space to the feature space.
 """
 function map_from_latent(
     counterfactual_explanation::CounterfactualExplanation,
-    x::Union{AbstractArray,Nothing} = nothing,
+    x::Union{AbstractArray,Nothing}=nothing,
 )
 
     # Unpack:
@@ -371,13 +371,13 @@ Reconstructs all categorical encodings. See [`DataPreprocessing.reconstruct_cat_
 """
 function reconstruct_cat_encoding(
[…]
