### 2023-03-09 9802d4c
benchmarking nearly sorted
