### 2023-03-09 cd8c8cd
benchmarking nearly sorted
