+++ src/evaluation/benchmark.jl
@@ -6,6 +6,18 @@ struct Benchmark
 end
 
 """
+    Base.vcat(bmk1::Benchmark, bmk2::Benchmark)
+
+Vertically concatenates two `Benchmark` objects.
+"""
+function Base.vcat(bmk1::Benchmark, bmk2::Benchmark)
+    ces = vcat(bmk1.counterfactual_explanations, bmk2.counterfactual_explanations)
+    evaluation = vcat(bmk1.evaluation, bmk2.evaluation)
+    bmk = Benchmark(ces, evaluation)
+    return bmk
+end
+
+"""
     benchmark(
         counterfactual_explanations::Vector{CounterfactualExplanation};
         meta_data::Union{Nothing,<:Vector{<:Dict}}=nothing,
@@ -24,12 +36,25 @@ function benchmark(
     return bmk
 end
 
+"""
+    benchmark(
+        x::Union{AbstractArray,Base.Iterators.Zip},
+        target::RawTargetType,
+        data::CounterfactualData;
+        models::Dict{<:Any, <:AbstractFittedModel},
+        generators::Dict{<:Any, <:AbstractGenerator},
+        measure::Union{Function,Vector{Function}}=default_measures,
+        kwrgs...
+    )
+
+First generates counterfactual explanations for factual `x`, the `target` and `data` using each of the provided `models` and `generators`. Then generates a `Benchmark` for the vector of counterfactual explanations as in [`benchmark(counterfactual_explanations::Vector{CounterfactualExplanation})`](@ref).
+"""
 function benchmark(
     x::Union{AbstractArray,Base.Iterators.Zip},
     target::RawTargetType,
     data::CounterfactualData;
-    models::Dict{<:Any, <:AbstractFittedModel},
-    generators::Dict{<:Any, <:AbstractGenerator},
+    models::Dict{<:Any,<:AbstractFittedModel},
+    generators::Dict{<:Any,<:AbstractGenerator},
     measure::Union{Function,Vector{Function}}=default_measures,
     kwrgs...
 )
@@ -37,9 +62,11 @@ function benchmark(
     meta_data = Vector{Dict}()
     ces = Vector{CounterfactualExplanation}()
     for (model_name, model) in models, (gen_name, generator) in generators
-        ce = generate_counterfactual(x,target,data,model,generator; kwrgs...)
-        push!(ces, ce)
-        push!(meta_data, Dict(:model => model_name, :generator => gen_name))
+        _ces = generate_counterfactual(x, target, data, model, generator; kwrgs...)
+        _ces = typeof(_ces) <: CounterfactualExplanation ? [_ces] : _ces
+        push!(ces, _ces...)
+        _meta_data = [Dict(:model => model_name, :generator => gen_name) for i in eachindex(_ces)]
+        push!(meta_data, _meta_data...)
     end
 
     # Performance Evaluation:
@@ -47,22 +74,52 @@ function benchmark(
     return bmk
 end
 
-# function benchmark(
-#     data::CounterfactualData;
-#     models::Union{Nothing,Dict{<:Any,<:AbstractFittedModel}},
-#     generators::Dict{<:Any,<:AbstractGenerator},
-#     measure::Union{Function,Vector{Function}}=default_measures,
-#     n_individuals::Int=1,
-#     kwrgs...
-# )
-#     # Setup
-#     factual = rand(data.y_levels)
-#     target = rand(data.y_levels[data.y_levels .!= factual])
-#     chosen = rand(findall(predict_label(M, counterfactual_data) .== factual), n_individuals)
-#     x = select_factual(counterfactual_data,chosen)
+"""
+    benchmark(
+        data::CounterfactualData;
+        models::Dict{Symbol,Any}=model_catalogue,
[…]
