+++ docs/Project.toml
@@ -18,8 +18,6 @@ ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
 ImageMagick = "6218d12a-5da1-5696-b52f-db25d2ecc6d1"
 ImageTransformations = "02fcd773-0e25-5acc-982a-7f6622650795"
 Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
-JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
-JSONTables = "b9914132-a727-11e9-1322-f18e41205b0b"
 LaplaceRedux = "c52c1a26-f7c5-402b-80be-ba1e638ad478"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 MLDatasets = "eb30cadb-4394-5ae3-aed4-317e484a6458"
@@ -40,5 +38,6 @@ Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 SliceMap = "82cb661a-3f19-5665-9e27-df437c7e54c8"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
+Tidier = "f0413319-3358-4bb0-8e7c-0c83523a93bd"
 Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
