+++ src/CounterfactualExplanations.jl
@@ -33,7 +33,7 @@ export CounterfactualData,
 include("models/Models.jl")
 using .Models
 export AbstractFittedModel, AbstractDifferentiableModel
-export FluxModel, FluxEnsemble, LaplaceReduxModel
+export Linear, FluxModel, FluxEnsemble, LaplaceReduxModel
 export flux_training_params
 export probs, logits
 export model_catalogue, fit_model, model_evaluation, predict_label
@@ -50,7 +50,7 @@ export GravitationalGenerator, GravitationalGeneratorParams
 export GreedyGenerator, GreedyGeneratorParams
 export REVISEGenerator, REVISEGeneratorParams
 export DiCEGenerator, DiCEGeneratorParams
-export generator_catalog
+export generator_catalogue
 export generate_perturbations, conditions_satisified, mutability_constraints
 
 ### CounterfactualExplanation
