+++ docs/src/tutorials/benchmarking.md
@@ -1,31 +1,17 @@
----
-execute: 
-  output: true
----
 
-
-```@meta
+``` @meta
 CurrentModule = CounterfactualExplanations 
-```
-
-```{julia}
-#| echo: false
-#| output: false
-include("docs/setup_docs.jl")
-eval(setup_docs)
 ```
 
 # Performance Benchmarks
 
-In the previous tutorial, we have seen how counterfactual explanations can be evaluated. An important follow-up task is to compare the performance of different counterfactual generators is an important task. Researchers can use benchmarks to test new ideas they want to implement. Practitioners can find the right counterfactual generator for their specific use case through benchmarks. In this tutorial, we will see how to run benchmarks for counterfactual generators. 
+In the previous tutorial, we have seen how counterfactual explanations can be evaluated. An important follow-up task is to compare the performance of different counterfactual generators is an important task. Researchers can use benchmarks to test new ideas they want to implement. Practitioners can find the right counterfactual generator for their specific use case through benchmarks. In this tutorial, we will see how to run benchmarks for counterfactual generators.
 
 ## Post Hoc Benchmarking
 
 We begin by continuing the discussion from the previous tutorial: suppose you have generated multiple counterfactual explanations for multiple individuals, like below:
 
-```{julia}
-#| output: false
-
+``` julia
 # Factual and target:
 n_individuals = 5
 ids = rand(findall(predict_label(M, counterfactual_data) .== factual), n_individuals)
@@ -35,8 +21,7 @@ ces = generate_counterfactual(xs, target, counterfactual_data, M, generator; num
 
 You may be interested in comparing the outcomes across individuals. To benchmark the various counterfactual explanations using default evaluation measures, you can simply proceed as follows:
 
-```{julia}
-#| output: false
+``` julia
 bmk = benchmark(ces)
 ```
 
@@ -46,19 +31,19 @@ Under the hood, the [`benchmark(counterfactual_explanations::Vector{Counterfactu
 
 For convenience, the `DataFrame` containing the evaluation can be returned by simply calling the `Benchmark` object. By default, the aggregated evaluation measures across `id` (in line with the default behaviour of `evaluate`).
 
-```{julia}
+``` julia
 bmk()
 ```
 
 To retrieve the granular dataset, simply do:
 
-```{julia}
+``` julia
 bmk(agg=nothing)
 ```
 
 Since benchmarks return a `DataFrame` object on call, post-processing is straightforward. For example, we could use [`Tidier.jl`](https://kdpsingh.github.io/Tidier.jl/dev/):
 
-```{julia}
+``` julia
 using Tidier
 @chain bmk() begin
     @filter(variable == "distance")
@@ -70,7 +55,7 @@ end
 
 Benchmarks always report metadata for each counterfactual explanation, which is automatically inferred by default. The default metadata concerns the explained `model` and the employed `generator`. In the current example, we used the same model and generator for each individual:
 
-```{julia}
+``` julia
 @chain bmk() begin
     @group_by(sample)
     @select(sample, model, generator)
@@ -81,7 +66,7 @@ end
 
 Metadata can also be provided as an optional key argument.
 
-```{julia}
+``` julia
 meta_data = Dict(
     :generator => "Generic",
     :model => "MLP",
@@ -100,9 +85,9 @@ end
 
 So far we have assumed the following workflow:
 
-1. Fit some machine learning model.
-2. Generate counterfactual explanations for some individual(s) (`generate_counterfactual`).
-3. Evaluate and benchmark them (`benchmark(ces::Vector{CounterfactualExplanation})`).
+1.  Fit some machine learning model.
+2.  Generate counterfactual explanations for some individual(s) (`generate_counterfactual`).
[…]
