+++ docs/src/tutorials/benchmarking.md
@@ -35,11 +35,60 @@ For convenience, the `DataFrame` containing the evaluation can be returned by si
 bmk()
 ```
 
+    15×7 DataFrame
+     Row │ sample  variable    value     generator                          model  ⋯
+         │ Int64   String      Float64   Symbol                             Symbol ⋯
+    ─────┼──────────────────────────────────────────────────────────────────────────
+       1 │      1  distance    0.752493  GenericGenerator(nothing, Linear…  FluxMo ⋯
+       2 │      1  redundancy  0.0       GenericGenerator(nothing, Linear…  FluxMo
+       3 │      1  validity    1.0       GenericGenerator(nothing, Linear…  FluxMo
+       4 │      2  distance    0.733512  GenericGenerator(nothing, Linear…  FluxMo
+       5 │      2  redundancy  0.0       GenericGenerator(nothing, Linear…  FluxMo ⋯
+       6 │      2  validity    1.0       GenericGenerator(nothing, Linear…  FluxMo
+       7 │      3  distance    0.727664  GenericGenerator(nothing, Linear…  FluxMo
+       8 │      3  redundancy  0.0       GenericGenerator(nothing, Linear…  FluxMo
+       9 │      3  validity    1.0       GenericGenerator(nothing, Linear…  FluxMo ⋯
+      10 │      4  distance    0.598759  GenericGenerator(nothing, Linear…  FluxMo
+      11 │      4  redundancy  0.0       GenericGenerator(nothing, Linear…  FluxMo
+      12 │      4  validity    1.0       GenericGenerator(nothing, Linear…  FluxMo
+      13 │      5  distance    0.725115  GenericGenerator(nothing, Linear…  FluxMo ⋯
+      14 │      5  redundancy  0.0       GenericGenerator(nothing, Linear…  FluxMo
+      15 │      5  validity    1.0       GenericGenerator(nothing, Linear…  FluxMo
+                                                                   3 columns omitted
+
 To retrieve the granular dataset, simply do:
 
 ``` julia
 bmk(agg=nothing)
 ```
+
+    75×8 DataFrame
+     Row │ sample  num_counterfactual  variable    value     generator             ⋯
+         │ Int64   Int64               String      Float64   Symbol                ⋯
+    ─────┼──────────────────────────────────────────────────────────────────────────
+       1 │      1                   1  distance    0.74913   GenericGenerator(noth ⋯
+       2 │      1                   2  distance    0.758537  GenericGenerator(noth
+       3 │      1                   3  distance    0.760605  GenericGenerator(noth
+       4 │      1                   4  distance    0.76034   GenericGenerator(noth
+       5 │      1                   5  distance    0.733853  GenericGenerator(noth ⋯
+       6 │      1                   1  redundancy  0.0       GenericGenerator(noth
+       7 │      1                   2  redundancy  0.0       GenericGenerator(noth
+       8 │      1                   3  redundancy  0.0       GenericGenerator(noth
+       9 │      1                   4  redundancy  0.0       GenericGenerator(noth ⋯
+      10 │      1                   5  redundancy  0.0       GenericGenerator(noth
+      11 │      1                   1  validity    1.0       GenericGenerator(noth
+      ⋮  │   ⋮             ⋮               ⋮          ⋮                      ⋮     ⋱
+      66 │      5                   1  redundancy  0.0       GenericGenerator(noth
+      67 │      5                   2  redundancy  0.0       GenericGenerator(noth ⋯
+      68 │      5                   3  redundancy  0.0       GenericGenerator(noth
+      69 │      5                   4  redundancy  0.0       GenericGenerator(noth
[…]
