+++ docs/src/tutorials/evaluation.qmd
@@ -127,7 +127,13 @@ Markdown.parse(
 ids = rand(findall(predict_label(M, counterfactual_data) .== factual), n_individuals)
 xs = select_factual(counterfactual_data, ids)
 ces = generate_counterfactual(xs, target, counterfactual_data, M, generator; num_counterfactuals=5)
-evaluate(ces)
+evaluation = evaluate(ces)
+```
+
+```{julia}
+#| echo: false
+
+show(evaluation)
 ```
 
 This leads us to our next topic: Performance Benchmarks.
