+++ docs/src/reference.md
@@ -25,7 +25,7 @@ Depth = 3
 ``` @autodocs
 Modules = [
     CounterfactualExplanations, 
-    CounterfactualExplanations.Benchmark,
+    CounterfactualExplanations.Evaluation,
     CounterfactualExplanations.Data,
     CounterfactualExplanations.DataPreprocessing,
     CounterfactualExplanations.Models, 
@@ -40,7 +40,7 @@ Private = false
 ``` @autodocs
 Modules = [
     CounterfactualExplanations, 
-    CounterfactualExplanations.Benchmark,
+    CounterfactualExplanations.Evaluation,
     CounterfactualExplanations.Data,
     CounterfactualExplanations.DataPreprocessing,
     CounterfactualExplanations.Models, 

[…]
