+++ docs/src/index.md
@@ -66,14 +66,14 @@ The counterfactual was produced using REVISE (Joshi et al. 2019):
 # Define generator:
 generator = REVISEGenerator(
   opt = Descent(0.1),
-  decision_threshold = 0.95,
-  λ = 0.01
+  decision_threshold = 0.99,
+  λ = 0.001
 )
 # Generate recourse:
 ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
 ```
 
-![](index_files/figure-commonmark/cell-10-output-1.svg)
+![](index_files/figure-commonmark/cell-11-output-1.svg)
 
 ## 🔍 Usage example
 
