+++ dev/tutorials/evaluation.qmd
@@ -1,0 +1,139 @@
+---
+execute: 
+  output: true
+---
+
+```@meta
+CurrentModule = CounterfactualExplanations 
+```
+
+```{julia}
+#| echo: false
+#| output: false
+include("docs/setup_docs.jl")
+eval(setup_docs)
+```
+
+# Performance Evaluation
+
+Now that we know how to generate counterfactual explanations in Julia, you may have a few follow-up questions: How do I know if the counterfactual search has been successful? How good is my counterfactual explanation? What does 'good' even mean in this context? In this tutorial, we will see how counterfactual explanations can be evaluated with respect to their performance.
+
+## Default Measures
+
+Numerous evaluation measures for counterfactual explanations have been proposed. In what follows, we will cover some of the most important measures.
+
+### Single Measure, Single Counterfactual
+
+One of the most important measures is [`validity`](@ref), which simply determines whether or not a counterfactual explanation $x^{\prime}$ is valid in the sense that it yields the target prediction: $M(x^{\prime})=t$. We can evaluate the validity of a single counterfactual explanation `ce` using the [`Evaluation.evaluate`](@ref) function as follows:
+
+```{julia}
+using CounterfactualExplanations.Evaluation: evaluate, validity
+evaluate(ce; measure=validity)
+```
+
+For a single counterfactual explanation, this evaluation measure can only take two values: it is either equal to `1`, if the explanation is valid or `0` otherwise. Another important measure is [`distance`](@ref), which relates to the distance between the factual $x$ and the counterfactual $x^{\prime}$. In the context of Algorithmic Recourse, higher distances are typically associated with higher costs to individuals seeking recourse. 
+
+```{julia}
+using CounterfactualExplanations.Evaluation: distance
+evaluate(ce; measure=distance)
+```
+
+By default, `distance` computes the L2 (Euclidean) distance. 
+
+### Multiple Measures, Single Counterfactual
+
+You might be interested in computing not just the L2 distance, but various LP norms. This can be done by supplying a vector of functions to the `measure` key argument. For convenience, all default distance measures have already been collected in a vector:
+
+```{julia}
+using CounterfactualExplanations.Evaluation: distance_measures
+distance_measures
+```
+
+We can use this vector of evaluation measures as follows:
+
+```{julia}
+evaluate(ce; measure=distance_measures)
+```
+
+If no `measure` is specified, the `evaluate` method will return all default measures,
+
+```{julia}
+evaluate(ce)
+```
+
+which include:
+
+```{julia}
+CounterfactualExplanations.Evaluation.default_measures
+```
+
+### Multiple Measures and Counterfactuals
+
+We can also evaluate multiple counterfactual explanations at once:
+
+```{julia}
+generator = DiCEGenerator()
+ces = generate_counterfactual(x, target, counterfactual_data, M, generator; num_counterfactuals=5)
+evaluate(ces)
+```
+
+By default, each evaluation measure is aggregated across all counterfactual explanations. To return individual measures for each counterfactual explanation you can specify `report_each=true`
+
+```{julia}
+evaluate(ces; report_each=true)
+```
+
+## Custom Measures
+
+A `measure` is just a method that takes a `CounterfactualExplanation` as its only positional argument. Defining custom measures is therefore straightforward. For example, we could define a measure to compute the inverse target probability as follows:
+
+```{julia}
+my_measure(ce::CounterfactualExplanation) = 1 .- CounterfactualExplanations.target_probs(ce)
+evaluate(ce; measure=my_measure)
+```
+
+## Tidy Output
+
[…]
