+++ dev/contribute.qmd
@@ -10,14 +10,14 @@ Our goal is to provide a go-to place for Counterfactual Explanations in Julia. T
 2. Additional datasets for testing, evaluation and benchmarking.
 3. Support for regression models.
 
-For a complete list, have a look at outstanding [issue](https://github.com/pat-alt/CounterfactualExplanations.jl/issues).
+For a complete list, have a look at outstanding [issue](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues).
 
 ## How to contribute?
 
 Any sort of contribution is welcome, in particular:
 
-1. Should you spot any errors or something is not working, please just open an [issue](https://github.com/pat-alt/CounterfactualExplanations.jl/issues).
+1. Should you spot any errors or something is not working, please just open an [issue](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues).
 2. If you want to contribute your code, please proceed as follows:
    - Fork this repo and clone your fork: `git clone https://github.com/your_username/CounterfactualExplanations.jl`.
    - Implement your modifications and submit a pull request.
-3. For any other questions or comments, you can also start a [discussion](https://github.com/pat-alt/CounterfactualExplanations.jl/discussions).
+3. For any other questions or comments, you can also start a [discussion](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/discussions).
