+++ versions.js
@@ -3,5 +3,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.1.6";
+var DOCUMENTER_NEWEST = "v0.1.8";
 var DOCUMENTER_STABLE = "stable";
