### 2023-03-11 f4310dd
some work on penalties. I think penalties can be reused as evaluation metrics.
