+++ src/generators/Generators.jl
@@ -9,6 +9,7 @@ using Flux
 using LinearAlgebra
 using ..Models
 using ..Objectives
+using Statistics
 
 export AbstractGradientBasedGenerator
 export ClaPROARGenerator, ClaPROARGeneratorParams
@@ -47,13 +48,11 @@ The default method to apply the generator complexity penalty to the current coun
 """
 function h(
     generator::AbstractGenerator,
-    counterfactual_explanation::AbstractCounterfactualExplanation,
+    counterfactual_explanation::AbstractCounterfactualExplanation;
+    agg=mean
 )
-    dist_ = generator.complexity(
-        counterfactual_explanation.x .-
-        CounterfactualExplanations.decode_state(counterfactual_explanation),
-    )
-    penalty = generator.λ * dist_
+    cost = agg(generator.complexity(counterfactual_explanation))
+    penalty = generator.λ * cost
     return penalty
 end
 
