+++ src/generators/gradient_based/functions.jl
@@ -51,8 +51,6 @@ function ∇(
     M::Models.AbstractDifferentiableModel,
     counterfactual_explanation::AbstractCounterfactualExplanation,
 )
-    println("loss: $(ℓ(generator, counterfactual_explanation))")
-    # println(∂ℓ(generator, M, counterfactual_explanation))
     ∂ℓ(generator, M, counterfactual_explanation) + ∂h(generator, counterfactual_explanation)
 end
 
