+++ src/objectives/Objectives.jl
@@ -2,9 +2,29 @@ module Objectives
 
 using ..CounterfactualExplanations
 
+# Loss functions:
 include("loss_functions.jl")
 export logitbinarycrossentropy, logitcrossentropy, mse
 
+# Catalogue:
+const losses_catalogue = Dict(
+    :logitbinarycrossentropy => logitbinarycrossentropy,
+    :logitcrossentropy => logitcrossentropy,
+    :mse => mse,
+)
+export losses_catalogue
+
+# Penalities
 include("penalties.jl")
+export distance, distance_l0, distance_l1, distance_l2, distance_linf
+
+# Catalogue:
+const penalties_catalogue = Dict(
+    :distance_l0 => distance_l0,
+    :distance_l1 => distance_l1,
+    :distance_l2 => distance_l2,
+    :distance_linf => distance_linf,
+)
+export losses_catalogue
 
 end
