+++ src/evaluation/measures.jl
@@ -1,14 +1,14 @@
 using LinearAlgebra
 using SliceMap
-using Statistics
+using Statistics: mean
 
 """
     validity(counterfactual_explanation::CounterfactualExplanation; γ=0.5)
 
 Checks of the counterfactual search has been successful with respect to the probability treshold `γ`. In case multiple counterfactuals were generated, the function returns the proportion of successful counterfactuals.
 """
-function validity(counterfactual_explanation::CounterfactualExplanation; γ=0.5)
-    CounterfactualExplanations.target_probs(counterfactual_explanation) .>= γ
+function validity(counterfactual_explanation::CounterfactualExplanation; agg=mean, γ=0.5)
+    agg(CounterfactualExplanations.target_probs(counterfactual_explanation) .>= γ)
 end
 
 """
@@ -23,9 +23,9 @@ validity_strict(counterfactual_explanation::CounterfactualExplanation) = validit
 
 Computes the feature redundancy: that is, the number of features that remain unchanged from their original, factual values.
 """
-function redundancy(counterfactual_explanation::CounterfactualExplanation)
+function redundancy(counterfactual_explanation::CounterfactualExplanation; agg=mean)
     x′ = CounterfactualExplanations.counterfactual(counterfactual_explanation)
-    redundant_x = mapslices(x -> sum(x .== 0) / size(x, 1), x′, dims=[1, 2])
+    redundant_x = agg(mapslices(x -> sum(x .== 0) / size(x, 1), x′, dims=[1, 2]))
     return redundant_x
 end
 
