+++ src/generators/gradient_based/REVISEGenerator.jl
@@ -35,7 +35,7 @@ generator = REVISEGenerator()
 """
 function REVISEGenerator(;
     loss::Union{Nothing,Function}=nothing,
-    complexity::Function=LinearAlgebra.norm,
+    complexity::Function=Objectives.distance_l2,
     λ::AbstractFloat=0.1,
     decision_threshold=0.5,
     kwargs...
