+++ src/objectives/Objectives.jl
@@ -17,6 +17,7 @@ export losses_catalogue
 # Penalities
 include("penalties.jl")
 export distance, distance_l0, distance_l1, distance_l2, distance_linf
+export ddp_diversity
 
 # Catalogue:
 const penalties_catalogue = Dict(
@@ -24,7 +25,8 @@ const penalties_catalogue = Dict(
     :distance_l1 => distance_l1,
     :distance_l2 => distance_l2,
     :distance_linf => distance_linf,
+    :ddp_diversity => ddp_diversity,
 )
-export losses_catalogue
+export penalties_catalogue
 
 end
