+++ docs/src/tutorials/generators.md
@@ -1,38 +1,32 @@
-```@meta
-CurrentModule = CounterfactualExplanations 
-```
 
-```{julia}
-#| echo: false
-include("docs/setup_docs.jl")
-eval(setup_docs)
+``` @meta
+CurrentModule = CounterfactualExplanations 
 ```
 
 # Handling Generators
 
 Generating Counterfactual Explanations can be seen as a generative modelling task because it involves generating samples in the input space: $x \sim \mathcal{X}$. In this tutorial, we will introduce how Counterfactual `Generator`s are used. They are discussed in more detail in the explanatory section of the documentation: [🤓 Explanation](@ref "🤓 Explanation").
 
-## Off-the-Shelf Generators 
+## Off-the-Shelf Generators
 
-Currently, the following off-the-shelf counterfactual generators are implemented in the package. 
+Currently, the following off-the-shelf counterfactual generators are implemented in the package.
 
-```{julia}
+``` julia
 generator_catalogue
 ```
 
 These `Generator`s are just composite types that contain information about how counterfactuals ought to be generated. To specify the type of generator you want to use, you can simply instantiate it:
 
-```{julia}
+``` julia
 # Search:
 generator = GenericGenerator()
 ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
 plot(ce)
 ```
 
-## Composable Generators 
+## Composable Generators
 
-!!! warning "Breaking Changes Expected"
-    Work on this feature is still in its very early stages and breaking changes should be expected. 
+!!! warning “Breaking Changes Expected” Work on this feature is still in its very early stages and breaking changes should be expected.
 
 One of the key objectives for this package is **Composability**. It turns out that many of the various counterfactual generators that have been proposed in the literature, essentially do the same thing: they optimize an objective function. Formally we have,
 
@@ -40,30 +34,31 @@ $$
 \begin{aligned}
 \mathbf{s}^\prime &= \arg \min_{\mathbf{s}^\prime \in \mathcal{S}} \left\{  {\text{yloss}(M(f(\mathbf{s}^\prime)),y^*)}+ \lambda {\text{cost}(f(\mathbf{s}^\prime)) }  \right\} 
 \end{aligned} 
-$$ {#eq-general}
+ \qquad(1)$$
 
-where $\text{yloss}$ denotes the main loss function and $\text{cost}$ is a penalty term [@altmeyer2023endogenous]. 
+where $\text{yloss}$ denotes the main loss function and $\text{cost}$ is a penalty term (Altmeyer et al. 2023).
 
-Without going into further detail here, the important thing to mention is that @eq-general very closely describes how counterfactual search is actually implemented in the package. In other words, all off-the-shelf generators currently implemented work with that same objective. They just vary in the way that penalties are defined, for example. This gives rise to an interesting idea: 
+Without going into further detail here, the important thing to mention is that **?@eq-eq-general** very closely describes how counterfactual search is actually implemented in the package. In other words, all off-the-shelf generators currently implemented work with that same objective. They just vary in the way that penalties are defined, for example. This gives rise to an interesting idea:
 
 > Why not compose generators that combine ideas from different off-the-shelf generators?
 
 The [`ComposableGenerator`](@ref) class provides a straightforward way to do this, without requiring users to build custom `Generator`s from scratch. It can be instantiated as follows:
 
-```{julia}
+``` julia
 generator = ComposableGenerator()
 ```
 
[…]
