+++ docs/src/tutorials/generators.md
@@ -15,6 +15,14 @@ Currently, the following off-the-shelf counterfactual generators are implemented
 generator_catalogue
 ```
 
+    Dict{Symbol, DataType} with 6 entries:
+      :gravitational => GravitationalGenerator
+      :revise        => REVISEGenerator
+      :dice          => DiCEGenerator
+      :generic       => GenericGenerator
+      :greedy        => GreedyGenerator
+      :claproar      => ClaPROARGenerator
+
 These `Generator`s are just composite types that contain information about how counterfactuals ought to be generated. To specify the type of generator you want to use, you can simply instantiate it:
 
 ``` julia
@@ -24,9 +32,12 @@ ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
 plot(ce)
 ```
 
+![](generators_files/figure-commonmark/cell-4-output-1.svg)
+
 ## Composable Generators
 
-!!! warning “Breaking Changes Expected” Work on this feature is still in its very early stages and breaking changes should be expected.
+!!! warning “Breaking Changes Expected”  
+    Work on this feature is still in its very early stages and breaking changes should be expected.
 
 One of the key objectives for this package is **Composability**. It turns out that many of the various counterfactual generators that have been proposed in the literature, essentially do the same thing: they optimize an objective function. Formally we have,
 
@@ -38,7 +49,7 @@ $$
 
 where $\text{yloss}$ denotes the main loss function and $\text{cost}$ is a penalty term (Altmeyer et al. 2023).
 
-Without going into further detail here, the important thing to mention is that **?@eq-eq-general** very closely describes how counterfactual search is actually implemented in the package. In other words, all off-the-shelf generators currently implemented work with that same objective. They just vary in the way that penalties are defined, for example. This gives rise to an interesting idea:
+Without going into further detail here, the important thing to mention is that [Equation 1](#eq-general) very closely describes how counterfactual search is actually implemented in the package. In other words, all off-the-shelf generators currently implemented work with that same objective. They just vary in the way that penalties are defined, for example. This gives rise to an interesting idea:
 
 > Why not compose generators that combine ideas from different off-the-shelf generators?
 
@@ -48,10 +59,10 @@ The [`ComposableGenerator`](@ref) class provides a straightforward way to do thi
 generator = ComposableGenerator()
 ```
 
-By default, this creates a `generator` that simply performs gradient descent without any penalties. To modify the behaviour of the `generator`, you can define the counterfactual search objective function using the [`objective`](@ref) macro:
+By default, this creates a `generator` that simply performs gradient descent without any penalties. To modify the behaviour of the `generator`, you can define the counterfactual search objective function using the [`@objective`](@ref) macro:
 
 ``` julia
-@objective(generator, logitbinarycrossentropy + 0.1distance_l2 + 10ddp_diversity)
+@objective(generator, logitbinarycrossentropy + 0.1distance_l2 + 1.0ddp_diversity)
 ```
 
 Here we have essentially created a version of the [`DiCEGenerator`](@ref):
@@ -60,5 +71,9 @@ Here we have essentially created a version of the [`DiCEGenerator`](@ref):
 ce = generate_counterfactual(x, target, counterfactual_data, M, generator; num_counterfactuals=5)
 plot(ce)
 ```
+
+![](generators_files/figure-commonmark/cell-7-output-1.svg)
+
+## References
 
 Altmeyer, Patrick, Giovan Angela, Aleksander Buszydlik, Karol Dobiczek, Arie van Deursen, and Cynthia Liem. 2023. “Endogenous Macrodynamics in Algorithmic Recourse.” In *First IEEE Conference on Secure and Trustworthy Machine Learning*.
