+++ docs/src/tutorials/generators.qmd
@@ -10,7 +10,7 @@ eval(setup_docs)
 
 # Handling Generators
 
-Generating Counterfactual Explanations can be seen as a generative modelling task because it involves generating samples in the input space: $x \sim \mathcal{X}$. In this tutorial, we will introduce how Counterfactual `Generator`s are used. They are discussed in more detail in the explanatory section of the documentation: [🤓 Explanation](@ref "🤓 Explanation").
+Generating Counterfactual Explanations can be seen as a generative modelling task because it involves generating samples in the input space: $x \sim \mathcal{X}$. In this tutorial, we will introduce how Counterfactual `Generator`s are used. They are discussed in more detail in the explanatory section of the documentation.
 
 ## Off-the-Shelf Generators 
 
