+++ docs/src/explanation/generators/overview.qmd
@@ -8,7 +8,7 @@ include("docs/setup_docs.jl")
 eval(setup_docs)
 ```
 
-# Counterfactual Generators
+# Counterfactual Generators [@id generators_explanation]
 
 Counterfactual generators form the very core of this package. The [`generator_catalogue`](@ref) can be used to inspect the available generators:
 
