+++ src/CounterfactualExplanations.jl
@@ -58,7 +58,7 @@ export REVISEGenerator, REVISEGeneratorParams
 export DiCEGenerator, DiCEGeneratorParams
 export generator_catalogue
 export generate_perturbations, conditions_satisified, mutability_constraints
-export ComposableGenerator, @objective, @threshold
+export Generator, @objective, @threshold
 
 macro sayhello(name)
     return :( println("Hello, $name") )
