### 2023-03-14 18a85db
not 100percent sure that moving to contructor-only architecture really makes sense
