+++ test/generators.jl
@@ -7,13 +7,11 @@ using Random
 
     @testset "Generic" begin
         generator = GenericGenerator()
-        @test hasfield(GenericGenerator, :loss)
         @test typeof(generator) <: AbstractGradientBasedGenerator
     end
 
     @testset "Greedy" begin
         generator = GreedyGenerator()
-        @test hasfield(GreedyGenerator, :loss)
         @test typeof(generator) <: AbstractGradientBasedGenerator
     end
 
