+++ src/generators/Generators.jl
@@ -12,12 +12,12 @@ using ..Objectives
 using Statistics
 
 export AbstractGradientBasedGenerator
-export ClaPROARGenerator, ClaPROARGeneratorParams
-export GenericGenerator, GenericGeneratorParams
-export GravitationalGenerator, GravitationalGeneratorParams
+export ClaPROARGenerator
+export GenericGenerator
+export GravitationalGenerator
 export GreedyGenerator, GreedyGeneratorParams
-export REVISEGenerator, REVISEGeneratorParams
-export DiCEGenerator, DiCEGeneratorParams
+export REVISEGenerator
+export DiCEGenerator
 export generator_catalogue
 export generate_perturbations, conditions_satisified, mutability_constraints
 export Generator, @objective, @threshold
