+++ src/generators/gradient_based/functions.jl
@@ -122,13 +122,3 @@ end
 
 # Baseline
 include("GreedyGenerator.jl")           # Schut et al. (2021)
-
-# Latent space
-"""
-    AbstractLatentSpaceGenerator
-
-An abstract type that serves as the base type for gradient-based counterfactual generators that search in a latent space. 
-"""
-abstract type AbstractLatentSpaceGenerator <: AbstractGradientBasedGenerator end
-
-include("REVISEGenerator.jl") # Joshi et al. (2019)
