### 2023-03-15 1480a4b
work on streamlining step rules
