+++ src/generators/gradient_based/GreedyGenerator.jl
@@ -1,145 +1,0 @@
-using Parameters
-using LinearAlgebra
-using SliceMap
-
-"Class for Greedy counterfactual generator following Schut et al (2020)."
-mutable struct GreedyGenerator <: AbstractGradientBasedGenerator
-    loss::Union{Nothing,Function} # loss function
-    complexity::Function # complexity function
-    λ::AbstractFloat # strength of penalty
-    decision_threshold::Union{Nothing,AbstractFloat} # probability threshold
-    latent_space::Bool 
-    ϵ::AbstractFloat # learning rate
-    τ::AbstractFloat # tolerance for convergence
-    n::Int # maximum number of times any feature can be changed
-    passes::Int # number of full passes (`n` times) through all features
-end
-
-# API streamlining:
-@with_kw struct GreedyGeneratorParams
-    ϵ::Union{AbstractFloat,Nothing} = nothing
-    τ::AbstractFloat = CounterfactualExplanations.parameters[:τ]
-    n::Union{Int,Nothing} = nothing
-end
-
-"""
-    GreedyGenerator(;
-        loss::Union{Nothing,Function} = nothing,
-        complexity::Function = LinearAlgebra.norm,
-        λ::AbstractFloat = 0.0,
-        decision_threshold = 0.5,
-        latent_space::Bool=false,
-        opt::Union{Nothing,Flux.Optimise.AbstractOptimiser} = nothing, # learning rate
-        kwargs...,
-    )
-
-An outer constructor method that instantiates a greedy generator.
-
-# Examples
-
-```julia-repl
-generator = GreedyGenerator()
-```
-"""
-function GreedyGenerator(;
-    loss::Union{Nothing,Function}=nothing,
-    complexity::Function=norm,
-    λ::AbstractFloat=0.0,
-    decision_threshold=0.5,
-    latent_space::Bool=false,
-    opt::Union{Nothing,Flux.Optimise.AbstractOptimiser}=nothing, # learning rate
-    kwargs...
-)
-
-    if !isnothing(opt)
-        @warn "The `GreedyGenerator` does not not work with a `Flux` optimiser. Argument `opt` will be ignored."
-        opt = nothing
-    end
-
-    # Load hyperparameters:
-    params = GreedyGeneratorParams(; kwargs...)
-    ϵ = params.ϵ
-    n = params.n
-    if all(isnothing.([ϵ, n]))
-        ϵ = 0.1
-        n = 10
-    elseif isnothing(ϵ) && !isnothing(n)
-        ϵ = 1 / n
-    elseif !isnothing(ϵ) && isnothing(n)
-        n = 1 / ϵ
-    end
-
-    # Sanity checks:
-    if λ != 0.0
-        @warn "Choosing λ different from 0 has no effect on `GreedyGenerator`, since no penalty term is involved."
-    end
-    if complexity != LinearAlgebra.norm
-        @warn "Specifying `complexity` has no effect on `GreedyGenerator`, since no penalty term is involved."
-    end
-
-    generator = GreedyGenerator(loss, complexity, λ, decision_threshold, latent_space, ϵ, params.τ, n, 0)
-
-    return generator
-end
-
-"""
-    ∇(generator::GreedyGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)    
-
-he default method to compute the gradient of the counterfactual search objective for a greedy generator. Since no complexity penalty is needed, this gradients just correponds to the partial derivative with respect to the loss function.
-
-"""
-∇(
-    generator::GreedyGenerator,
-    M::Models.AbstractDifferentiableModel,
-    counterfactual_explanation::AbstractCounterfactualExplanation,
[…]
