### 2023-03-16 4993f0f
docs updated and more macros added. it's possible to chain them.
