+++ docs/src/tutorials/generators.qmd
@@ -12,27 +12,6 @@ eval(setup_docs)
 
 Generating Counterfactual Explanations can be seen as a generative modelling task because it involves generating samples in the input space: $x \sim \mathcal{X}$. In this tutorial, we will introduce how Counterfactual `Generator`s are used. They are discussed in more detail in the explanatory section of the documentation.
 
-## Off-the-Shelf Generators 
-
-Currently, the following off-the-shelf counterfactual generators are implemented in the package. 
-
-```{julia}
-#| output: true
-
-generator_catalogue
-```
-
-These `Generator`s are just composite types that contain information about how counterfactuals ought to be generated. To specify the type of generator you want to use, you can simply instantiate it:
-
-```{julia}
-#| output: true
-
-# Search:
-generator = GenericGenerator()
-ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
-plot(ce)
-```
-
 ## Composable Generators 
 
 | !!! warning \"Breaking Changes Expected\"
@@ -73,5 +52,99 @@ ce = generate_counterfactual(x, target, counterfactual_data, M, generator; num_c
 plot(ce)
 ```
 
+Multiple macros can be chained using `Chains.jl` making it easy to create entirely new flavours of counterfactual generators. The following generator, for example, combines ideas from DiCE [@mothilal2020explaining], REVISE [@joshi2019realistic] and Greedy [@schut2021generating]:
+
+```{julia}
+@chain generator begin
+    @objective logitcrossentropy + 5.0ddp_diversity     # DiCE (Mothilal et al. 2020)
+    @threshold 0.95
+    @with_optimiser JSMADescent(η=0.5)                  # Greedy (Schut et al. 2021)
+    @search_latent_space                                # REVISE (Joshi et al. 2019)
+end
+```
+
+```{julia}
+#| echo: false
+using CounterfactualExplanations.Models: load_mnist_ensemble, load_mnist_vae
+counterfactual_data = load_mnist()
+X, y = unpack_data(counterfactual_data)
+input_dim, n_obs = size(counterfactual_data.X)
+M = load_mnist_ensemble()
+vae = load_mnist_vae()
+counterfactual_data.generative_model = vae # assign generative model
+```
+
+```{julia}
+#| echo: false
+
+Random.seed!(42)
+factual_label = 0
+x = reshape(X[:,rand(findall(predict_label(M, counterfactual_data).==factual_label))],input_dim,1)
+target = 8
+factual = predict_label(M, counterfactual_data, x)[1]
+ce = generate_counterfactual(x, target, counterfactual_data, M, generator; num_counterfactuals=5)
+```
+
+```{julia}
+#| echo: false
+#| output: true
+
+Markdown.parse(
+  """
+  Let's take this generator to our MNIST dataset and generate a counterfactual explanation for turning a $(factual) into a $(target). 
+  """
+)
+```
+
+```{julia}
+#| echo: false
+#| output: true
+
+# Factual:
+image_size = 300
+p1 = plot(
+    convert2image(MNIST, reshape(x,28,28)),
+    axis=nothing, 
+    size=(image_size, image_size),
+    title="Factual"
+)
+
+plts = []
+for i in 1:ce.num_counterfactuals
+    target_prob = round(target_probs(ce)[:,:,i][1],digits=2)
+    ŷ = CounterfactualExplanations.counterfactual_label(ce)[:,:,i]
+    plt = plot(
+        convert2image(MNIST, reshape(CounterfactualExplanations.counterfactual(ce)[:,:,i],28,28)),
+        axis=nothing, 
+        size=(image_size, image_size),
+        title="CE$i"
+    )
+    push!(plts, plt)
+end
+plot(p1,plts...;size=(image_size * 6,image_size), layout=(1,6))
+```
[…]
