+++ docs/src/explanation/generators/optimisers/jsma.qmd
@@ -1,32 +1,0 @@
-```@meta
-CurrentModule = CounterfactualExplanations 
-```
-
-```{julia}
-#| echo: false
-include("docs/setup_docs.jl")
-eval(setup_docs)
-```
-
-# Jacobian-based Saliency Map Attack
-
-@schut2021generating propose to use a ...
-
-```{julia}
-generator = GenericGenerator()
-generator.opt = Descent(0.01)
-ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
-plot(ce)
-```
-
-```{julia}
-using CounterfactualExplanations.Generators: JSMADescent
-opt = JSMADescent(;η=0.1)
-```
-
-```{julia}
-generator.opt = opt
-ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
-plot(ce)
-```
-
