+++ src/CounterfactualExplanations.jl
@@ -57,7 +57,7 @@ export GreedyGenerator
 export REVISEGenerator
 export DiCEGenerator
 export generator_catalogue
-export generate_perturbations, conditions_satisified, mutability_constraints
+export generate_perturbations, conditions_satisfied, mutability_constraints
 export Generator, @objective, @threshold
 
 macro sayhello(name)
