### 2023-03-16 7f5b945
minor things on convergence
