### 2023-03-18 9095a9e
removed unnecessary SliceMap
