### 2023-03-18 bf4fdc0
Clean README, docs and docstrings
I've just:
- Fixed a few typos in `docs`
- Fixed a few typos in `src`
- Fixed a typo in README
- Fixed the format of JuliaFormatter config file
