+++ src/generators/Generators.jl
@@ -33,7 +33,7 @@ include("gradient_based/functions.jl")
 include("gradient_based/generators.jl")
 include("gradient_based/optimisers.jl")
 
-"A dictionary countaining the contructors of all available counterfactual generators."
+"A dictionary containing the constructors of all available counterfactual generators."
 generator_catalogue = Dict(
     :claproar => Generators.ClaPROARGenerator,
     :generic => Generators.GenericGenerator,
