+++ docs/src/_intro.qmd
@@ -97,7 +97,7 @@ generator = GravitationalGenerator(
 )
 ```
 
-Consider the following real-world scenario: a retail bank is using a black-box model trained on their clients' credit history to decide whether they will provide credit to new applicants. To simulate this scenario, we have pre-trained a binary classifier on the publically available Give Me Some Credit dataset that ships with this package [@kaggle2011give].
+Consider the following real-world scenario: a retail bank is using a black-box model trained on their clients' credit history to decide whether they will provide credit to new applicants. To simulate this scenario, we have pre-trained a binary classifier on the publicly available Give Me Some Credit dataset that ships with this package [@kaggle2011give].
 
 ```{julia}
 #| echo: false
