### 2023-03-20 d8e9bb9
Format code and add checking workflow
