+++ src/data/tabular.jl
@@ -12,13 +12,13 @@ data_dir = joinpath(artifact"data-tabular", "data-tabular")
 
 Loads and pre-processes California Housing data.
 """
-function load_california_housing(n::Union{Nothing,Int} = 5000)
+function load_california_housing(n::Union{Nothing,Int}=5000)
 
     # Load:
     df = CSV.read(joinpath(data_dir, "cal_housing.csv"), DataFrame)
 
     # Pre-process features:
-    transformer = Standardizer(count = true)
+    transformer = Standardizer(; count=true)
     mach = MLJBase.fit!(machine(transformer, df[:, Not(:target)]))
     X = MLJBase.transform(mach, df[:, Not(:target)])
     X = Matrix(X)
@@ -31,8 +31,9 @@ function load_california_housing(n::Union{Nothing,Int} = 5000)
 
     # Undersample:
     if !isnothing(n)
-        counterfactual_data =
-            CounterfactualExplanations.DataPreprocessing.undersample(counterfactual_data, n)
+        counterfactual_data = CounterfactualExplanations.DataPreprocessing.undersample(
+            counterfactual_data, n
+        )
     end
 
     return counterfactual_data
@@ -43,13 +44,13 @@ end
 
 Loads and pre-processes Give Me Some Credit (GMSC) data.
 """
-function load_gmsc(n::Union{Nothing,Int} = 5000)
+function load_gmsc(n::Union{Nothing,Int}=5000)
 
     # Load:
     df = CSV.read(joinpath(data_dir, "gmsc.csv"), DataFrame)
 
     # Pre-process features:
-    transformer = Standardizer(count = true)
+    transformer = Standardizer(; count=true)
     mach = MLJBase.fit!(machine(transformer, df[:, Not(:target)]))
     X = MLJBase.transform(mach, df[:, Not(:target)])
     X = Matrix(X)
@@ -62,8 +63,9 @@ function load_gmsc(n::Union{Nothing,Int} = 5000)
 
     # Undersample:
     if !isnothing(n)
-        counterfactual_data =
-            CounterfactualExplanations.DataPreprocessing.undersample(counterfactual_data, n)
+        counterfactual_data = CounterfactualExplanations.DataPreprocessing.undersample(
+            counterfactual_data, n
+        )
     end
 
     return counterfactual_data
@@ -74,7 +76,7 @@ end
 
 Loads and pre-processes UCI Credit Default data.
 """
-function load_credit_default(n::Union{Nothing,Int} = 5000)
+function load_credit_default(n::Union{Nothing,Int}=5000)
 
     # Load:
     df = CSV.read(joinpath(data_dir, "credit_default.csv"), DataFrame)
@@ -83,7 +85,7 @@ function load_credit_default(n::Union{Nothing,Int} = 5000)
     df.SEX = categorical(df.SEX)
     df.EDUCATION = categorical(df.EDUCATION)
     df.MARRIAGE = categorical(df.MARRIAGE)
-    transformer = Standardizer(count = true) |> ContinuousEncoder()
+    transformer = ContinuousEncoder()(Standardizer(; count=true))
     mach = MLJBase.fit!(machine(transformer, df[:, Not(:target)]))
     X = MLJBase.transform(mach, df[:, Not(:target)])
     X = permutedims(Matrix(X))
@@ -95,14 +97,16 @@ function load_credit_default(n::Union{Nothing,Int} = 5000)
 
     # Counterfactual data:
     y = df.target
-    counterfactual_data =
-        CounterfactualData(X, y; features_categorical = features_categorical)
+    counterfactual_data = CounterfactualData(
+        X, y; features_categorical=features_categorical
+    )
     counterfactual_data.X = Float32.(counterfactual_data.X)
 
     # Undersample:
[…]
