+++ src/generators/gradient_based/base.jl
@@ -38,7 +38,3 @@ function Generator(;
 )
     return Generator(loss, penalty, λ, latent_space, opt)
 end
-
-
-
-
