+++ src/data/tabular.jl
@@ -85,7 +85,7 @@ function load_credit_default(n::Union{Nothing,Int}=5000)
     df.SEX = categorical(df.SEX)
     df.EDUCATION = categorical(df.EDUCATION)
     df.MARRIAGE = categorical(df.MARRIAGE)
-    transformer = ContinuousEncoder()(Standardizer(; count=true))
+    transformer = Standardizer(count = true) |> ContinuousEncoder()
     mach = MLJBase.fit!(machine(transformer, df[:, Not(:target)]))
     X = MLJBase.transform(mach, df[:, Not(:target)])
     X = permutedims(Matrix(X))
