+++ dev/utils.jl
@@ -19,14 +19,13 @@ nn = build_model()
 
 """
 function build_model(;
-    input_dim = 2,
-    n_hidden = 32,
-    output_dim = 1,
-    batch_norm = false,
-    dropout = false,
-    activation = Flux.relu,
+    input_dim=2,
+    n_hidden=32,
+    output_dim=1,
+    batch_norm=false,
+    dropout=false,
+    activation=Flux.relu,
 )
-
     if batch_norm
         nn = Chain(
             Dense(input_dim, n_hidden),
@@ -45,7 +44,6 @@ function build_model(;
     end
 
     return nn
-
 end
 
 """
@@ -53,11 +51,10 @@ end
 
 Wrapper function to train neural network and generate an animation showing the training loss evolution.
 """
-function forward_nn(nn, loss, data, opt; n_epochs = 200, plotting = nothing)
-
+function forward_nn(nn, loss, data, opt; n_epochs=200, plotting=nothing)
     avg_l = []
 
-    for epoch = 1:n_epochs
+    for epoch in 1:n_epochs
         for d in data
             gs = Flux.gradient(Flux.params(nn)) do
                 l = loss(d...)
@@ -71,12 +68,11 @@ function forward_nn(nn, loss, data, opt; n_epochs = 200, plotting = nothing)
             avg_loss(data) = mean(map(d -> loss(d[1], d[2]), data))
             avg_l = vcat(avg_l, avg_loss(data))
             if epoch % plotting[4] == 0
-                plot!(plt, avg_l, color = idx)
+                plot!(plt, avg_l; color=idx)
                 frame(anim, plt)
             end
         end
     end
-
 end
 
 """
@@ -92,8 +88,8 @@ using LaplaceRedux
 ```
 
 """
-function build_ensemble(K = 5; kw = (input_dim = 2, n_hidden = 32, output_dim = 1))
-    ensemble = [build_model(; kw...) for i = 1:K]
+function build_ensemble(K=5; kw=(input_dim=2, n_hidden=32, output_dim=1))
+    ensemble = [build_model(; kw...) for i in 1:K]
     return ensemble
 end
 
@@ -106,39 +102,33 @@ function forward(
     ensemble,
     data,
     opt;
-    loss_type = :logitbinarycrossentropy,
-    plot_loss = true,
-    n_epochs = 10,
-    plot_every = 1,
+    loss_type=:logitbinarycrossentropy,
+    plot_loss=true,
+    n_epochs=10,
+    plot_every=1,
 )
-
     anim = nothing
     if plot_loss
         anim = Animation()
-        plt = plot(
-            ylim = (0, 1),
-            xlim = (0, n_epochs),
-            legend = false,
-            xlab = "Epoch",
-            title = "Average (training) loss",
+        plt = plot(;
+            ylim=(0, 1),
+            xlim=(0, n_epochs),
+            legend=false,
+            xlab="Epoch",
+            title="Average (training) loss",
         )
-        for i = 1:length(ensemble)
+        for i in 1:length(ensemble)
             nn = ensemble[i]
             loss(x, y) = getfield(Flux.Losses, loss_type)(nn(x), y)
             forward_nn(
-                nn,
-                loss,
-                data,
-                opt,
-                n_epochs = n_epochs,
-                plotting = (plt, anim, i, plot_every),
+                nn, loss, data, opt; n_epochs=n_epochs, plotting=(plt, anim, i, plot_every)
             )
         end
     else
         plt = nothing
         for nn in ensemble
             loss(x, y) = getfield(Flux.Losses, loss_type)(nn(x), y)
[…]
