### 2023-03-21 3db7e09
Allow pipe syntax in the config file
