+++ src/generators/gradient_based/generators.jl
@@ -30,7 +30,8 @@ end
 
 "Constructor for `GreedyGenerator`."
 function GreedyGenerator(;
-    λ::AbstractFloat=0.1, opt=CounterfactualExplanations.Generators.JSMADescent(), kwargs...
-)
+    λ::AbstractFloat=0.1, η=0.1, n=nothing, kwargs...
+)   
+    opt = CounterfactualExplanations.Generators.JSMADescent(η=η,n=n)
     return Generator(; penalty=Objectives.distance_l2, λ=λ, opt=opt, kwargs...)
 end
