### 2023-04-05 f943e2f
fixes minor bug in undersample() function
