+++ src/data/vision.jl
@@ -12,7 +12,7 @@ function load_mnist(n::Union{Nothing,Int}=nothing)
     X, y = MNIST(:train)[:]
     X = Flux.flatten(X)
     y = categorical(y)
-    counterfactual_data = CounterfactualData(X, y; domain=(0, 1))
+    counterfactual_data = CounterfactualData(X, y; domain=(0, 1), standardize=false)
     counterfactual_data.X = Float32.(counterfactual_data.X)
     # Undersample:
     if !isnothing(n)
