### 2023-04-05 #103 @pat-alt
No worries!

Just to flag that I've developed the package quite a bit since Jan. Some major updates that may be worth mentioning in the paper include:

- [ ] Improved benchmarking ([0.1.8](https://github.com/JuliaTrustworthyAI/CounterfactualExplanations.jl/releases/tag/v0.1.8))
- [ ] Full composability of counterfactual generators ([0.1.9](https://github.com/JuliaTrustworthyAI/CounterfactualExplanations.jl/releases/tag/v0.1.9))

Unfortunately, I won't really have time to make further edits to the paper until mid-May, but thought I'd still mention it here so reviewers are aware.

Thanks!
