### 2023-04-05 135644f
increased perturbation for DICE to avoid singularity
