+++ dev/contribute/performance.qmd
@@ -1,0 +1,35 @@
+```{julia}
+#| echo: false
+include("docs/setup_docs.jl")
+eval(setup_docs)
+```
+
+
+```{julia}
+# Counteractual data and model:
+counterfactual_data = load_linearly_separable()
+M = fit_model(counterfactual_data, :Linear)
+target = 2
+factual = 1
+chosen = rand(findall(predict_label(M, counterfactual_data) .== factual))
+x = select_factual(counterfactual_data, chosen)
+
+# Search:
+generator = GenericGenerator()
+ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
+```
+
+
+```{julia}
+counterfactual_data_large = load_linearly_separable(100000)
+```
+
+
+```{julia}
+@time generate_counterfactual(x, target, counterfactual_data, M, generator)
+```
+
+
+```{julia}
+@time generate_counterfactual(x, target, counterfactual_data_large, M, generator)
+```
