+++ src/counterfactuals/plotting.jl
@@ -5,7 +5,7 @@ using SliceMap
 
 """
     Plots.plot(
-        counterfactual_explanation::CounterfactualExplanation;
+        ce::CounterfactualExplanation;
         alpha_ = 0.5,
         plot_up_to::Union{Nothing,Int} = nothing,
         plot_proba::Bool = false,
@@ -26,26 +26,26 @@ plot(ce)
 ```
 """
 function Plots.plot(
-    counterfactual_explanation::CounterfactualExplanation;
+    ce::CounterfactualExplanation;
     alpha_=0.5,
     plot_up_to::Union{Nothing,Int}=nothing,
     plot_proba::Bool=false,
     kwargs...,
 )
-    max_iter = total_steps(counterfactual_explanation)
+    max_iter = total_steps(ce)
     max_iter = if isnothing(plot_up_to)
-        total_steps(counterfactual_explanation)
+        total_steps(ce)
     else
         minimum([plot_up_to, max_iter])
     end
     max_iter += 1
     ingredients = set_up_plots(
-        counterfactual_explanation; alpha=alpha_, plot_proba=plot_proba, kwargs...
+        ce; alpha=alpha_, plot_proba=plot_proba, kwargs...
     )
 
     for t in 1:max_iter
         final_state = t == max_iter
-        plot_state(counterfactual_explanation, t, final_state; ingredients...)
+        plot_state(ce, t, final_state; ingredients...)
     end
 
     plt = if plot_proba
@@ -58,7 +58,7 @@ function Plots.plot(
 end
 
 """
-    animate_path(counterfactual_explanation::CounterfactualExplanation, path=tempdir(); plot_proba::Bool=false, kwargs...)
+    animate_path(ce::CounterfactualExplanation, path=tempdir(); plot_proba::Bool=false, kwargs...)
 
 Returns and animation of the counterfactual path.
 
@@ -73,27 +73,27 @@ animate_path(ce)
 ```
 """
 function animate_path(
-    counterfactual_explanation::CounterfactualExplanation,
+    ce::CounterfactualExplanation,
     path=tempdir();
     alpha_=0.5,
     plot_up_to::Union{Nothing,Int}=nothing,
     plot_proba::Bool=false,
     kwargs...,
 )
-    max_iter = total_steps(counterfactual_explanation)
+    max_iter = total_steps(ce)
     max_iter = if isnothing(plot_up_to)
-        total_steps(counterfactual_explanation)
+        total_steps(ce)
     else
         minimum([plot_up_to, max_iter])
     end
     max_iter += 1
     ingredients = set_up_plots(
-        counterfactual_explanation; alpha=alpha_, plot_proba=plot_proba, kwargs...
+        ce; alpha=alpha_, plot_proba=plot_proba, kwargs...
     )
 
     anim = @animate for t in 1:max_iter
         final_state = t == max_iter
-        plot_state(counterfactual_explanation, t, final_state; ingredients...)
+        plot_state(ce, t, final_state; ingredients...)
         if plot_proba
             plot(ingredients.p1, ingredients.p2; kwargs...)
         else
@@ -105,7 +105,7 @@ end
 
 """
     plot_state(
-        counterfactual_explanation::CounterfactualExplanation,
+        ce::CounterfactualExplanation,
         t::Int,
         final_sate::Bool;
         kwargs...
@@ -114,7 +114,7 @@ end
 Helper function that plots a single step of the counterfactual path.
 """
 function plot_state(
-    counterfactual_explanation::CounterfactualExplanation,
+    ce::CounterfactualExplanation,
     t::Int,
     final_sate::Bool;
     kwargs...,
@@ -124,7 +124,7 @@ function plot_state(
[…]
