+++ src/counterfactuals/functions.jl
@@ -86,13 +86,13 @@ function CounterfactualExplanation(;
     )
 
     # Instantiate: 
-    counterfactual_explanation = CounterfactualExplanation(
+    ce = CounterfactualExplanation(
         x,
         target,
         target_encoded,
         x,
         data,
-        deepcopy(M),
+        M,
         deepcopy(generator),
         generative_model_params,
         params,
@@ -103,48 +103,48 @@ function CounterfactualExplanation(;
     )
 
     # Initialization:
-    adjust_shape!(counterfactual_explanation)                                           # adjust shape to specified number of counterfactuals
-    counterfactual_explanation.s′ = encode_state(counterfactual_explanation)            # encode the counterfactual state
-    counterfactual_explanation.s′ = initialize_state(counterfactual_explanation)        # initialize the counterfactual state
+    adjust_shape!(ce)                                           # adjust shape to specified number of counterfactuals
+    ce.s′ = encode_state(ce)            # encode the counterfactual state
+    ce.s′ = initialize_state(ce)        # initialize the counterfactual state
 
     # Initialize search:
-    counterfactual_explanation.search = Dict(
+    ce.search = Dict(
         :iteration_count => 0,
-        :times_changed_features => zeros(size(decode_state(counterfactual_explanation))),
-        :path => [counterfactual_explanation.s′],
+        :times_changed_features => zeros(size(decode_state(ce))),
+        :path => [ce.s′],
         :terminated =>
-            threshold_reached(counterfactual_explanation, counterfactual_explanation.x),
-        :converged => converged(counterfactual_explanation),
+            threshold_reached(ce, ce.x),
+        :converged => converged(ce),
     )
 
     # Check for redundancy:
-    if terminated(counterfactual_explanation)
+    if terminated(ce)
         @info "Factual already in target class and probability exceeds threshold γ."
     end
 
-    return counterfactual_explanation
+    return ce
 end
 
 # 1.) Convenience methods:
 """
-    output_dim(counterfactual_explanation::CounterfactualExplanation)
+    output_dim(ce::CounterfactualExplanation)
 
 A convenience method that returns the output dimension of the predictive model.
 """
-function output_dim(counterfactual_explanation::CounterfactualExplanation)
-    return size(Models.probs(counterfactual_explanation.M, counterfactual_explanation.x))[1]
+function output_dim(ce::CounterfactualExplanation)
+    return size(Models.probs(ce.M, ce.x))[1]
 end
 
 """
-    guess_loss(counterfactual_explanation::CounterfactualExplanation)
+    guess_loss(ce::CounterfactualExplanation)
 
 Guesses the loss function to be used for the counterfactual search in case `likelihood` field is specified for the [`AbstractFittedModel`](@ref) instance and no loss function was explicitly declared for [`AbstractGenerator`](@ref) instance.
 """
-function guess_loss(counterfactual_explanation::CounterfactualExplanation)
-    if :likelihood in fieldnames(typeof(counterfactual_explanation.M))
-        if counterfactual_explanation.M.likelihood == :classification_binary
+function guess_loss(ce::CounterfactualExplanation)
+    if :likelihood in fieldnames(typeof(ce.M))
+        if ce.M.likelihood == :classification_binary
             loss_fun = Objectives.logitbinarycrossentropy
-        elseif counterfactual_explanation.M.likelihood == :classification_multi
+        elseif ce.M.likelihood == :classification_multi
[…]
