+++ src/generators/gradient_based/functions.jl
@@ -2,82 +2,82 @@ using Flux
 using Statistics
 
 """
-    ∂ℓ(generator::AbstractGradientBasedGenerator, M::Union{Models.LogisticModel, Models.BayesianLogisticModel}, counterfactual_explanation::AbstractCounterfactualExplanation)
+    ∂ℓ(generator::AbstractGradientBasedGenerator, M::Union{Models.LogisticModel, Models.BayesianLogisticModel}, ce::AbstractCounterfactualExplanation)
 
 The default method to compute the gradient of the loss function at the current counterfactual state for gradient-based generators. It assumes that `Zygote.jl` has gradient access.
 """
 function ∂ℓ(
     generator::AbstractGradientBasedGenerator,
     M::Models.AbstractDifferentiableModel,
-    counterfactual_explanation::AbstractCounterfactualExplanation,
+    ce::AbstractCounterfactualExplanation,
 )
     gs = gradient(
-        () -> ℓ(generator, counterfactual_explanation),
-        Flux.params(counterfactual_explanation.s′),
-    )[counterfactual_explanation.s′]
+        () -> ℓ(generator, ce),
+        Flux.params(ce.s′),
+    )[ce.s′]
     return gs
 end
 
 """
-    ∂h(generator::AbstractGradientBasedGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)
+    ∂h(generator::AbstractGradientBasedGenerator, ce::AbstractCounterfactualExplanation)
 
 The default method to compute the gradient of the complexity penalty at the current counterfactual state for gradient-based generators. It assumes that `Zygote.jl` has gradient access.
 """
 function ∂h(
     generator::AbstractGradientBasedGenerator,
-    counterfactual_explanation::AbstractCounterfactualExplanation,
+    ce::AbstractCounterfactualExplanation,
 )
     return gradient(
-        () -> h(generator, counterfactual_explanation),
-        Flux.params(counterfactual_explanation.s′),
-    )[counterfactual_explanation.s′]
+        () -> h(generator, ce),
+        Flux.params(ce.s′),
+    )[ce.s′]
 end
 
 # Gradient:
 """
-    ∇(generator::AbstractGradientBasedGenerator, M::Models.AbstractDifferentiableModel, counterfactual_explanation::AbstractCounterfactualExplanation)
+    ∇(generator::AbstractGradientBasedGenerator, M::Models.AbstractDifferentiableModel, ce::AbstractCounterfactualExplanation)
 
 The default method to compute the gradient of the counterfactual search objective for gradient-based generators. It simply computes the weighted sum over partial derivates. It assumes that `Zygote.jl` has gradient access.
 """
 function ∇(
     generator::AbstractGradientBasedGenerator,
     M::Models.AbstractDifferentiableModel,
-    counterfactual_explanation::AbstractCounterfactualExplanation,
+    ce::AbstractCounterfactualExplanation,
 )
-    return ∂ℓ(generator, M, counterfactual_explanation) +
-           ∂h(generator, counterfactual_explanation)
+    return ∂ℓ(generator, M, ce) +
+           ∂h(generator, ce)
 end
 
 """
-    propose_state(generator::AbstractGradientBasedGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)
+    propose_state(generator::AbstractGradientBasedGenerator, ce::AbstractCounterfactualExplanation)
 
 Proposes new state based on backpropagation.
 """
 function propose_state(
     generator::AbstractGradientBasedGenerator,
-    counterfactual_explanation::AbstractCounterfactualExplanation,
+    ce::AbstractCounterfactualExplanation,
 )
-    grads = ∇(generator, counterfactual_explanation.M, counterfactual_explanation) # gradient
[…]

[…]
