+++ src/models/differentiable/flux_mlp.jl
@@ -85,6 +85,7 @@ function forward!(
             n_epochs; desc="Progress on epochs:", showspeed=true, color=:green
         )
     end
+
     for epoch in 1:n_epochs
         for d in data
             gs = Flux.gradient(Flux.params(model)) do
@@ -116,32 +117,39 @@ function build_mlp(;
     n_layers::Int=2,
     output_dim::Int=1,
     dropout::Bool=false,
+    batch_norm::Bool=false,
     activation=Flux.relu,
     p_dropout=0.25,
 )
     @assert n_layers >= 1 "Need at least one layer."
 
     if n_layers == 1
-
-        # Logistic regression:
         model = Chain(Dense(input_dim, output_dim))
-
     elseif dropout
         hidden_ = repeat(
             [Dense(n_hidden, n_hidden, activation), Dropout(p_dropout)], n_layers - 2
         )
-
         model = Chain(
             Dense(input_dim, n_hidden, activation),
             Dropout(p_dropout),
             hidden_...,
             Dense(n_hidden, output_dim),
         )
+    elseif batch_norm
+        hidden_ = repeat([Dense(n_hidden,n_hidden),BatchNorm(n_hidden,activation)],n_layers-2)
+        model = Chain(
+            Dense(input_dim, n_hidden),
+            BatchNorm(n_hidden, activation),
+            hidden_...,
+            Dense(n_hidden, output_dim),
+            BatchNorm(output_dim)
+        )  
     else
         hidden_ = repeat([Dense(n_hidden, n_hidden, activation)], n_layers - 2)
-
         model = Chain(
-            Dense(input_dim, n_hidden, activation), hidden_..., Dense(n_hidden, output_dim)
+            Dense(input_dim, n_hidden, activation), 
+            hidden_..., 
+            Dense(n_hidden, output_dim)
         )
     end
 
