+++ src/counterfactuals/functions.jl
@@ -39,12 +39,12 @@ end
 
 Outer method to construct a `CounterfactualExplanation` structure.
 """
-function CounterfactualExplanation(;
+function CounterfactualExplanation(
     x::AbstractArray,
     target::RawTargetType,
     data::CounterfactualData,
     M::Models.AbstractFittedModel,
-    generator::Generators.AbstractGenerator,
+    generator::Generators.AbstractGenerator;
     num_counterfactuals::Int=1,
     initialization::Symbol=:add_perturbation,
     generative_model_params::NamedTuple=(;),
@@ -167,15 +167,7 @@ A convenience method that adjusts the dimensions of `x`.
 function adjust_shape(
     ce::CounterfactualExplanation, x::AbstractArray
 )
-    size_ =
-        Int.(
-            vcat(
-                ones(maximum([ndims(x), 2])), ce.num_counterfactuals
-            )
-        )
-    s′ = copy(x)
-    s′ = repeat(x; outer=size_)
-
+    s′ = repeat(x; outer=(1, ce.num_counterfactuals))
     return s′
 end
 
@@ -215,8 +207,7 @@ Applies all required encodings to `x`:
 Finally, it returns the encoded state variable.
 """
 function encode_state(
-    ce::CounterfactualExplanation,
-    x::Union{AbstractArray,Nothing}=nothing,
+    ce::CounterfactualExplanation, x::Union{AbstractArray,Nothing}=nothing
 )
 
     # Unpack:
@@ -226,20 +217,18 @@ function encode_state(
     # Latent space:
     if ce.params[:latent_space]
         s′ = map_to_latent(ce, s′)
-        return s′
     end
 
     # Standardize data unless latent space:
-    if !ce.params[:latent_space]
+    if !ce.params[:latent_space] && data.standardize
         dt = data.dt
         idx = transformable_features(data)
-        SliceMap.slicemap(s′; dims=(1, 2)) do s
-            _s = s[idx, :]
-            StatsBase.transform!(dt, _s)
-            s[idx, :] = _s
-        end
-        return s′
+        s = s′[idx, :]
+        StatsBase.transform!(dt, s)
+        s′[idx, :] = s
     end
+
+    return s′
 end
 
 """
@@ -322,16 +311,14 @@ function decode_state(
     end
 
     # Standardization:
-    if !ce.params[:latent_space]
+    if !ce.params[:latent_space] && data.standardize
         dt = data.dt
 
         # Continuous:
         idx = transformable_features(data)
-        SliceMap.slicemap(s′; dims=(1, 2)) do s
-            _s = s[idx, :]
-            StatsBase.reconstruct!(dt, _s)
-            s[idx, :] = _s
-        end
+        s = s′[idx, :]
+        StatsBase.reconstruct!(dt, s′[idx, :])
+        s′[idx, :] = s
     end
 
     # Categorical:
@@ -389,11 +376,7 @@ function reconstruct_cat_encoding(
     s′ = isnothing(x) ? deepcopy(ce.s′) : x
     data = ce.data
 
-    s′ = SliceMap.slicemap(s′; dims=(1, 2)) do s
-        s_encoded = DataPreprocessing.reconstruct_cat_encoding(data, s)
-        s = reshape(s_encoded, size(s)...)
-        return s
-    end
+    s′ = DataPreprocessing.reconstruct_cat_encoding(data, s′)
 
     return s′
 end
@@ -424,11 +407,9 @@ function initialize_state(ce::CounterfactualExplanation)
 
     # Add random perturbation following Slack (2021): https://arxiv.org/abs/2106.02666
     if ce.initialization == :add_perturbation
[…]
