### 2023-04-21 09dc770
Update src/counterfactuals/functions.jl
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
