### 2023-04-21 2d533fd
Update src/counterfactuals/plotting.jl
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
