### 2023-04-21 dfb8af3
Update src/generate_counterfactual.jl
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
