### 2023-04-21 65e1339
Update src/models/differentiable/flux_mlp.jl
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
