+++ src/models/differentiable/flux_mlp.jl
@@ -153,8 +153,6 @@ function build_mlp(;
         hidden_ = repeat([Dense(n_hidden, n_hidden, activation)], n_layers - 2)
         model = Chain(
             Dense(input_dim, n_hidden, activation), hidden_..., Dense(n_hidden, output_dim)
-        )
-    end
 
     return model
 end
