+++ src/evaluation/measures.jl
@@ -17,10 +17,7 @@ end
 Checks if the counterfactual search has been strictly valid in the sense that it has converged with respect to the pre-specified target probability `γ`.
 """
 function validity_strict(ce::CounterfactualExplanation)
-    return validity(
-        ce;
-        γ=ce.convergence[:decision_threshold],
-    )
+    return validity(ce; γ=ce.convergence[:decision_threshold])
 end
 
 """
