+++ src/counterfactuals/plotting.jl
@@ -39,9 +39,7 @@ function Plots.plot(
         minimum([plot_up_to, max_iter])
     end
     max_iter += 1
-    ingredients = set_up_plots(
-        ce; alpha=alpha_, plot_proba=plot_proba, kwargs...
-    )
+    ingredients = set_up_plots(ce; alpha=alpha_, plot_proba=plot_proba, kwargs...)
 
     for t in 1:max_iter
         final_state = t == max_iter
@@ -87,9 +85,7 @@ function animate_path(
         minimum([plot_up_to, max_iter])
     end
     max_iter += 1
-    ingredients = set_up_plots(
-        ce; alpha=alpha_, plot_proba=plot_proba, kwargs...
-    )
+    ingredients = set_up_plots(ce; alpha=alpha_, plot_proba=plot_proba, kwargs...)
 
     anim = @animate for t in 1:max_iter
         final_state = t == max_iter
@@ -113,12 +109,7 @@ end
 
 Helper function that plots a single step of the counterfactual path.
 """
-function plot_state(
-    ce::CounterfactualExplanation,
-    t::Int,
-    final_sate::Bool;
-    kwargs...,
-)
+function plot_state(ce::CounterfactualExplanation, t::Int, final_sate::Bool; kwargs...)
     args = PlotIngredients(; kwargs...)
     x1 = selectdim(args.path_embedded[t], 1, 1)
     x2 = selectdim(args.path_embedded[t], 1, 2)
@@ -173,16 +164,8 @@ end
 
 A helper method that prepares data for plotting.
 """
-function set_up_plots(
-    ce::CounterfactualExplanation; alpha, plot_proba, kwargs...
-)
-    p1 = Models.plot(
-        ce.M,
-        ce.data;
-        target=ce.target,
-        alpha=alpha,
-        kwargs...,
-    )
+function set_up_plots(ce::CounterfactualExplanation; alpha, plot_proba, kwargs...)
+    p1 = Models.plot(ce.M, ce.data; target=ce.target, alpha=alpha, kwargs...)
     p2 = plot(; xlims=(1, total_steps(ce) + 1), ylims=(0, 1))
     path_embedded = embed_path(ce)
     path_labels = counterfactual_label_path(ce)
