+++ src/counterfactuals/functions.jl
@@ -113,8 +113,7 @@ function CounterfactualExplanation(
         :iteration_count => 0,
         :times_changed_features => zeros(size(decode_state(ce))),
         :path => [ce.s′],
-        :terminated =>
-            threshold_reached(ce, ce.x),
+        :terminated => threshold_reached(ce, ce.x),
         :converged => converged(ce),
     )
 
@@ -182,9 +181,7 @@ function adjust_shape!(ce::CounterfactualExplanation)
     s′ = adjust_shape(ce, x)      # augment to account for specified number of counterfactuals
     ce.s′ = s′
     target_encoded = ce.target_encoded
-    ce.target_encoded = adjust_shape(
-        ce, target_encoded
-    )
+    ce.target_encoded = adjust_shape(ce, target_encoded)
 
     # Parameters:
     params = ce.params
@@ -246,9 +243,7 @@ function wants_latent_space(ce::CounterfactualExplanation)
     latent_space = ce.params[:latent_space]
 
     # If threshold is already reached, training GM is redundant:
-    latent_space =
-        latent_space &&
-        !threshold_reached(ce, ce.x)
+    latent_space = latent_space && !threshold_reached(ce, ce.x)
 
     return latent_space
 end
@@ -262,8 +257,7 @@ end
 Maps `x` from the feature space $\mathcal{X}$ to the latent space learned by the generative model.
 """
 function map_to_latent(
-    ce::CounterfactualExplanation,
-    x::Union{AbstractArray,Nothing}=nothing,
+    ce::CounterfactualExplanation, x::Union{AbstractArray,Nothing}=nothing
 )
 
     # Unpack:
@@ -298,8 +292,7 @@ Applies all the applicable decoding functions:
 Finally, the decoded counterfactual is returned.
 """
 function decode_state(
-    ce::CounterfactualExplanation,
-    x::Union{AbstractArray,Nothing}=nothing,
+    ce::CounterfactualExplanation, x::Union{AbstractArray,Nothing}=nothing
 )
 
     # Unpack:
@@ -339,8 +332,7 @@ end
 Maps the state variable back from the latent space to the feature space.
 """
 function map_from_latent(
-    ce::CounterfactualExplanation,
-    x::Union{AbstractArray,Nothing}=nothing,
+    ce::CounterfactualExplanation, x::Union{AbstractArray,Nothing}=nothing
 )
 
     # Unpack:
@@ -372,8 +364,7 @@ end
 Reconstructs all categorical encodings. See [`DataPreprocessing.reconstruct_cat_encoding`](@ref) for details.
 """
 function reconstruct_cat_encoding(
-    ce::CounterfactualExplanation,
-    x::Union{AbstractArray,Nothing}=nothing,
+    ce::CounterfactualExplanation, x::Union{AbstractArray,Nothing}=nothing
 )
     # Unpack:
     s′ = isnothing(x) ? deepcopy(ce.s′) : x
@@ -494,8 +485,7 @@ end
 Returns the predicted probability of the target class for `x`. If `x` is `nothing`, the predicted probability corresponding to the counterfactual value is returned.
 """
 function target_probs(
-    ce::CounterfactualExplanation,
-    x::Union{AbstractArray,Nothing}=nothing,
+    ce::CounterfactualExplanation, x::Union{AbstractArray,Nothing}=nothing
 )
     data = ce.data
     likelihood = ce.data.likelihood
@@ -533,9 +523,7 @@ end
 
 Returns the counterfactual probabilities for each step of the search.
 """
-function counterfactual_probability_path(
-    ce::CounterfactualExplanation
-)
+function counterfactual_probability_path(ce::CounterfactualExplanation)
     M = ce.M
     p = map(X -> counterfactual_probability(ce, X), path(ce))
     return p
[…]

[…]
