+++ Project.toml
@@ -35,6 +35,7 @@ UMAP = "c4f8c510-2410-5be4-91d7-4fbaeb39457e"
 CSV = "0.10"
 CUDA = "3, 4"
 CategoricalArrays = "0.10"
+ChainRulesCore = "1.15"
 DataFrames = "1"
 Flux = "0.12, 0.13"
 LaplaceRedux = "0.1"
