+++ v0.1.10/tutorials/data_preprocessing.qmd
@@ -1,0 +1,280 @@
+```@meta
+CurrentModule = CounterfactualExplanations 
+```
+
+```{julia}
+#| echo: false
+include("docs/setup_docs.jl")
+eval(setup_docs)
+```
+
+# Handling Data
+
+The package works with custom data containers that contain the input and output data as well as information about the type and mutability of features. In this tutorial, we will see how data can be prepared for use with the package. 
+
+## Basic Functionality
+
+To demonstrate the basic way to prepare data, let's look at a standard benchmark dataset: Fisher's classic iris dataset. We can use `MLDatasets` to load this data.
+
+```{julia}
+dataset = Iris()
+```
+
+Our data constructor `CounterfactualData` needs at least two inputs: features `X` and targets `y`. 
+
+```{julia}
+X = dataset.features
+y = dataset.targets
+```
+
+Next, we convert the input data to a `Tables.MatrixTable` (following `MLJ.jl`) convention. Concerning the target variable, we just assign grab the first column of the data frame. 
+
+```{julia}
+X = table(Tables.matrix(X))
+y = y[:,1]
+```
+
+Now we can feed these two ingredients to our constructor:
+
+```{julia}
+counterfactual_data = CounterfactualData(X, y)
+```
+
+Under the hood, the constructor performs basic preprocessing steps. For example, the output variable `y` is automatically one-hot encoded:
+
+```{julia}
+#| output: true
+
+counterfactual_data.y
+```
+
+Similarly, a transformer used to scale continuous input features is automatically fitted:
+
+```{julia}
+#| output: true
+
+counterfactual_data.dt
+```
+
+## Categorical Features
+
+For the counterfactual search, it is important to distinguish between continuous and categorical features. This is because categorical features cannot be perturbed arbitrarily: they can take specific discrete values, but not just any value on the real line. 
+
+Consider the following example:
+
+```{julia}
+#| output: true
+y = rand([1,0],4)
+X = (
+    name=categorical(["Danesh", "Lee", "Mary", "John"]),
+    grade=categorical(["A", "B", "A", "C"], ordered=true),
+    sex=categorical(["male","female","male","male"]),
+    height=[1.85, 1.67, 1.5, 1.67],
+)
+schema(X)
+```
+
+Typically, in the context of Unserpervised Learning, categorical features are one-hot or dummy encoded. To this end, we could use `MLJ`, for example:
+
+```{julia}
+hot = OneHotEncoder()
+mach = MLJBase.fit!(machine(hot, X))
+W = MLJBase.transform(mach, X)
+X = permutedims(MLJBase.matrix(W))
+```
+
+In all likelihood, this pre-processing step already happens at the stage, when the supervised model is trained. Since our counterfactual generators need to work in the same feature domain as the model they are intended to explain, we assume that categorical features are already encoded.
+
+The `CounterfactualData` constructor takes two optional arguments that can be used to specify the indices of categorical and continuous features. By default, all features are assumed to be continuous. For categorical features, the constructor expects an array of arrays of integers (`Vector{Vector{Int}}`) where each subarray includes the indices of all one-hot encoded rows related to a single categorical feature. In the example above, the `name` feature is one-hot encoded across rows 1, 2, 3 and 4 of `X`, the `grade` feature is encoded across the following three rows, etc.
+
+```{julia}
+#| output: true
+schema(W)
+```
+
+The code chunk below assigns the categorical and continuous feature indices:
+
+```{julia}
+features_categorical = [
+    [1,2,3,4],      # name
+    [5,6,7],        # grade
+    [8,9]           # sex
+]
+features_continuous = [10]
+```
+
[…]
