+++ src/objectives/penalties.jl
@@ -1,18 +1,40 @@
+using ChainRulesCore
 using ..CounterfactualExplanations
 using LinearAlgebra
 using SliceMap
-using Statistics: mean
+using Statistics: mean, median
 
 """
     distance(ce::AbstractCounterfactualExplanation, p::Real=2)
 
 Computes the distance of the counterfactual to the original factual.
 """
-function distance(ce::AbstractCounterfactualExplanation, p::Real=2; agg=mean)
+function distance(ce::AbstractCounterfactualExplanation; agg=mean, p::Real=1, weights::Union{Nothing,AbstractArray}=nothing)
     x = CounterfactualExplanations.factual(ce)
     x′ = CounterfactualExplanations.counterfactual(ce)
-    Δ = agg(norm(x′ .- x, p))
+    xs = eachslice(x′, dims=ndims(x′))                      # slices along the last dimension (i.e. the number of counterfactuals)
+    if isnothing(weights)
+        Δ = agg(map(x′ -> norm(x′ .- x, p), xs))            # aggregate across counterfactuals
+    else
+        @assert length(weights) == size(first(xs), ndims(first(xs))) "The length of the weights vector must match the number of features."
+        Δ = agg(map(x′ -> norm.(x′ .- x, p)'weights, xs))   # aggregate across counterfactuals
+    end            
     return Δ
+end
+
+"""
+    distance_mad(ce::AbstractCounterfactualExplanation; agg=mean)
+
+This is the distance measure proposed by Wachter et al. (2017).
+"""
+function distance_mad(ce::AbstractCounterfactualExplanation; agg=mean)
+    X = ce.data.X
+    mad = []
+    ignore_derivatives() do
+        _mad = median(abs.(X .- median(X, dims=ndims(X))), dims=ndims(X))
+        push!(mad, _mad)
+    end
+    return distance(ce; agg=agg, weights=1.0 ./ mad[1])[1]
 end
 
 """
@@ -21,7 +43,7 @@ end
 Computes the L0 distance of the counterfactual to the original factual.
 """
 function distance_l0(ce::AbstractCounterfactualExplanation; agg=mean)
-    return distance(ce, 0; agg=agg)
+    return distance(ce; p=0, agg=agg)
 end
 
 """
@@ -30,7 +52,7 @@ end
 Computes the L1 distance of the counterfactual to the original factual.
 """
 function distance_l1(ce::AbstractCounterfactualExplanation; agg=mean)
-    return distance(ce, 1; agg=agg)
+    return distance(ce; p=1, agg=agg)
 end
 
 """
@@ -39,7 +61,7 @@ end
 Computes the L2 (Euclidean) distance of the counterfactual to the original factual.
 """
 function distance_l2(ce::AbstractCounterfactualExplanation; agg=mean)
-    return distance(ce, 2; agg=agg)
+    return distance(ce; p=3, agg=agg)
 end
 
 """
@@ -48,7 +70,7 @@ end
 Computes the L-inf distance of the counterfactual to the original factual.
 """
 function distance_linf(ce::AbstractCounterfactualExplanation; agg=mean)
-    return distance(ce, Inf; agg=agg)
+    return distance(ce; p=Inf, agg=agg)
 end
 
 """
