+++ src/CounterfactualExplanations.jl
@@ -85,6 +85,6 @@ include("evaluation/Evaluation.jl")
 using .Evaluation
 
 # Precompile:
-# include("precompile.jl")
+include("precompile.jl")
 
 end
