+++ src/generators/gradient_based/generators.jl
@@ -1,37 +1,40 @@
-const default_distance = Objectives.distance_mad
+const default_distance = Objectives.distance_l1
 
 "Constructor for `GenericGenerator`."
 function GenericGenerator(; λ::AbstractFloat=0.1, kwargs...)
     return Generator(; penalty=default_distance, λ=λ, kwargs...)
 end
 
+"Constructor for `WachterGenerator`."
+function WachterGenerator(; λ::AbstractFloat=0.1, kwargs...)
+    return Generator(; penalty=Objectives.distance_mad, λ=λ, kwargs...)
+end
+
 "Constructor for `DiCEGenerator`."
-function DiCEGenerator(; λ::Vector{<:AbstractFloat}=[0.1, 1.0], kwargs...)
+function DiCEGenerator(; λ::Vector{<:AbstractFloat}=[0.1, 0.1], kwargs...)
     _penalties = [default_distance, Objectives.ddp_diversity]
     return Generator(; penalty=_penalties, λ=λ, kwargs...)
 end
 
 "Constructor for `ClaPGenerator`."
-function ClaPROARGenerator(; λ::Vector{<:AbstractFloat}=[0.1, 1.0], kwargs...)
+function ClaPROARGenerator(; λ::Vector{<:AbstractFloat}=[0.1, 0.1], kwargs...)
     _penalties = [default_distance, Objectives.model_loss_penalty]
     return Generator(; penalty=_penalties, λ=λ, kwargs...)
 end
 
 "Constructor for `GravitationalGenerator`."
-function GravitationalGenerator(; λ::Vector{<:AbstractFloat}=[0.1, 1.0], kwargs...)
+function GravitationalGenerator(; λ::Vector{<:AbstractFloat}=[0.1, 0.1], kwargs...)
     _penalties = [default_distance, Objectives.distance_from_target]
     return Generator(; penalty=_penalties, λ=λ, kwargs...)
 end
 
 "Constructor for `REVISEGenerator`."
 function REVISEGenerator(; λ::AbstractFloat=0.1, latent_space=true, kwargs...)
-    return Generator(;
-        penalty=default_distance, λ=λ, latent_space=latent_space, kwargs...
-    )
+    return Generator(; penalty=default_distance, λ=λ, latent_space=latent_space, kwargs...)
 end
 
 "Constructor for `GreedyGenerator`."
-function GreedyGenerator(; λ::AbstractFloat=0.1, η=0.1, n=nothing, kwargs...)
+function GreedyGenerator(; η=0.1, n=nothing, kwargs...)
     opt = CounterfactualExplanations.Generators.JSMADescent(; η=η, n=n)
-    return Generator(; penalty=default_distance, λ=λ, opt=opt, kwargs...)
+    return Generator(; penalty=default_distance, λ=0.0, opt=opt, kwargs...)
 end
