+++ src/objectives/penalties.jl
@@ -9,15 +9,24 @@ using Statistics: mean, median
 
 Computes the distance of the counterfactual to the original factual.
 """
-function distance(ce::AbstractCounterfactualExplanation; agg=mean, p::Real=1, weights::Union{Nothing,AbstractArray}=nothing)
+function distance(
+    ce::AbstractCounterfactualExplanation; 
+    from::Union{Nothing,AbstractArray}=nothing, 
+    agg=mean, 
+    p::Real=1, 
+    weights::Union{Nothing,AbstractArray}=nothing
+)
+    if isnothing(from)
+        from = CounterfactualExplanations.factual(ce)
+    end
     x = CounterfactualExplanations.factual(ce)
     x′ = CounterfactualExplanations.counterfactual(ce)
     xs = eachslice(x′, dims=ndims(x′))                      # slices along the last dimension (i.e. the number of counterfactuals)
     if isnothing(weights)
-        Δ = agg(map(x′ -> norm(x′ .- x, p), xs))            # aggregate across counterfactuals
+        Δ = agg(map(x′ -> norm(x′ .- from, p), xs))            # aggregate across counterfactuals
     else
         @assert length(weights) == size(first(xs), ndims(first(xs))) "The length of the weights vector must match the number of features."
-        Δ = agg(map(x′ -> norm.(x′ .- x, p)'weights, xs))   # aggregate across counterfactuals
+        Δ = agg(map(x′ -> norm.(x′ .- from, p)'weights, xs))   # aggregate across counterfactuals
     end            
     return Δ
 end
@@ -42,8 +51,8 @@ end
 
 Computes the L0 distance of the counterfactual to the original factual.
 """
-function distance_l0(ce::AbstractCounterfactualExplanation; agg=mean)
-    return distance(ce; p=0, agg=agg)
+function distance_l0(ce::AbstractCounterfactualExplanation; kwrgs...)
+    return distance(ce; p=0, kwrgs...)
 end
 
 """
@@ -51,8 +60,8 @@ end
 
 Computes the L1 distance of the counterfactual to the original factual.
 """
-function distance_l1(ce::AbstractCounterfactualExplanation; agg=mean)
-    return distance(ce; p=1, agg=agg)
+function distance_l1(ce::AbstractCounterfactualExplanation; kwrgs...)
+    return distance(ce; p=1, kwrgs...)
 end
 
 """
@@ -60,8 +69,8 @@ end
 
 Computes the L2 (Euclidean) distance of the counterfactual to the original factual.
 """
-function distance_l2(ce::AbstractCounterfactualExplanation; agg=mean)
-    return distance(ce; p=3, agg=agg)
+function distance_l2(ce::AbstractCounterfactualExplanation; kwrgs...)
+    return distance(ce; p=3, kwrgs...)
 end
 
 """
@@ -69,8 +78,8 @@ end
 
 Computes the L-inf distance of the counterfactual to the original factual.
 """
-function distance_linf(ce::AbstractCounterfactualExplanation; agg=mean)
-    return distance(ce; p=Inf, agg=agg)
+function distance_linf(ce::AbstractCounterfactualExplanation; kwrgs...)
+    return distance(ce; p=Inf, kwrgs...)
 end
 
 """
@@ -103,13 +112,12 @@ end
 Computes the distance of the counterfactual from a point in the target main.
 """
 function distance_from_target(
-    ce::AbstractCounterfactualExplanation, p::Int=2; agg=mean, K::Int=50
+    ce::AbstractCounterfactualExplanation; K::Int=50, kwrgs...
 )
     ids = rand(1:size(ce.params[:potential_neighbours], 2), K)
[…]
