### 2023-05-01 cc1cfb6
added separate wachter implementation with exact distance function
