+++ src/generators/Generators.jl
@@ -41,6 +41,7 @@ generator_catalogue = Dict(
     :greedy => Generators.GreedyGenerator,
     :revise => Generators.REVISEGenerator,
     :dice => Generators.DiCEGenerator,
+    :wachter => Generators.WachterGenerator,
 )
 
 end
