+++ src/generators/Generators.jl
@@ -18,6 +18,7 @@ export GravitationalGenerator
 export GreedyGenerator
 export REVISEGenerator
 export DiCEGenerator
+export WachterGenerator
 export generator_catalogue
 export generate_perturbations, conditions_satisfied, mutability_constraints
 export Generator
