+++ src/CounterfactualExplanations.jl
@@ -56,6 +56,7 @@ export GravitationalGenerator
 export GreedyGenerator
 export REVISEGenerator
 export DiCEGenerator
+export WachterGenerator
 export generator_catalogue
 export generate_perturbations, conditions_satisfied, mutability_constraints
 export Generator, @objective, @threshold
