+++ test/evaluation.jl
@@ -3,4 +3,3 @@ using DataFrames
 
 bmk = benchmark(counterfactual_data; converge_when=:generator_conditions)
 @test typeof(bmk.counterfactual_explanations) <: Vector{CounterfactualExplanation}
-@test typeof(bmk()) <: DataFrame
