### 2023-05-02 fc5e653
adds wrapper function to load all tabular data
