### 2023-05-03 6bd8a1f
this is sorted and also som minor bug fixes
