+++ src/evaluation/benchmark.jl
@@ -2,9 +2,8 @@ using DataFrames
 using Statistics
 using ..Models: train
 
-"A container for benchmarks of counterfactual explanations."
+"A container for benchmarks of counterfactual explanations. Instead of subtyping `DataFrame`, it contains a `DataFrame` of evaluation measures (see [this discussion](https://discourse.julialang.org/t/creating-an-abstractdataframe-subtype/36451/6?u=pat-alt) for why we don't subtype `DataFrame` directly)."
 struct Benchmark
-    counterfactual_explanations::Vector{CounterfactualExplanation}
     evaluation::DataFrame
 end
 
@@ -36,31 +35,30 @@ function Base.vcat(
     idcol_name="dataset",
 )
     @assert isnothing(ids) || length(ids) == 2
-    ces = vcat(bmk1.counterfactual_explanations, bmk2.counterfactual_explanations)
     if !isnothing(ids)
         bmk1.evaluation[!, idcol_name] .= ids[1]
         bmk2.evaluation[!, idcol_name] .= ids[2]
     end
     evaluation = vcat(bmk1.evaluation, bmk2.evaluation)
-    bmk = Benchmark(ces, evaluation)
+    bmk = Benchmark(evaluation)
     return bmk
 end
-
-function aggregate(bmk::Benchmark) end
 
 """
     benchmark(
         counterfactual_explanations::Vector{CounterfactualExplanation};
         meta_data::Union{Nothing,<:Vector{<:Dict}}=nothing,
-        measure::Union{Function,Vector{Function}}=default_measures
+        measure::Union{Function,Vector{Function}}=default_measures,
+        store_ce::Bool=false,
     )
 
-Generates a `Benchmark` for a vector of counterfactual explanations. Optionally `meta_data` describing each individual counterfactual explanation can be supplied. This should be a vector of dictionaries of the same length as the vector of counterfactuals. If no `meta_data` is supplied, it will be automatically inferred. 
+Generates a `Benchmark` for a vector of counterfactual explanations. Optionally `meta_data` describing each individual counterfactual explanation can be supplied. This should be a vector of dictionaries of the same length as the vector of counterfactuals. If no `meta_data` is supplied, it will be automatically inferred. All `measure` functions are applied to each counterfactual explanation. If `store_ce=true`, the counterfactual explanations are stored in the benchmark.
 """
 function benchmark(
     counterfactual_explanations::Vector{CounterfactualExplanation};
     meta_data::Union{Nothing,<:Vector{<:Dict}}=nothing,
     measure::Union{Function,Vector{Function}}=default_measures,
+    store_ce::Bool=false,
 )
     evaluations = evaluate(
         counterfactual_explanations;
@@ -68,9 +66,9 @@ function benchmark(
         report_each=true,
         report_meta=true,
         meta_data=meta_data,
-        store_ce=true,
+        store_ce=store_ce,
     )
-    bmk = Benchmark(counterfactual_explanations, evaluations)
+    bmk = Benchmark(evaluations)
     return bmk
 end
 
