+++ src/evaluation/measures.jl
@@ -8,7 +8,7 @@ using Statistics: mean
 Checks of the counterfactual search has been successful with respect to the probability threshold `γ`. In case multiple counterfactuals were generated, the function returns the proportion of successful counterfactuals.
 """
 function validity(ce::CounterfactualExplanation; agg=mean, γ=0.5)
-    return agg(CounterfactualExplanations.target_probs(ce) .>= γ)[1]
+    return vec(agg(CounterfactualExplanations.target_probs(ce) .>= γ))
 end
 
 """
@@ -27,6 +27,6 @@ Computes the feature redundancy: that is, the number of features that remain unc
 """
 function redundancy(ce::CounterfactualExplanation; agg=mean, tol=1e-5)
     x′ = CounterfactualExplanations.counterfactual(ce)
-    redundant_x = agg(sum(abs.(x′ .- ce.x) .< tol) / size(x′, 1))
+    redundant_x = [agg(sum(abs.(x .- ce.x) .< tol) / size(x, 1)) for x in eachslice(x′, dims=ndims(x′))]
     return redundant_x
 end
