+++ docs/src/tutorials/benchmarking.md
@@ -39,21 +39,21 @@ bmk()
      Row │ sample  variable    value     generator                          model  ⋯
          │ Int64   String      Float64   Symbol                             Symbol ⋯
     ─────┼──────────────────────────────────────────────────────────────────────────
-       1 │      1  distance    0.752493  GenericGenerator(nothing, Linear…  FluxMo ⋯
-       2 │      1  redundancy  0.0       GenericGenerator(nothing, Linear…  FluxMo
-       3 │      1  validity    1.0       GenericGenerator(nothing, Linear…  FluxMo
-       4 │      2  distance    0.733512  GenericGenerator(nothing, Linear…  FluxMo
-       5 │      2  redundancy  0.0       GenericGenerator(nothing, Linear…  FluxMo ⋯
-       6 │      2  validity    1.0       GenericGenerator(nothing, Linear…  FluxMo
-       7 │      3  distance    0.727664  GenericGenerator(nothing, Linear…  FluxMo
-       8 │      3  redundancy  0.0       GenericGenerator(nothing, Linear…  FluxMo
-       9 │      3  validity    1.0       GenericGenerator(nothing, Linear…  FluxMo ⋯
-      10 │      4  distance    0.598759  GenericGenerator(nothing, Linear…  FluxMo
-      11 │      4  redundancy  0.0       GenericGenerator(nothing, Linear…  FluxMo
-      12 │      4  validity    1.0       GenericGenerator(nothing, Linear…  FluxMo
-      13 │      5  distance    0.725115  GenericGenerator(nothing, Linear…  FluxMo ⋯
-      14 │      5  redundancy  0.0       GenericGenerator(nothing, Linear…  FluxMo
-      15 │      5  validity    1.0       GenericGenerator(nothing, Linear…  FluxMo
+       1 │      1  distance    0.956787  Generator(nothing, Counterfactua…  FluxMo ⋯
+       2 │      1  redundancy  0.0       Generator(nothing, Counterfactua…  FluxMo
+       3 │      1  validity    0.8       Generator(nothing, Counterfactua…  FluxMo
+       4 │      2  distance    1.05868   Generator(nothing, Counterfactua…  FluxMo
+       5 │      2  redundancy  0.0       Generator(nothing, Counterfactua…  FluxMo ⋯
+       6 │      2  validity    1.0       Generator(nothing, Counterfactua…  FluxMo
+       7 │      3  distance    1.04356   Generator(nothing, Counterfactua…  FluxMo
+       8 │      3  redundancy  0.0       Generator(nothing, Counterfactua…  FluxMo
+       9 │      3  validity    0.8       Generator(nothing, Counterfactua…  FluxMo ⋯
+      10 │      4  distance    0.92757   Generator(nothing, Counterfactua…  FluxMo
+      11 │      4  redundancy  0.0       Generator(nothing, Counterfactua…  FluxMo
+      12 │      4  validity    0.8       Generator(nothing, Counterfactua…  FluxMo
+      13 │      5  distance    1.02967   Generator(nothing, Counterfactua…  FluxMo ⋯
+      14 │      5  redundancy  0.0       Generator(nothing, Counterfactua…  FluxMo
+      15 │      5  validity    0.8       Generator(nothing, Counterfactua…  FluxMo
                                                                    3 columns omitted
 
 To retrieve the granular dataset, simply do:
@@ -66,28 +66,28 @@ bmk(agg=nothing)
      Row │ sample  num_counterfactual  variable    value     generator             ⋯
          │ Int64   Int64               String      Float64   Symbol                ⋯
     ─────┼──────────────────────────────────────────────────────────────────────────
-       1 │      1                   1  distance    0.74913   GenericGenerator(noth ⋯
[…]
