+++ test/evaluation.jl
@@ -2,4 +2,4 @@ using CounterfactualExplanations.Evaluation
 using DataFrames
 
 bmk = benchmark(counterfactual_data; converge_when=:generator_conditions)
-@test typeof(bmk.counterfactual_explanations) <: Vector{CounterfactualExplanation}
+@test typeof(bmk()) <: DataFrame
