+++ src/evaluation/measures.jl
@@ -8,7 +8,9 @@ using Statistics: mean
 Checks of the counterfactual search has been successful with respect to the probability threshold `γ`. In case multiple counterfactuals were generated, the function returns the proportion of successful counterfactuals.
 """
 function validity(ce::CounterfactualExplanation; agg=mean, γ=0.5)
-    return vec(agg(CounterfactualExplanations.target_probs(ce) .>= γ))
+    val = agg(CounterfactualExplanations.target_probs(ce) .>= γ)
+    val = val isa AbstractMatrix ? vec(val) : val
+    return val
 end
 
 """
@@ -30,5 +32,6 @@ function redundancy(ce::CounterfactualExplanation; agg=mean, tol=1e-5)
     redundant_x = [
         agg(sum(abs.(x .- ce.x) .< tol) / size(x, 1)) for x in eachslice(x′; dims=ndims(x′))
     ]
+    redundant_x = length(redundant_x) == 1 ? redundant_x[1] : redundant_x
     return redundant_x
 end
