+++ dev/artifacts/vision.qmd
@@ -45,7 +45,7 @@ Serialization.serialize(joinpath(model_path,"mnist_ensemble.jls"),M_ens)
 
 ```{julia}
 using CounterfactualExplanations.GenerativeModels: VAE, train!
-vae = VAE(input_dim; nll=Flux.logitbinarycrossentropy, epochs=100, λ=0.0001, latent_dim=28, hidden_dim=50)
+vae = VAE(input_dim; nll=Flux.Losses.mse, epochs=100, λ=0.0001, latent_dim=28, hidden_dim=50)
 X = counterfactual_data.X
 y = counterfactual_data.output_encoder.y        # raw output vector
 flux_training_params.verbose = true
@@ -54,10 +54,60 @@ Serialization.serialize(joinpath(model_path,"mnist_vae_strong.jls"),vae)
 ```
 
 ```{julia}
-vae = VAE(input_dim; nll=Flux.logitbinarycrossentropy, epochs=100, λ=0.05, latent_dim=2, hidden_dim=10)
+using CounterfactualExplanations.GenerativeModels: reconstruct
+images = []
+rec_images = []
+for i in 0:9
+    j = 0
+    while j < 10
+        x = X[:,rand(findall(y .== i))]
+        x̂ = reconstruct(vae, x)[1] |> 
+            x̂ -> clamp.((x̂ .+ 1.0) ./ 2.0, 0.0, 1.0) |>
+            x̂ -> reshape(x̂, 28,28) |>
+            x̂ -> convert2image(MNIST, x̂)
+        x = clamp.((x .+ 1.0) ./ 2.0, 0.0, 1.0) |> 
+            x -> reshape(x, 28,28) |>
+            x -> convert2image(MNIST, x)
+        push!(images, x)
+        push!(rec_images, x̂)
+        j += 1
+    end
+end
+p1 = plot(mosaic(images..., ncol=10), title="Images")
+p2 = plot(mosaic(rec_images..., ncol=10), title="Reconstructions")
+plot(p1, p2, axis=false, size=(800,375))
+```
+
+```{julia}
+vae = VAE(input_dim; nll=Flux.Losses.mse, epochs=100, λ=0.05, latent_dim=2, hidden_dim=10)
 flux_training_params.verbose = true
 train!(vae, X, y)
 Serialization.serialize(joinpath(model_path,"mnist_vae_weak.jls"),vae)
+```
+
+```{julia}
+using CounterfactualExplanations.GenerativeModels: reconstruct
+images = []
+rec_images = []
+for i in 0:9
+    j = 0
+    while j < 10
+        x = X[:,rand(findall(y .== i))]
+        x̂ = reconstruct(vae, x)[1] |> 
+            x̂ -> clamp.((x̂ .+ 1.0) ./ 2.0, 0.0, 1.0) |>
+            x̂ -> reshape(x̂, 28,28) |>
+            x̂ -> convert2image(MNIST, x̂)
+        x = clamp.((x .+ 1.0) ./ 2.0, 0.0, 1.0) |> 
+            x -> reshape(x, 28,28) |>
+            x -> convert2image(MNIST, x)
+        push!(images, x)
+        push!(rec_images, x̂)
+        j += 1
+    end
+end
+p1 = plot(mosaic(images..., ncol=10), title="Images")
+p2 = plot(mosaic(rec_images..., ncol=10), title="Reconstructions")
+plot(p1, p2, axis=false, size=(800,375))
 ```
 
 ## Generate Artifacts
