+++ dev/artifacts/vision.qmd
@@ -75,7 +75,8 @@ for i in 0:9
 end
 p1 = plot(mosaic(images..., ncol=10), title="Images")
 p2 = plot(mosaic(rec_images..., ncol=10), title="Reconstructions")
-plot(p1, p2, axis=false, size=(800,375))
+plt = plot(p1, p2, axis=false, size=(800,375))
+savefig(plt, joinpath(www_path, "mnist_vae_strong.png"))
 ```
 
 ```{julia}
@@ -107,12 +108,14 @@ for i in 0:9
 end
 p1 = plot(mosaic(images..., ncol=10), title="Images")
 p2 = plot(mosaic(rec_images..., ncol=10), title="Reconstructions")
-plot(p1, p2, axis=false, size=(800,375))
+plt = plot(p1, p2, axis=false, size=(800,375))
+savefig(plt, joinpath(www_path, "mnist_vae_weak.png"))
 ```
 
 ## Generate Artifacts
 
 ```{julia}
-generate_artifacts(model_path; artifact_name=model_name)
+artifact_name = "$(model_name)-$(Int(VERSION.major)).$(Int(VERSION.minor))"
+generate_artifacts(model_path; artifact_name=artifact_name)
 ```
 
