+++ src/data/Data.jl
@@ -26,7 +26,7 @@ const data_catalogue = Dict(
         :credit_default => load_credit_default,
         :gmsc => load_gmsc,
     ),
-    :vision => Dict(:mnist => load_mnist),
+    :vision => Dict(:mnist => load_mnist, :fashion_mnist => load_fashion_mnist),
 )
 
 """
@@ -72,5 +72,6 @@ export load_synthetic_data
 export load_california_housing, load_credit_default, load_gmsc
 export load_tabular_data
 export load_mnist, load_mnist_test
+export load_fashion_mnist, load_fashion_mnist_test
 
 end
