+++ Artifacts.toml
@@ -87,11 +87,11 @@ lazy = true
     url = "https://github.com/JuliaTrustworthyAI/CounterfactualExplanations.jl/releases/download/artifacts-1.8/model-vision.tar.gz"
 
 ["model-vision-1.8"]
-git-tree-sha1 = "f667444ce5302f195a28655322926c97a563eade"
+git-tree-sha1 = "f5e8433ca19481ad309be889ddac364c7847da0b"
 lazy = true
 
     [["model-vision-1.8".download]]
-    sha256 = "dfa000284cdbeab3e2ffc17c12cc3b9261bd7b499682f6da87976d3c6076d2f0"
+    sha256 = "48608adcda321e351680b2af26182a0e5356f785b80254892c41c6aaabca3397"
     url = "https://github.com/JuliaTrustworthyAI/CounterfactualExplanations.jl/releases/download/artifacts-1.8/model-vision-1.8.tar.gz"
 
 [paper-images]
