### 2023-05-09 9dbfa22
small changed to synthetic datasets. Previously scaling let to issues with VAE underlying revise
