+++ docs/src/explanation/generators/feature_tweak.md
@@ -1,0 +1,7 @@
+``` @meta
+CurrentModule = CounterfactualExplanations 
+```
+
+# `FeatureTweakGenerator`
+
+**Feature Tweaking** refers to the generator introduced by Tolomei et al. (2017)
