### 2023-05-10 b4b2d9f
Fix equation attempt 1
