+++ src/generators/gradient_based/base.jl
@@ -9,8 +9,8 @@ An abstract type that serves as the base type for gradient-based counterfactual 
 """
 abstract type AbstractGradientBasedGenerator <: AbstractGenerator end
 
-"Base class for counterfactual generators."
-mutable struct Generator <: AbstractGradientBasedGenerator
+"Base class for gradient-based counterfactual generators."
+mutable struct GradientBasedGenerator <: AbstractGradientBasedGenerator
     loss::Union{Nothing,Function}
     penalty::Union{Nothing,Function,Vector{Function}}
     λ::Union{Nothing,AbstractFloat,Vector{<:AbstractFloat}}
@@ -29,12 +29,12 @@ end
 
 Default outer constructor for `Generator`.
 """
-function Generator(;
+function GradientBasedGenerator(;
     loss::Union{Nothing,Function}=nothing,
     penalty::Union{Nothing,Function,Vector{Function}}=nothing,
     λ::Union{Nothing,AbstractFloat,Vector{<:AbstractFloat}}=nothing,
     latent_space::Bool=false,
     opt::Flux.Optimise.AbstractOptimiser=Flux.Descent(),
 )
-    return Generator(loss, penalty, λ, latent_space, opt)
+    return GradientBasedGenerator(loss, penalty, λ, latent_space, opt)
 end
