### 2023-05-11 97ca257
Add search_path function
