### 2023-05-11 2b9e497
Add source for inspiration to documentation
