+++ src/models/pretrained.jl
@@ -21,3 +21,22 @@ function load_mnist_vae(; strong=true)
     end
     return vae
 end
+
+function load_fashion_mnist_mlp()
+    M = deserialize(joinpath(vision_dir, "fashion_mnist_mlp.jls"))
+    return M
+end
+
+function load_fashion_mnist_ensemble()
+    M = deserialize(joinpath(vision_dir, "fashion_mnist_ensemble.jls"))
+    return M
+end
+
+function load_fashion_mnist_vae(; strong=true)
+    if strong
+        vae = deserialize(joinpath(vision_dir, "fashion_mnist_vae_strong.jls"))
+    else
+        vae = deserialize(joinpath(vision_dir, "fashion_mnist_vae_weak.jls"))
+    end
+    return vae
+end
