+++ src/generators/non_gradient_based/generators.jl
@@ -1,1 +1,3 @@
-function FeatureTweakGenerator(; )
+function FeatureTweakGenerator(; ϵ::AbstractFloat, kwargs...)
+    return HeuristicBasedGenerator(; ϵ=ϵ, kwargs...)
+end
