+++ src/generators/non_gradient_based/base.jl
@@ -7,23 +7,20 @@ abstract type AbstractNonGradientBasedGenerator <: AbstractGenerator end
 "Base class for heuristic/tree based counterfactual generators."
 mutable struct HeuristicBasedGenerator <: AbstractNonGradientBasedGenerator
     loss::Union{Nothing,Function}
-    penalty::Union{Nothing,Function,Vector{Function}}
-    λ::Union{Nothing,AbstractFloat,Vector{<:AbstractFloat}}
+    ϵ::Union{Nothing,AbstractFloat}
 end
 
 """
     HeuristicBasedGenerator(;
         loss::Union{Nothing,Function}=nothing,
-        penalty::Union{Nothing,Function,Vector{Function}}=nothing,
-        λ::Union{Nothing,AbstractFloat,Vector{AbstractFloat}}=nothing,
+        ϵ::Union{Nothing,AbstractFloat}=nothing,
     )
 
-Default outer constructor for `Generator`.
+Default outer constructor for `HeuristicBasedGenerator`.
 """
 function HeuristicBasedGenerator(;
     loss::Union{Nothing,Function}=nothing,
-    penalty::Union{Nothing,Function,Vector{Function}}=nothing,
-    λ::Union{Nothing,AbstractFloat,Vector{<:AbstractFloat}}=nothing,
+    ϵ::Union{Nothing,AbstractFloat}=nothing,
 )
-    return HeuristicBasedGenerator(loss, penalty, λ)
+    return HeuristicBasedGenerator(loss, ϵ)
 end
