### 2023-05-12 5749268
Update vision.qmd notebook with Cifar10 dataset
