### 2023-05-12 36926f2
Update vision.qmd notebook with Cifar10 dataset
