### 2023-05-13 cb5109e
Add test suite for Cifar10 dataset loaders
