+++ docs/setup_docs.jl
@@ -29,8 +29,8 @@ setup_docs = quote
     # Setup:
     theme(:wong)
     Random.seed!(2023)
-    www_path = "docs/src/www"
-    include("docs/src/utils.jl")
+    www_path = "$(pwd())/docs/src/www"
+    include("$(pwd())/docs/src/utils.jl")
     synthetic = CounterfactualExplanations.Data.load_synthetic_data()
     ENV["DATADEPS_ALWAYS_ACCEPT"] = "true"
 
