### 2023-05-15 fe737e6
Update src/data/vision.jl
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
