+++ test/data.jl
@@ -1,5 +1,4 @@
 ENV["DATADEPS_ALWAYS_ACCEPT"] = "true"
-
 @testset "Construction" begin
     for (name, loader) in merge(values(data_catalogue)...)
         @testset "$name" begin
@@ -14,20 +13,18 @@ end
     @test size(counterfactual_data.X)[2] == 50000
     @test size(counterfactual_data.X)[1] == 3072
     @test size(counterfactual_data.y)[2] == 50000
-    @test all(counterfactual_data.domain .== (-1.0, 1.0))
+    @test all(counterfactual_data.domain[i] == (-1.0, 1.0) for i in eachindex(counterfactual_data.domain))
     @test counterfactual_data.standardize == false
     @test eltype(counterfactual_data.X) == Float32
-    @test eltype(counterfactual_data.y) == Float32
 
     # Test loading CIFAR10 dataset with subsampled data
     counterfactual_data = load_cifar_10(1000)
     @test size(counterfactual_data.X)[2] == 1000
     @test size(counterfactual_data.X)[1] == 3072
     @test size(counterfactual_data.y)[2] == 1000
-    @test all(counterfactual_data.domain .== (-1.0, 1.0))
+    @test all(counterfactual_data.domain[i] == (-1.0, 1.0) for i in eachindex(counterfactual_data.domain))
     @test counterfactual_data.standardize == false
     @test eltype(counterfactual_data.X) == Float32
-    @test eltype(counterfactual_data.y) == Float32
 
     # Test loading CIFAR10 test dataset
     counterfactual_data = load_cifar_10_test()
@@ -36,5 +33,4 @@ end
     @test size(counterfactual_data.y)[2] == 10000
     @test counterfactual_data.standardize == false
     @test eltype(counterfactual_data.X) == Float32
-    @test eltype(counterfactual_data.y) == Float32
 end
