+++ dev/artifacts/vision.qmd
@@ -289,8 +289,8 @@ for i in 0:9
         x = X[:,rand(findall(y .== i))]
         x̂ = reconstruct(vae, x)[1] |> 
             x̂ -> clamp.((x̂ .+ 1.0) ./ 2.0, 0.0, 1.0) |>
-            x̂ -> reshape(x̂, 28,28) |>
-            x̂ -> convert2image(FashionMNIST, x̂)
+            x̂ -> reshape(x̂, 32, 32, 3) |>
+            x̂ -> convert2image(CIFAR10, x̂)
         push!(rec_images, x̂)
         j += 1
     end
@@ -298,14 +298,14 @@ end
 p1 = plot(mosaic(images..., ncol=10), title="Images")
 p2 = plot(mosaic(rec_images..., ncol=10), title="Reconstructions")
 plt = plot(p1, p2, axis=false, size=(800,375))
-savefig(plt, joinpath(www_path, "fashion_mnist_vae_strong.png"))
+savefig(plt, joinpath(www_path, "cifar_10_vae_strong.png"))
 ```
 
 ```{julia}
 vae = VAE(input_dim; nll=Flux.Losses.mse, epochs=100, λ=0.05, latent_dim=2, hidden_dim=10)
 flux_training_params.verbose = true
 train!(vae, X, y)
-Serialization.serialize(joinpath(model_path,"fashion_mnist_vae_weak.jls"),vae)
+Serialization.serialize(joinpath(model_path,"cifar_10_vae_weak.jls"),vae)
 ```
 
 ```{julia}
@@ -316,8 +316,8 @@ for i in 0:9
         x = X[:,rand(findall(y .== i))]
         x̂ = reconstruct(vae, x)[1] |> 
             x̂ -> clamp.((x̂ .+ 1.0) ./ 2.0, 0.0, 1.0) |>
-            x̂ -> reshape(x̂, 32,32) |>
-            x̂ -> convert2image(FashionMNIST, x̂)
+            x̂ -> reshape(x̂, 32, 32, 3) |>
+            x̂ -> convert2image(CIFAR10, x̂)
         push!(rec_images, x̂)
     end
 end
