+++ test/data.jl
@@ -25,7 +25,10 @@ end
     @test size(counterfactual_data.X)[2] == 1000
     @test size(counterfactual_data.X)[1] == 3072
     @test size(counterfactual_data.y)[2] == 1000
-    @test all(counterfactual_data.domain[i] == (-1.0, 1.0) for i in eachindex(counterfactual_data.domain))
+    @test all(
+        counterfactual_data.domain[i] == (-1.0, 1.0) for
+        i in eachindex(counterfactual_data.domain)
+    )
     @test counterfactual_data.standardize == false
     @test eltype(counterfactual_data.X) == Float32
 
