+++ src/generators/non_gradient_based/generators.jl
@@ -1,3 +1,3 @@
 function FeatureTweakGenerator(; ϵ::AbstractFloat, kwargs...)
-    return HeuristicBasedGenerator(; ϵ=ϵ, kwargs...)
+    return HeuristicBasedGenerator(; penalty=Objectives.distance_l2, ϵ=ϵ, kwargs...)
 end
