+++ students_documentation/README.md
@@ -1,0 +1,15 @@
+![counterfactual_explanations_logo](./pictures/logo.png)
+
+## Description of project
+
+In this project, we focus on contributing to CounterfactualExplanations.jl, a trustworthy artificial intelligence package for generating Counterfactual Explanations and Algorithmic Recourse for black-box algorithms in Julia. As contributors, we work on feature enhancements, including increasing the scope of predictive models that are compatible with the package. Our expected outcomes include familiarizing ourselves with the package, increasing our understanding of Counterfactual Explanations and Algorithmic Recourse, gaining relevant experience with FOSS development, and becoming familiar with Julia. This project provides us with an opportunity to get involved in Taija projects and contribute to the open-source community.
+
+## Group members
+
+| Name           | Email                         |
+| -------------- | ----------------------------- |
+| Lauri Keskul   |                               |
+| Rauno Arike    |                               |
+| Vincent Pikand |                               |
+| Simon Kasdorp  |                               |
+| Mariusz Kicior | m.a.kicior@student.tudelft.nl |
