+++ students_documentation/meeting_notes/TA meeting 16.05 - notes.md
@@ -1,0 +1,10 @@
+# TA meeting 16.05
+
+## Address TA feedback
+
+- Difference between start and dart issues: dart issue is a core of the project (client need them the most) and star issues are additional things we can do. It was originally created by our client to indicate which issues are meant to solved by students.
+- Definition of done: we created a definition of done in our project plan which can be basically applied to every issue we have.
+- Smaller issues: we're going to create smaller issues during the project going once we realise any issue is particularly big and should be tackled by multiple people
+- Dev branch: we created a branch `development` and there's a branch with a readme file already
+- What is the repository structure we're supposed to include? (We're going to add this to the readme file)
+- We added an article to our CoC about commit messages
