### 2023-05-16 4e65b34
Upload meeting notes, CoC and Project Plan
