+++ students_documentation/meeting_notes/TA meeting 25.04 - notes.md
@@ -1,0 +1,53 @@
+# TA Meeting 25.04
+
+## Practical Parts
+
+- We will be graded based on what we deliver from the "must-haves" and "could-haves." Be pessimistic and ensure that we actually accomplish what's important.
+
+- Focus on the problem, not the solution, in the project plan. Elicit all the requirements from the client in the Friday meeting.
+
+- The repository is set up, but we haven't been added yet.
+
+- Pipelines: we need to create our own pipeline. Keep it concise and not too long. Run pipelines only on specific branches (main and dev). We need to connect to Docker Hub. The TUDelft server can download it automatically if we specify the version. Avoid using the latest tag and use a specific version. If we encounter any problems, contact TA.
+
+- Consider mirroring repositories.
+
+- We are in control of the TA meetings. Set up the same routine for each week. TA is here to answer our questions and provide feedback from time to time. It's up to us to decide what we're doing. Reserve a place or find any location to sit down together for 15 minutes.
+
+## Q&A
+
+- No requirements for pull requests every week.
+
+- TA responsibilities: They only looks to see if we managed to work throughout the 10 weeks. It's important to work consistently throughout the whole project and not rush it at the end.
+
+- We don't need to add coach or TA to our Team's team.
+
+## What's the goal of the Software Project course?
+
+- We are supposed to create software around the problem. Show software architecture, UI, backend, or any relevant aspect. Communicate with a client that the course requires creating some sort of software, and implementing an algorithm is not sufficient. We should include visualizations, UI, backend, or other software-related aspects. If the client disagrees, approach TA for guidance.
+
+- Our focus is not to show that we can create ML models but to demonstrate that we can work together and create a piece of software.
+
+## TW: How to Write a Final Report?
+
+- Writing the final report is not as horrible as it seems but it requires dedicating a couple of days to it.
+
+## Deadlines
+
+- Code of Conduct and Project Plan should be completed before TW and teamwork-related tasks (week 5).
+
+- TA won't be available in person in week 4.
+
+- The course aims to conduct everything in person.
+
+- If there are any questions contact TA via Mattermost.
+
+## TODO
+
+- Write an email to the coach and inform them that we'll meet in the 3rd week.
+
+- Go through the example project plan and come up with as many questions as possible for our client.
+
+- Try to understand the problem and be as specific as possible. Gather requirements and validation techniques. Clearly state how we can verify things. Include numerical actions (e.g., testing above 90% coverage). Do not mention our solution in the project plan; we should have an idea of how we're going to approach it but avoid discussing it in the project plan. Keep the plan concise. The project plan is not graded; it's mainly for us to receive feedback on how well we understand the project.
+
+- Project plan, code of conduct, set up our workspace, set up pipelines.
