+++ src/generators/Generators.jl
@@ -12,7 +12,9 @@ using ..Objectives
 using Statistics
 
 export AbstractGradientBasedGenerator
+export AbstractNonGradientBasedGenerator
 export ClaPROARGenerator
+export FeatureTweakGenerator
 export GenericGenerator
 export GravitationalGenerator
 export GreedyGenerator
@@ -22,6 +24,7 @@ export WachterGenerator
 export generator_catalogue
 export generate_perturbations, conditions_satisfied, mutability_constraints
 export GradientBasedGenerator
+export HeuristicBasedGenerator
 export @objective, @threshold, @with_optimiser, @search_feature_space, @search_latent_space
 export JSMADescent
 
@@ -34,9 +37,15 @@ include("gradient_based/functions.jl")
 include("gradient_based/generators.jl")
 include("gradient_based/optimisers.jl")
 
+# Non-Gradient-Based Generators:
+include("non_gradient_based/base.jl")
+include("non_gradient_based/functions.jl")
+include("non_gradient_based/generators.jl")
+
 "A dictionary containing the constructors of all available counterfactual generators."
 generator_catalogue = Dict(
     :claproar => Generators.ClaPROARGenerator,
+    :feature_tweak => Generators.FeatureTweakGenerator,
     :generic => Generators.GenericGenerator,
     :gravitational => Generators.GravitationalGenerator,
     :greedy => Generators.GreedyGenerator,
