+++ dev/artifacts/test_feature_tweak.qmd
@@ -6,5 +6,38 @@ eval(setup)
 ```
 
 ```{julia}
-model_catalogue
+generator_catalogue
+```
+
+```{julia}
+n = 500
+counterfactual_data = load_moons(n)
+X = counterfactual_data.X
+y = counterfactual_data.y
+plt = plot()
+scatter!(counterfactual_data)
+```
+
+```{julia}
+X = float32.(X)
+y = string.(y[2,:])
+```
+
+```{julia}
+using DecisionTree
+model = DecisionTreeClassifier(max_depth=2)
+DecisionTree.fit!(model, X', y)
+```
+
+```{julia}
+DecisionTree.predict_proba(model, float32.([1.0,1.0]))
+DecisionTree.predict(model, float32.([-1.0,0.5]))
+```
+
+```{julia}
+M = model_catalogue[:DecisionTree](model, :classification_binary)
+x = float32.([-1.0,0.5])
+target = "true"
+generator = FeatureTweakGenerator(ϵ=0.1)
+ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
 ```
