+++ docs/src/explanation/generators/feature_tweak.qmd
@@ -1,0 +1,15 @@
+```@meta
+CurrentModule = CounterfactualExplanations 
+```
+
+```{julia}
+#| echo: false
+include("docs/setup_docs.jl")
+eval(setup_docs)
+```
+
+# `FeatureTweakGenerator`
+
+**Feature Tweak** refers to the generator introduced by Tolomei et al. (2017)
+
+Our implementation takes inspiration from the [featureTweakPy library](https://github.com/upura/featureTweakPy).
