+++ students_documentation/meeting_notes/TA meeting 16.05 - notes.md
@@ -2,9 +2,33 @@
 
 ## Address TA feedback
 
-- Difference between start and dart issues: dart issue is a core of the project (client need them the most) and star issues are additional things we can do. It was originally created by our client to indicate which issues are meant to solved by students.
+- Difference between start and dart issues: dart issue is a core of the project (client need them the most) and star issues are additional things we can do. It was originally created by our client to indicate which issues are meant to be solved by students.
+  - TA asked us to create only one tag "students" for our issues.
+
 - Definition of done: we created a definition of done in our project plan which can be basically applied to every issue we have.
-- Smaller issues: we're going to create smaller issues during the project going once we realise any issue is particularly big and should be tackled by multiple people
-- Dev branch: we created a branch `development` and there's a branch with a readme file already
+  - TA wants us to create a checkpoint list of to-do boxes explaining what has to be done to close the issue (e.g., proper documentation, how user can call it).
+ 
+- Smaller issues: we're going to create smaller issues during the project going once we realise any issue is too big and should be tackled by multiple people.
+
+- Dev branch: we created a branch `development` and there's a branch with a readme file already.
+
+- We added an article to our CoC about commit messages.
+  
 - What is the repository structure we're supposed to include? (We're going to add this to the readme file)
-- We added an article to our CoC about commit messages
+  - Pointers where to go if you want to find something (documentation for the code, for process, generators, datasets, etc.), e.g., "You can find X in a folder Y".
+
+## Questions
+
+- Milestones: are we supposed to add more than two milestones? Do we need to assign milestones to both issues and pull requests? 
+  - Ideally, split it up into smaller pieces, like for sprints (it's normal to move issues from one milestone to another)
+- Issues: are we supposed to create issues about the documentation (i.e., TW assignments)? 
+  - We don't have to.
+- Technical Writting: are we supposed to throw as many references to our report as possible? 
+  - At the very end they want we're expected to have at least 10 references, they should come from the introductions chapters, so now is the moment when we should add most of the references.
+
+## Midterm presentation
+
+- There's a rubric on Brightspace where we can find all relevant information (product and a process part rubric)
+- Presentation for 20 mins + 10 mins Q&A + 15 mins feedback
+- Book a room for the presentation (double check with coach)
+- No TA meeting next meeting (presentation is enough but if we want one we can ask our TA)
