+++ src/generators/gradient_based/functions.jl
@@ -4,7 +4,8 @@ using Statistics
 """
     ∂ℓ(generator::AbstractGradientBasedGenerator, M::Union{Models.LogisticModel, Models.BayesianLogisticModel}, ce::AbstractCounterfactualExplanation)
 
-The default method to compute the gradient of the loss function at the current counterfactual state for gradient-based generators. It assumes that `Zygote.jl` has gradient access.
+The default method to compute the gradient of the loss function at the current counterfactual state for gradient-based generators.
+It assumes that `Zygote.jl` has gradient access.
 """
 function ∂ℓ(
     generator::AbstractGradientBasedGenerator,
@@ -18,7 +19,8 @@ end
 """
     ∂h(generator::AbstractGradientBasedGenerator, ce::AbstractCounterfactualExplanation)
 
-The default method to compute the gradient of the complexity penalty at the current counterfactual state for gradient-based generators. It assumes that `Zygote.jl` has gradient access.
+The default method to compute the gradient of the complexity penalty at the current counterfactual state for gradient-based generators.
+It assumes that `Zygote.jl` has gradient access.
 """
 function ∂h(
     generator::AbstractGradientBasedGenerator, ce::AbstractCounterfactualExplanation
@@ -30,7 +32,8 @@ end
 """
     ∇(generator::AbstractGradientBasedGenerator, M::Models.AbstractDifferentiableModel, ce::AbstractCounterfactualExplanation)
 
-The default method to compute the gradient of the counterfactual search objective for gradient-based generators. It simply computes the weighted sum over partial derivates. It assumes that `Zygote.jl` has gradient access.
+The default method to compute the gradient of the counterfactual search objective for gradient-based generators.
+It simply computes the weighted sum over partial derivates. It assumes that `Zygote.jl` has gradient access.
 """
 function ∇(
     generator::AbstractGradientBasedGenerator,
