### 2023-05-17 e6f8f97
Added delta_hat calculation in the notebook
