### 2023-05-17 cbcf671
Added delta_hat calculation in the notebook
