+++ src/models/model_utils.jl
@@ -51,7 +51,6 @@ function predict_proba(
     @assert !(isnothing(counterfactual_data) && isnothing(X))
     X = isnothing(X) ? counterfactual_data.X : X
     p = probs(M, X)
-    # println(p)
     binary = M.likelihood == :classification_binary
     p = binary ? binary_to_onehot(p) : p
     return p
