+++ dev/artifacts/test_feature_tweak.qmd
@@ -1,6 +1,6 @@
 ```{julia}
 #| echo: false
-
+using Revise
 include("$(pwd())/dev/artifacts/setup.jl")
 eval(setup)
 ```
@@ -35,9 +35,13 @@ DecisionTree.predict(model, float32.([-1.0,0.5]))
 ```
 
 ```{julia}
+DecisionTree.predict(model, float32.([1.0,-0.5]))
+```
+
+```{julia}
 M = model_catalogue[:DecisionTree](model, :classification_binary)
 x = float32.([-1.0,0.5])
-target = "true"
+target = 1
 generator = FeatureTweakGenerator(ϵ=0.1)
 ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
 ```
