### 2023-05-17 8a106c6
Update readme with CIFAR 10 information
