+++ src/counterfactuals/functions.jl
@@ -108,6 +108,10 @@ function CounterfactualExplanation(
     ce.s′ = encode_state(ce)            # encode the counterfactual state
     ce.s′ = initialize_state(ce)        # initialize the counterfactual state
 
+    if generator isa Generators.HeuristicBasedGenerator
+        return ce
+    end
+
     # Initialize search:
     ce.search = Dict(
         :iteration_count => 0,
