### 2023-05-17 2fbbe44
Add CIFAR10 dataset to the summary.qmd
