+++ students_documentation/summary.qmd
@@ -1,33 +1,0 @@
-# Added Features to CounterfactualExplanations.jl
-
-In this notebook, we document the new features that we have added to the CounterfactualExplanations.jl package. These features include increasing the scope of predictive models that are compatible with the package, adding unit testing and documentation, and improving the package infrastructure. We hope that this notebook will serve as a useful reference for both ourselves and the course staff.
-
-## Must haves
-
-### <feature_name>
-
-<feature_description>
-
-```{julia}
-print("Hello world!")
-```
-
-## Should haves
-
-### <feature_name>
-
-<feature_description>
-
-```{julia}
-print("Hello world!")
-```
-
-## Could haves
-
-### <feature_name>
-
-<feature_description>
-
-```{julia}
-print("Hello world!")
-```
