### 2023-05-17 9524a12
Add CIFAR10 dataset to the summary.qmd
