### 2023-05-17 6bd3b12
Fix search_path function
