### 2023-05-17 e1937eb
Add CIFAR10 to the package documentation
