+++ dev/artifacts/test_feature_tweak.qmd
@@ -1,5 +1,4 @@
 ```{julia}
-#| echo: false
 using Revise
 include("$(pwd())/dev/artifacts/setup.jl")
 eval(setup)
@@ -36,13 +35,13 @@ DecisionTree.predict(model, float32.([-1.0,0.5]))
 ```
 
 ```{julia}
-DecisionTree.predict(model, float32.([1.0,-0.5]))
+DecisionTree.predict(model, float32.([2, 0]))
 ```
 
 ```{julia}
 M = model_catalogue[:DecisionTree](model, :classification_binary)
-x = float32.([-1.0,0.5])
-target = 1
+x = float32.([2, 0])
+target = 0
 generator = FeatureTweakGenerator(ϵ=0.1)
 ce = generate_counterfactual(x, target, counterfactual_data, M, generator)
 ```
