### 2023-05-18 8a50c7a
Small changes to logits and probs
