+++ .github/workflows/CI.yml
@@ -3,7 +3,6 @@ on:
   push:
     branches:
       - main
-      - develop
     tags: '*'
   pull_request:
 concurrency:
