+++ .github/workflows/CI-development.yml
@@ -1,8 +1,8 @@
-name: CI-development
+name: CI-tests
 on:
   push:
     branches:
-      - development
+      - *
 jobs:
   test:
     name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{ github.event_name }}
