+++ .github/workflows/CI-development.yml
@@ -2,7 +2,7 @@ name: CI-tests
 on:
   push:
     branches:
-      - *
+      - '*'
 jobs:
   test:
     name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{ github.event_name }}
