### 2023-05-19 3e31202
adding new version to be able to move to non-dev mode for ECCCo
