+++ Project.toml
@@ -51,7 +51,7 @@ Plots = "1.38.2"
 ProgressMeter = "1"
 SliceMap = "0.2"
 SnoopPrecompile = "1.0"
-StatsBase = "0.33"
+StatsBase = "0.33, 0.34"
 Tables = "1"
 UMAP = "0.1"
 julia = "1.7, 1.8, 1.9"
