### 2023-05-22 #103 @pat-alt
@matbesancon with JuliaCon 2023 approaching fast, is there anything I can do to accelerate this? 😄 Happy to ask for interest in reviewing this on discourse, slack and socials if that's acceptable.

We've had some major updates in meantime and now have a group of TU Delft students working on this package full time for a couple of weeks, so at this point it may well be worth updating the paper as well.

Let me know how if I can help in any way, would be great to see this move forward
