### 2023-06-20 #103 @kartikeyrinwa
### Introduction
In this project the authors have developed a Julia library to construct counterfactual explanations for Machine Learning models. The goal is to develop a library that can easily be extended to new ML models and counterfactual generators.
The authors achieve this by allowing for easy extensibility of the allowable input model and of the generating functions.
The present illustrative examples on the _Give Me Some Credit_ and _MNIST_ data sets.

### Review
Overall, I believe that the provided library is a key contribution to Julia's machine learning ecosystem. It comes along with extensive documentation and will be a key resource for researchers and practitioners. **I recommend acceptance with minor revision.**

### Minor Revisions
* In section 2.2 you mention different desiderata for counterfactual explanations. It would be a good idea to provide functions to evaluate the different generated counterfactuals on some standard metrics for these properties.
* Figure 1 is difficult to read. I recommend increasing the font size of the text in the boxes.
* It might be a good idea to give a description of counterfactual search outside gradient descent so that people unfamiliar with the topic understand what approaches there can be besides gradient descent.
* In the paper you refer to the code in various locations as Listings. However, the code blocks have no numbering. I recommend provide such numbering to the blocks to make them easy to find.
* Section 7.1 has a missing reference for the PROBE method.
* Section 2.2: realistic plausible -> realistic and plausible

### Reviewer Checklist

 - [x]   As the reviewer I confirm that I have read the JuliaCon conflict of interest policy and that there are no conflicts of interest for me to review this work.

- [x] I confirm that I read and will adhere to the JuliaCon code of conduct.

- [x]    Repository: Is the source code for this software available at the [repository url]( https://github.com/pat-alt/CounterfactualExplanations.jl)?
- [x]    License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- [x]   Authorship: Has the submitting author (@pat-alt) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [ ]    Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

- [x]    Authors: Does the paper.tex file include a list of authors with their affiliations?
- [x]    References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
- [x]    Page limit: Is the page limit for full papers respected by the submitted document?

- [x]    Context: is the scientific context motivating the work correctly presented?
- [x]    Methodology: is the approach taken in the work justified, presented with enough details and reference to reproduce it?
- [ ]    Results: are the results presented and compared to approaches with similar goals?
