### 2023-06-20 #130 @whedon
[REVIEW]: Explaining Black-Box Models through Counterfactuals
**Submitting author:** @pat-alt (<a href="http://orcid.org/0000-0003-4726-8613">Patrick Altmeyer</a>)
**Repository:** <a href="https://github.com/pat-alt/CounterfactualExplanations.jl" target ="_blank">https://github.com/pat-alt/CounterfactualExplanations.jl</a>
**Version:** v0.1.14
