### 2023-06-21 #103 @pat-alt
Thanks for this very helpful list of comments @kartikeyrinwa - I plan to address them all this week. Tracking/comments below:

- [ ] In section 2.2 you mention different desiderata for counterfactual explanations. It would be a good idea to provide functions to evaluate the different generated counterfactuals on some standard metrics for these properties.

We now actually have a whole new module for [evaluating](https://juliatrustworthyai.github.io/CounterfactualExplanations.jl/stable/tutorials/evaluation/) and [benchmarking](https://juliatrustworthyai.github.io/CounterfactualExplanations.jl/stable/tutorials/benchmarking/) CE. This probably deserves its own subsection, but I'm already pushing the page limit. Is there some flexibility for this?

- [ ] Figure 1 is difficult to read. I recommend increasing the font size of the text in the boxes.

Good point, done.

- [ ] It might be a good idea to give a description of counterfactual search outside gradient descent so that people unfamiliar with the topic understand what approaches there can be besides gradient descent.

I have a group of students currently working full-time on the package. Among other things they have added support for a gradient-free approach.  Will add a sentence or two on this.

- [ ] In the paper you refer to the code in various locations as Listings. However, the code blocks have no numbering. I recommend provide such numbering to the blocks to make them easy to find.

Not sure how to address this, because of this [issue](https://github.com/JuliaCon/JuliaConSubmission.jl/issues/31) I opened a while ago.

- [ ] Section 7.1 has a missing reference for the PROBE method.

Will fix, thanks.

- [ ] Section 2.2: realistic plausible -> realistic and plausible

Fixed.
