### 2023-06-22 JuliaTrustworthyAI/CounterfactualExplanations.jl#242 @pat-alt
Additionally, I have done the following:

- [ ] Where appropriate, I have added hyperlinks to corresponding sections in the package documentation.
