### 2023-06-30 #130 @pat-alt
@matbesancon @kartikeyrinwa I've now edited according to https://github.com/JuliaCon/proceedings-review/issues/103#issuecomment-1598649885.

In addition to the helpful suggestions, I have incorporated recent updates to the package including:

- [Composability of Generators](https://juliatrustworthyai.github.io/CounterfactualExplanations.jl/dev/tutorials/generators/)
- [Evaluation and Benchmarking](https://juliatrustworthyai.github.io/CounterfactualExplanations.jl/dev/tutorials/evaluation/)
- Contributions that have come out of the [TU Delft Student Software Project](https://github.com/JuliaTrustworthyAI/CounterfactualExplanations.jl/pull/252)

I noted that the boxes for *testing* and *contribution guidelines* in the Documentation section were unticked. We do run tests, of course - should that be explained in the docs? As for contribution guidelines, we have a section in the [docs](https://juliatrustworthyai.github.io/CounterfactualExplanations.jl/dev/contribute/) and the README - should we improve this somehow?

Thanks again!
