### 2023-08-11 #130 @pat-alt
@matbesancon awesome, thanks. That's all done: [![DOI](https://zenodo.org/badge/440782065.svg)](https://zenodo.org/badge/latestdoi/440782065)

Had to push a new version  `v0.1.14` for this to take effect.
