### 2023-08-11 #130 @matbesancon
@pat-alt can you check the DOIs above and add them to your bib file?
